@@ -97,17 +90,17 @@
     "postcss-flexbugs-fixes": "^5.0.2",
     "prettier": "^2.3.1",
     "re-resizable": "^6.9.0",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-confetti": "^6.1.0",
     "react-contexify": "^5.0.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-ga": "^3.3.0",
     "react-gtm-module": "^2.0.11",
     "react-hook-form": "^7.0.7",
     "react-icons": "^4.9.0",
     "react-query": "^3.26.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.14.2",
     "react-scripts": "5.0.1",
     "react-use-localstorage": "^3.5.3",
     "shared": "workspace:*",
@@ -175,18 +172,21 @@
     "eslint-plugin-prettier": "^3.4.0",
     "eslint-plugin-react": "^7.24.0",
     "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "history": "^5.3.0",
     "identity-obj-proxy": "^3.0.0",
     "jest-canvas-mock": "^2.3.1",
     "jsdom": "^20.0.3",
     "minimist": "^1.2.5",
-    "msw": "^0.36.8",
+    "msw": "^1.2.3",
+    "react@latest": "link:@@testing-library/react@latest",
     "source-map-explorer": "^2.5.2",
+    "supabase": "^1.82.1",
     "tailwindcss": "^3.2.6",
     "ts-node": "^10.9.1",
     "ts-unused-exports": "^7.0.3",
-    "typescript": "^4.8.4",
+    "typescript": "^5.1.6",
     "web-worker": "^1.2.0"
   },
   "jest": {
