@@ -1,24 +1,31 @@
 import { Trans } from "@lingui/macro";
 import { CopySimple } from "phosphor-react";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
-import { randomChartName, titleToLocalStorageKey } from "../lib/helpers";
+import { slugify, titleToLocalStorageKey } from "../lib/helpers";
 import { docToString, useDoc } from "../lib/useDoc";
 import { Button2 } from "../ui/Shared";
+import { useContext, useState } from "react";
+import { AppContext } from "./AppContext";
+import { getFunFlowchartName } from "../lib/getFunFlowchartName";
+import { languages } from "../locales/i18n";
 
 export function CloneButton() {
-  const { push } = useHistory();
+  const navigate = useNavigate();
   const fullText = useDoc((s) => docToString(s));
+  const language = useContext(AppContext).language;
+  const [name] = useState<string>(
+    slugify(getFunFlowchartName(language as keyof typeof languages))
+  );
   return (
     <Button2
       color="blue"
       onClick={() => {
-        const newChartTitle = randomChartName();
         window.localStorage.setItem(
-          titleToLocalStorageKey(newChartTitle),
+          titleToLocalStorageKey(name),
           fullText ?? ""
         );
-        push(`/${newChartTitle}`);
+        navigate(`/${name}`);
       }}
       rightIcon={<CopySimple size={16} />}
     >
