@@ -1,14 +1,14 @@
 import { Trans } from "@lingui/macro";
 import * as Tabs from "@radix-ui/react-tabs";
 
-import { useIsValidSponsor } from "../../lib/hooks";
+import { useIsProUser } from "../../lib/hooks";
 import styles from "./EditorTabList.module.css";
 
 const btnDark =
   "dark:border-neutral-600 dark:text-neutral-300 dark:hover:bg-neutral-800";
 
 export function EditorTabList() {
-  const isValidSponsor = useIsValidSponsor();
+  const isProUser = useIsProUser();
 
   return (
     <Tabs.List className="grid grid-flow-col gap-2 border-b border-neutral-300 dark:border-neutral-600 pt-2 px-5 pb-0 justify-start">
@@ -39,7 +39,7 @@ export function EditorTabList() {
           <Trans>Style</Trans>
         </span>
       </Tabs.Trigger>
-      {isValidSponsor && (
+      {isProUser && (
         <Tabs.Trigger
           value="Advanced"
           data-testid="Editor Tab: Advanced"
