@@ -1,461 +1,5 @@
 /*eslint-disable*/ module.exports = {
-  messages: {
-    "$3 per month": "3 $ pro Monat",
-    "$30 per year": "30 $ pro Jahr",
-    "(before a colon creates a label)":
-      "(vor einem Doppelpunkt wird zu einer Beschriftung)",
-    "<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":
-      "<0>Flowchart Fun</0> ist ein Open-Source-Projekt von <1>Tone Row</1>",
-    "A new version of the app is available. Please reload to update.":
-      "Eine neue Version der App ist verfügbar. Bitte neu laden, um zu aktualisieren.",
-    About: "Über",
-    "Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":
-      "Greifen Sie von überall aus auf diese Diagramme zu.<0/>Teilen und betten Sie Flussdiagramme ein, die synchron bleiben.",
-    "Accessible from any device": "Von jedem Gerät aus zugänglich",
-    Account: "Konto",
-    "Activate your account": "Aktiviere dein Konto ",
-    Advanced: "Vorangeschritten",
-    Amount: "Betrag",
-    "An error occurred. Try resubmitting or email {0} directly.": [
-      "Es ist ein Fehler aufgetreten. Versuchen Sie, es erneut einzureichen oder senden Sie eine E-Mail direkt an ",
-      ["0"],
-      ".",
-    ],
-    Appearance: "Erscheinungsbild",
-    "Apply Style": "Stil anwenden",
-    Attributes: "Attribute",
-    "Auto-Layouts": "Automatische Layouts",
-    "Back To Editor": "Zurück zum Editor",
-    "Basic Flowchart": "Grundlegender Flussdiagramm",
-    "Become a Github Sponsor": "Werden Sie ein Github-Sponsor",
-    "Become a Pro User": "Werden Sie ein Pro-Benutzer",
-    "Begin with a simple example showing how <0>Flowchart Fun</0> works.":
-      "Beginnen Sie mit einem einfachen Beispiel, das zeigt, wie <0>Flowchart Fun</0> funktioniert.",
-    Blank: "Leer",
-    Blog: "Blog",
-    "Book a Meeting": "Ein Treffen buchen",
-    Border: "Rand",
-    "Bottom to Top": "Von unten nach oben",
-    Breadthfirst: "In die Breite",
-    Cancel: "Abbrechen",
-    "Cancel your subscription. Your hosted charts will become read-only.":
-      "Kündigen Sie Ihr Abonnement. Ihre gehosteten Diagramme werden schreibgeschützt.",
-    "Certain attributes can be used to customize the appearance or functionality of elements.":
-      "Bestimmte Attribute können verwendet werden, um das Aussehen oder die Funktionalität von Elementen anzupassen.",
-    "Change Email Address": "E-Mail Adresse ändern",
-    Changelog: "Änderungsprotokoll",
-    Charts: "Diagramme",
-    "Check your email for a link to log in.<0/>You can close this window.":
-      "Überprüfen Sie Ihre E-Mail auf einen Link zum Einloggen. Sie können dieses Fenster schließen.",
-    "Choose a Theme": "Wählen Sie ein Thema",
-    "Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":
-      "Wählen Sie aus einer Vielzahl von Pfeilformen für die Quelle und das Ziel einer Kante aus. Formen beinhalten Dreieck, Dreieck-Tee, Kreis-Dreieck, Dreieck-Kreuz, Dreieck-Rückbiegung, Vee, Tee, Quadrat, Kreis, Diamant, Chevron und keine.",
-    "Choose the plan that's right for you and start creating amazing flowcharts with Flowchart Fun Pro":
-      "Wählen Sie den Plan, der am besten zu Ihnen passt, und starten Sie mit Flowchart Fun Pro die Erstellung beeindruckender Flussdiagramme.",
-    Circle: "Kreis",
-    Classes: "Klassen",
-    Clay: "Ton",
-    Clear: "Löschen",
-    "Clear text?": "Text löschen?",
-    "Click on \\nLearn Syntax\\nto learn more":
-      "Klicke auf \\nSyntax lernen\\num mehr zu erfahren.",
-    "Click to preview": "Klicken Sie zum Vorschauen",
-    Clone: "Klon",
-    Close: "Schließen",
-    Color: "Farbe",
-    "Colors include red, orange, yellow, blue, purple, black, white, and gray.":
-      "Farben beinhalten Rot, Orange, Gelb, Blau, Violett, Schwarz, Weiß und Grau.",
-    Column: "Spalte",
-    "Comic Book": "Comic-Buch",
-    Comment: "Kommentar",
-    Concentric: "Konzentrisch",
-    "Confirm New Email": "Neue E-Mail bestätigen",
-    Containers: "Behälter",
-    "Containers are nodes that contain other nodes. They are declared using curly braces.":
-      "Container sind Knoten, die andere Knoten enthalten. Sie werden mit geschweiften Klammern deklariert.",
-    Continue: "Weiter",
-    "Convert this to a hosted chart?":
-      "Konvertiere dies in ein gehostetes Diagramm?",
-    "Convert to hosted chart?": "In gehostetes Diagramm konvertieren?",
-    Copy: "Kopieren",
-    "Copy PNG Image": "PNG-Bild kopieren",
-    "Copy SVG Code": "Kopieren Sie den SVG-Code",
-    "Copy your mermaid.js code or open it directly in the mermaid.js live editor.":
-      "Kopieren Sie Ihren mermaid.js-Code oder öffnen Sie ihn direkt im mermaid.js Live-Editor.",
-    "Create Flowcharts from a Prompt with AI":
-      "Erstellen Sie Diagramme aus einer Eingabeaufforderung mit KI",
-    "Create New Flowchart": "Neuen Flussdiagramm erstellen",
-    "Create a New Flowchart": "Erstellen Sie ein neues Diagramm",
-    "Create professional-quality visual aids for presentations or training materials":
-      "Erstellen Sie professionelle visuelle Hilfsmittel für Präsentationen oder Trainingsmaterialien",
-    "Creating an edge between two nodes is done by indenting the second node below the first":
-      "Eine Kante zwischen zwei Knoten wird erstellt, indem der zweite Knoten unter dem ersten eingerückt wird",
-    "Custom Sharing Options": "Benutzerdefinierte Freigabeoptionen",
-    "Customer Portal": "Kundenportal",
-    "Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":
-      "Passen Sie Ihr Thema an, indem Sie die <0>Cytoscape CSS</0> unten bearbeiten. Unsere Styling-Dokumentation ist bald verfügbar! In der Zwischenzeit ist die beste Ressource die <1>Cytoscape</1>-Dokumentation. Stellen Sie Fragen im <2>Discord</2>, wenn Sie steckenbleiben.",
-    Dark: "Dunkel",
-    "Dark Mode": "Dunkelmodus",
-    Date: "Datum",
-    "Deleted when browser data is cleared":
-      "Wird gelöscht, wenn Browserdaten gelöscht werden",
-    "Describe the flowchart you wish to create":
-      "Beschreiben Sie das Diagramm, das Sie erstellen möchten",
-    Direction: "Richtung",
-    Discard: "Verwerfen",
-    "Do you want to cancel your subscription?":
-      "Möchten Sie Ihr Abonnement kündigen?",
-    "Do you want to delete this?": "Möchten Sie dies löschen?",
-    Document: "Document",
-    "Don't have an account?": "Sie haben kein Konto?",
-    Download: "Herunterladen",
-    "Download JPG": "JPG herunterladen",
-    "Download PNG": "PNG herunterladen",
-    "Download SVG": "SVG herunterladen",
-    "Drag and drop a CSV file here, or click to select a file":
-      "Ziehen Sie eine CSV-Datei hierher oder klicken Sie, um eine Datei auszuwählen",
-    "Draw an edge from multiple nodes by beginning the line with a reference":
-      "Zeichnen Sie eine Kante von mehreren Knoten, indem Sie die Zeile mit einer Referenz beginnen.",
-    "Drop the file here ...": "Datei hier ablegen ...",
-    "Easily import existing flowcharts from other software":
-      "Importieren Sie einfach vorhandene Flussdiagramme aus anderer Software",
-    Edge: "Kante",
-    "Edge ID, Classes, Attributes": "Kante-ID, Klassen, Attribute",
-    "Edge Label": "Kantenbeschriftung",
-    "Edge Label Column": "Spalte mit Kantenbeschriftung",
-    "Edge Style": "Kantenstil",
-    "Edge missing indentation": "Kante fehlende Einrückung",
-    Edges: "Kanten",
-    "Edges are declared in the same row as their source node":
-      "Kanten werden in der gleichen Zeile wie ihr Quellknoten deklariert",
-    "Edges are declared in the same row as their target node":
-      "Kanten werden in der gleichen Zeile wie ihr Zielknoten deklariert",
-    "Edges are declared in their own row":
-      "Kanten werden in ihrer eigenen Zeile deklariert",
-    "Edges can also have ID's, classes, and attributes before the label":
-      "Kanten können auch ID's, Klassen und Attribute vor der Bezeichnung haben",
-    "Edges can be styled with dashed, dotted, or solid lines":
-      "Kanten können mit gestrichelten, punktierten oder soliden Linien gestaltet werden",
-    "Edges in Separate Rows": "Kanten in separaten Zeilen",
-    "Edges in Source Node Row": "Kanten in Quellknotenzeile",
-    "Edges in Target Node Row": "Kanten in Zielknotenzeile",
-    Editable: "Editierbar",
-    Editor: "Editor",
-    Eggs: "Eier",
-    Email: "E-Mail",
-    Empty: "Leer",
-    "Enter your email": "Geben Sie Ihre E-Mail-Adresse ein",
-    "Equal To": "Gleich",
-    "Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":
-      "Essay-Schreibprozess-Flowchart, der Schüler durch Brainstorming, Gliederung, Entwurf und Überarbeitung führt",
-    "Everything in Free": "Alles im kostenlosen Plan",
-    Excalidraw: "Excalidraw",
-    Export: "Exportieren",
-    "Export High-Resolution Images": "Hochauflösende Bilder exportieren",
-    "Export to SVG": "In SVG exportieren",
-    "Extra Large": "Extra groß",
-    Extract: "Extrakt",
-    Feedback: "Feedback",
-    "Follow Us on Twitter": "Folgen Sie uns auf Twitter",
-    "Found a bug? Have a feature request? We would love to hear from you!":
-      "Einen Fehler gefunden? Eine Funktionsanfrage? Wir würden uns freuen von Ihnen zu hören!",
-    Free: "Kostenlos",
-    Fullscreen: "Vollbild",
-    Futuristic: "Futuristisch",
-    "Get More Layouts": "Mehr Layouts erhalten",
-    "Get personalized attention and support from our expert team":
-      "Erhalten Sie personalisierte Aufmerksamkeit und Unterstützung von unserem Experten-Team",
-    "Getting Started": "Erste Schritte",
-    Graph: "Diagramm",
-    Grid: "Raster",
-    "Have complex questions or issues? We're here to help.":
-      "Haben Sie komplexe Fragen oder Probleme? Wir sind hier, um zu helfen.",
-    Help: "Hilfe",
-    History: "Verlauf",
-    Home: "Startseite",
-    "How are edges declared in this data?":
-      "Wie werden Kanten in diesen Daten deklariert?",
-    "ID's": "IDs",
-    "If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":
-      "Wenn Sie <0>Flowchart Fun</0> gerne verwenden, überlegen Sie bitte, das Projekt zu unterstützen.",
-    "If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\:</0>":
-      "Wenn Sie eine Kante erstellen möchten, rücken Sie diese Zeile ein. Wenn nicht, entkomme dem Doppelpunkt mit einem Backslash <0>\\:</0>",
-    "Image Export": "Bildexport",
-    Images: "Bilder",
-    "Import Data": "Daten importieren",
-    "Import data from a CSV file.": "Daten aus einer CSV-Datei importieren.",
-    "Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":
-      "Daten aus jeder CSV-Datei importieren und auf einem neuen Flussdiagramm abbilden. Dies ist eine großartige Möglichkeit, Daten aus anderen Quellen wie Lucidchart, Google Sheets und Visio zu importieren.",
-    "Import from Lucidchart, Visio, and other CSV files":
-      "Importieren Sie aus Lucidchart, Visio und anderen CSV-Dateien",
-    "Import your diagram it into Microsoft Visio using one of these CSV files.":
-      "Importieren Sie Ihr Diagramm mithilfe einer dieser CSV-Dateien in Microsoft Visio.",
-    "Important Changes are Coming to Flowchart Fun.":
-      "Wichtige Änderungen kommen zu Flowchart Fun.",
-    "Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":
-      "Der Datenimport ist eine Profifunktion. Sie können für nur 3 USD pro Monat oder 30 USD pro Jahr auf Flowchart Fun Pro upgraden.",
-    "Impress your audience with high-quality images produced by Flowchart Fun":
-      "Beeindrucken Sie Ihr Publikum mit hochwertigen Bildern, die von Flowchart Fun produziert werden",
-    "Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":
-      "Fügen Sie einen Titel mit einem <0>title</0>-Attribut ein. Um die Visio-Farbgebung zu verwenden, fügen Sie ein <1>roleType</1>-Attribut gleich einem der folgenden hinzu:",
-    "Indenting creates an edge": "Durch das Einrücken entsteht eine Kante",
-    Info: "Info",
-    Instruct: "Anweisen",
-    Is: "Ist",
-    Language: "Sprache",
-    Large: "Groß",
-    Layout: "Layout",
-    "Layout Frozen": "Layout eingefroren",
-    "Layout is Frozen": "Layout ist eingefroren",
-    "Leading References": "Führende Referenzen",
-    "Learn More": "Mehr erfahren",
-    "Learn Syntax": "Syntax lernen",
-    "Learn about Flowchart Fun Pro": "Erfahren Sie mehr über Flowchart Fun Pro",
-    "Left to Right": "Von links nach rechts",
-    "Let's get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":
-      "Los geht's! Geben Sie unten Ihre E-Mail-Adresse ein, um Ihr Flowchart Fun-Konto zu erstellen und unsere leistungsstarken Funktionen zu nutzen.",
-    Light: "Hell",
-    "Light Mode": "Heller Modus",
-    Link: "Link",
-    "Load Theme": "Thema laden",
-    "Loading a theme will erase any custom styles you have applied":
-      "Das Laden eines Themas wird alle benutzerdefinierten Stile löschen, die Sie angewendet haben",
-    "Lock Zoom to Graph": "Zoom an Graph anpassen",
-    "Log In": "Anmelden",
-    "Log Out": "Abmelden",
-    "Log in to start creating flowcharts.":
-      "Melden Sie sich an, um mit dem Erstellen von Flussdiagrammen zu beginnen.",
-    "Make a One-Time Donation": "Machen Sie eine einmalige Spende",
-    "Make publicly accessible": "Öffentlich zugänglich machen",
-    "Make sure you use the same email you will use to log in.":
-      "Stellen Sie sicher, dass Sie die gleiche E-Mail verwenden, mit der Sie sich anmelden werden.",
-    "Map Data": "Daten abbilden",
-    "Market understanding and competitive landscape maintenance for SaaS product development":
-      "Marktverständnis und Wettbewerbslandschaftspflege für die Entwicklung von SaaS-Produkten",
-    Medium: "Mittel",
-    Method: "Methode",
-    Monospace: "Monospace",
-    Monthly: "Monatlich",
-    "Multiple pointers on same line": "Mehrere Zeiger auf derselben Zeile",
-    Museum: "Museum",
-    Name: "Name",
-    New: "Neues",
-    "New Email": "Neue e-mail",
-    "Next charge": "Nächste Gebühr",
-    "No Edges": "Keine Kanten",
-    Node: "Knoten",
-    "Node Border Style": "Knotenrahmenstil",
-    "Node Colors": "Knotenfarben",
-    "Node ID": "Knoten-ID",
-    "Node ID, Classes, Attributes": "Knoten-ID, Klassen, Attribute",
-    "Node Label": "Knotenbeschriftung",
-    "Node Shapes": "Knotenformen",
-    "Node and pointer on same line": "Knoten und Zeiger auf derselben Zeile",
-    "Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":
-      "Knoten können mit gestrichelten, gepunkteten oder doppelten Linien gestaltet werden. Grenzen können auch mit border-none entfernt werden.",
-    "Not Empty": "Nicht leer",
-    "Office Hours": "Geschäftszeiten",
-    "Once in a while the magic link will end up in your spam folder. If you don't see it after a few minutes, check there or request a new link.":
-      "legentlich landet der magische Link in Ihrem Spam-Ordner. Wenn Sie ihn nach ein paar Minuten nicht sehen, überprüfen Sie dort oder fordern Sie einen neuen Link an.",
-    "One-on-One Support": "One-on-One-Support",
-    "One-time Share Links": "Einmalige Share-Links",
-    "Only available on this device.<0/>Clearing your browser cache will erase them.":
-      "Nur auf diesem Gerät verfügbar.<0/>Wenn Sie Ihren Browser-Cache leeren, werden sie gelöscht.",
-    "Open Customer Portal": "Öffnen Sie das Kundenportal",
-    "Organization Chart": "Organigramm",
-    "Paste the information you wish to convert to a flowchart":
-      "Fügen Sie die Informationen ein, die Sie in ein Flussdiagramm umwandeln möchten",
-    Permalinks: "Permalinks",
-    "Permanent Charts are a Pro Feature":
-      "Permanente Diagramme sind eine Pro-Funktion",
-    "Permanent Flowcharts": "Permanente Flussdiagramme",
-    Persistent: "Beständig",
-    "Persistent Flowcharts": "Beständige Diagramme",
-    Playbook: "Spielbuch",
-    "Pointer and container on same line":
-      "Zeiger und Container auf derselben Zeile",
-    Pricing: "Preisgestaltung",
-    "Process for corporate social responsibility initiatives development and implementation across company operations":
-      "Prozess zur Entwicklung und Implementierung von Corporate Social Responsibility-Initiativen in Unternehmensoperationen",
-    "Processing Data": "Datenverarbeitung",
-    Prompt: "Aufforderung",
-    Public: "Öffentlich",
-    Random: "Zufällig",
-    "Read-only": "Schreibgeschützt",
-    "Reference by Class": "Referenz nach Klasse",
-    "Reference by ID": "Referenz nach ID",
-    "Reference by Label": "Referenz nach Label",
-    References: "Referenzen",
-    "References are used to create edges between nodes that are created elsewhere in the document":
-      "Referenzen werden verwendet, um Kanten zwischen Knoten zu erstellen, die anderswo im Dokument erstellt werden",
-    "Referencing a node by its exact label":
-      "Referenzierung eines Knotens durch sein exaktes Label",
-    "Referencing a node by its unique ID":
-      "Referenzierung eines Knotens durch seine eindeutige ID",
-    "Referencing multiple nodes with the same assigned class":
-      "Mehrere Knoten mit der selben zugewiesenen Klasse referenzieren",
-    "Refresh Page": "Seite aktualisieren",
-    "Reload to Update": "Neu laden, um zu aktualisieren",
-    "Remove Image Watermark": "Bildwasserzeichen entfernen",
-    Rename: "Umbenennen",
-    Reset: "Zurücksetzen",
-    "Resume Subscription": "Abonnement fortsetzen",
-    Retro: "Retro",
-    Return: "Zurückkehren",
-    "Right to Left": "Von rechts nach links",
-    Roadmap: "Fahrplan",
-    Save: "Speichern",
-    "Save 20% (2 months free!)": "Sparen Sie 20% (2 Monate kostenlos!)",
-    "Save time and automate certain tasks with this cutting-edge feature":
-      "Sparen Sie Zeit und automatisieren Sie bestimmte Aufgaben mit dieser fortschrittlichen Funktion",
-    "Save time and streamline your workflow by importing data from other sources":
-      "Sparen Sie Zeit und vereinfachen Sie Ihren Arbeitsablauf, indem Sie Daten von anderen Quellen importieren",
-    "Select your plan!": "Wählen Sie Ihren Plan aus!",
-    "Send Feedback": "Feedback senden",
-    Settings: "Einstellungen",
-    Shape: "Form",
-    "Sign Up": "Registrieren",
-    "Sign Up Now": "Jetzt registrieren",
-    Size: "Größe",
-    Small: "Klein",
-    "Some classes are available to help style your flowchart":
-      "Einige Klassen stehen zur Verfügung, um Ihren Flussdiagramm zu gestalten",
-    "Sorry! This page is only available in English.":
-      "Entschuldigung! Diese Seite ist nur auf Englisch verfügbar.",
-    "Source Arrow": "Quellpfeil",
-    "Source Column": "Quellspalte",
-    "Source Delimiter": "Quell-Trennzeichen",
-    "Source/Target Arrow Shape": "Quelle/Ziel-Pfeilform",
-    Spacing: "Abstand",
-    "Special Attributes": "Spezielle Attribute",
-    Start: "Start",
-    "Start Over": "Von vorne anfangen",
-    Status: "Status",
-    "Store any data associated to a node":
-      "Speichern Sie alle Daten, die einem Knoten zugeordnet sind",
-    "Stored in the cloud": "In der Cloud gespeichert",
-    "Stored on this computer": "Auf diesem Computer gespeichert",
-    "Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":
-      "Vereinfachen Sie Ihren Arbeitsablauf und vereinfachen Sie die Prozessvisualisierung mit Flowchart Fun. Abonnieren Sie jetzt und nutzen Sie die leistungsstarken Pro-Funktionen!",
-    Style: "Stil",
-    "Style Classes": "Stil-Klassen",
-    Submit: "Einsenden",
-    Subscribe: "Abonnieren",
-    Subscription: "Abonnement",
-    "Subscription will end": "Abonnement wird beendet",
-    "Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":
-      "Supply-Chain-Analyse und -Optimierung: Kostenreduzierung, Effizienzverbesserung und Stakeholder-Kollaboration",
-    "Target Arrow": "Zielpfeil",
-    "Target Column": "Ziel-Spalte",
-    "Target Delimiter": "Ziel-Trennzeichen",
-    Temporary: "Temporäres",
-    "Temporary Flowchart Warning": "Vorübergehende Flussdiagramm Warnung",
-    "Temporary Flowcharts": "Temporäre Diagramme",
-    "Text Sizes": "Textgrößen",
-    "Text followed by colon+space creates an edge with the text as the label":
-      "Text gefolgt von Doppelpunkt + Leerzeichen erstellt eine Kante mit dem Text als Label",
-    "Text on a line creates a node with the text as the label":
-      "Text in einer Zeile erstellt einen Knoten mit dem Text als Label",
-    "Text sizes include small, regular, large, and extra-large.":
-      "Textgrößen umfassen klein, regulär, groß und extra groß.",
-    "Thank you for your feedback!": "Danke für Ihr Feedback!",
-    "The column that contains the edge label(s)":
-      "Die Spalte, die die Kantenbeschriftung(en) enthält",
-    "The column that contains the source node ID(s)":
-      "Die Spalte, die die Quellknoten-ID(en) enthält",
-    "The column that contains the target node ID(s)":
-      "Die Spalte, die die Zielknoten-ID(en) enthält",
-    "The delimiter used to separate multiple source nodes":
-      "Der Trennzeichen, das verwendet wird, um mehrere Quellknoten zu trennen",
-    "The delimiter used to separate multiple target nodes":
-      "Der Trennzeichen, das verwendet wird, um mehrere Zielknoten zu trennen",
-    "The possible shapes are:": "Die möglichen Formen sind:",
-    "Theme Editor": "Thema-Editor",
-    "There are no edges in this data": "Es gibt keine Kanten in diesen Daten",
-    "This action cannot be undone.":
-      "Diese Aktion kann nicht rückgängig gemacht werden.",
-    "This app works by typing": "Diese App funktioniert durch Tippen",
-    "This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":
-      "Dieses Dokument ist nur auf diesem Computer gespeichert. Wenn Sie eine Möglichkeit suchen, Dokumente zu speichern und von überall aus darauf zugreifen zu können, probieren Sie unseren Pro-Plan für nur 3 $/Monat aus!",
-    "This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":
-      "Diese Funktion ist nur für Pro-Benutzer verfügbar. <0>Werden Sie Pro-Nutzer</0>, um es freizuschalten.",
-    "This is a temporary chart, it will be deleted when you clear your browser cache.":
-      "Dies ist ein temporäres Diagramm, es wird gelöscht, wenn Sie den Browser-Cache löschen.",
-    Tip: "Tipp",
-    "To fix this change one of the edge IDs":
-      "Um dies zu beheben, ändern Sie eine der Kanten-IDs",
-    "To fix this change one of the node IDs":
-      "Um das zu beheben, ändern Sie eine der Knoten-IDs",
-    "To fix this move one pointer to the next line":
-      "Um das zu beheben, verschieben Sie einen Zeiger auf die nächste Zeile",
-    "To fix this move the pointer to the next line":
-      "Um das zu beheben, verschieben Sie den Zeiger auf die nächste Zeile",
-    "To fix this start the container <0/> on a different line":
-      "Um dies zu beheben, starten Sie den Container <0/> auf einer anderen Zeile.",
-    "Top to Bottom": "Von oben nach unten",
-    "Try again": "Erneut versuchen",
-    "Two edges have the same ID": "Zwei Kanten haben die gleiche ID",
-    "Two nodes have the same ID": "Zwei Knoten haben die gleiche ID",
-    Type: "Typ",
-    Unfreeze: "Freigeben",
-    "Unique text value to identify a node":
-      "Einzigartiger Textwert, um einen Knoten zu identifizieren",
-    "Unknown Parsing Error": "Unbekannter Parser-Fehler",
-    "Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":
-      "Entfesseln Sie Ihre Kreativität und vereinfachen Sie Ihren Arbeitsablauf mit Flowchart Fun Pro – ab nur 3 $/Monat!",
-    "Update Email": "E-Mail aktualisieren",
-    "Upgrade to Pro": "Auf Pro upgraden",
-    "Upload your File": "Laden Sie Ihre Datei hoch",
-    "Use AI to create flowcharts from data sets or algorithms":
-      "Verwenden Sie KI, um aus Datensätzen oder Algorithmen Flussdiagramme zu erstellen.",
-    "Use AI to generate a flowchart from a prompt. <0>Learn More</0>":
-      "Verwenden Sie AI, um aus einem Prompt ein Flowchart zu erstellen. <0>Erfahren Sie mehr</0>",
-    "Use classes to group nodes":
-      "Verwenden Sie Klassen, um Knoten zu gruppieren",
-    "Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":
-      "Verwenden Sie das Attribut <0>href</0>, um einem Knoten einen Link zu setzen, der in einem neuen Tab geöffnet wird.",
-    "Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":
-      "Verwenden Sie das Attribut <0>src</0>, um das Bild eines Knotens zu setzen. Das Bild wird an den Knoten angepasst, sodass Sie möglicherweise die Breite und Höhe des Knotens anpassen müssen, um das gewünschte Ergebnis zu erzielen. Es werden nur öffentliche Bilder (nicht von CORS blockiert) unterstützt.",
-    "Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":
-      "Verwenden Sie die Attribute <0>w</0> und <1>h</1>, um die Breite und Höhe eines Knotens explizit festzulegen.",
-    "Use the customer portal to change your billing information.":
-      "Verwenden Sie das Kundenportal, um Ihre Rechnungsinformationen zu ändern.",
-    "Use this file for org charts, hierarchies, and other organizational structures.":
-      "Verwenden Sie diese Datei für Organigramme, Hierarchien und andere Organisationsstrukturen.",
-    "Use this file for sequences, processes, and workflows.":
-      "Verwenden Sie diese Datei für Sequenzen, Prozesse und Workflows.",
-    User: "Benutzer",
-    "View on Github": "Auf Github ansehen",
-    "Water evaporates from the Earth's surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":
-      "Wasser verdunstet von der Erdoberfläche, steigt in die Atmosphäre auf und fällt als Niederschlag wieder herunter. Dieses Wasser fließt dann in Flüsse, Seen und Ozeane, wo es wieder verdunstet und in die Atmosphäre zurückgeführt wird.",
-    "We use magic links to log you in. Enter your email below to get started.":
-      "Wir verwenden magische Links, um Sie einzuloggen. Geben Sie unten Ihre E-Mail-Adresse ein, um zu beginnen.",
-    "Welcome to Flowchart Fun Pro!": "Willkommen bei Flowchart Fun Pro!",
-    "Width and Height": "Breite und Höhe",
-    "Would you like to continue?": "Möchten Sie fortfahren?",
-    Yearly: "Jährlich",
-    "You already have a flowchart with this name.":
-      "Sie haben bereits ein Diagramm mit diesem Namen.",
-    "You are about to add {numNodes} nodes and {numEdges} edges to your graph.":
-      [
-        "Sie sind dabei, ",
-        ["numNodes"],
-        " Knoten und ",
-        ["numEdges"],
-        " Kanten zu Ihrem Graphen hinzuzufügen.",
-      ],
-    'You can also wrap text in "\\( \\)" to create an edge':
-      'Sie können Text auch in "\\( \\)" umbrechen, um eine Kante zu erstellen',
-    "You must log in to create a persistent flowchart.":
-      "Sie müssen sich anmelden, um einen beständigen Flussdiagramm zu erstellen.",
-    "You're almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":
-      "Du bist fast da! Nur noch ein Schritt, um das volle Potenzial von <0>Flowchart Fun Pro</0> freizuschalten. Gib deine Zahlungsdetails unten ein, um dein Abonnement abzuschließen und heute mit dem Erstellen beeindruckender Flussdiagramme zu beginnen.",
-    "Your Charts": "Ihre Diagramme",
-    "Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":
-      "Ihr Abonnement ist nicht mehr aktiv. Wenn Sie dauerhafte Diagramme erstellen und bearbeiten möchten, upgraden Sie auf Pro.",
-    "Zoom In": "Vergrößern",
-    "Zoom Out": "Verkleinern",
-    "any text: before a colon creates a label":
-      "Jeder beliebige Text: vor einem Doppelpunkt wird zu einer Beschriftung",
-  },
+  messages: JSON.parse(
+    '{"$3 per month":"3 $ pro Monat","$30 per year":"30 $ pro Jahr","(before a colon creates a label)":"(vor einem Doppelpunkt wird zu einer Beschriftung)","<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":"<0>Flowchart Fun</0> ist ein Open-Source-Projekt von <1>Tone Row</1>","<0>Sign In</0> / <1>Sign Up</1> with email and password":"<0>Anmelden</0> / <1>Registrieren</1> mit E-Mail und Passwort","<0>Starting August 28th you will need a subscription to create and edit charts. <1>Learn more</1>.</0>":"<0>Ab dem 28. August benötigen Sie ein Abonnement, um Diagramme zu erstellen und zu bearbeiten. <1>Erfahren Sie mehr</1>.</0>","<0>You currently have a free account.<1/><2>Learn about our Pro Features and subscribe on our pricing page</2>.</0>":"<0>Sie haben derzeit ein kostenloses Konto.<1/><2>Erfahren Sie mehr über unsere Pro-Funktionen und abonnieren Sie unsere Preisseite</2>.</0>","A new version of the app is available. Please reload to update.":"Eine neue Version der App ist verfügbar. Bitte neu laden, um zu aktualisieren.","About":"Über","Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":"Greifen Sie von überall aus auf diese Diagramme zu.<0/>Teilen und betten Sie Flussdiagramme ein, die synchron bleiben.","Accessible from any device":"Von jedem Gerät aus zugänglich","Account":"Konto","Activate your account":"Aktiviere dein Konto ","Advanced":"Vorangeschritten","Amount":"Betrag","An error occurred. Try resubmitting or email {0} directly.":["Es ist ein Fehler aufgetreten. Versuchen Sie, es erneut einzureichen oder senden Sie eine E-Mail direkt an ",["0"],"."],"Appearance":"Erscheinungsbild","Apply Style":"Stil anwenden","Attributes":"Attribute","Auto-Layouts":"Automatische Layouts","Back To Editor":"Zurück zum Editor","Basic Flowchart":"Grundlegender Flussdiagramm","Become a Github Sponsor":"Werden Sie ein Github-Sponsor","Become a Pro User":"Werden Sie ein Pro-Benutzer","Begin with a simple example showing how <0>Flowchart Fun</0> works.":"Beginnen Sie mit einem einfachen Beispiel, das zeigt, wie <0>Flowchart Fun</0> funktioniert.","Blank":"Leer","Blog":"Blog","Book a Meeting":"Ein Treffen buchen","Border":"Rand","Bottom to Top":"Von unten nach oben","Breadthfirst":"In die Breite","Cancel":"Abbrechen","Cancel your subscription. Your hosted charts will become read-only.":"Kündigen Sie Ihr Abonnement. Ihre gehosteten Diagramme werden schreibgeschützt.","Certain attributes can be used to customize the appearance or functionality of elements.":"Bestimmte Attribute können verwendet werden, um das Aussehen oder die Funktionalität von Elementen anzupassen.","Change Email Address":"E-Mail Adresse ändern","Changelog":"Änderungsprotokoll","Charts":"Diagramme","Check your email for a link to log in.<0/>You can close this window.":"Überprüfen Sie Ihre E-Mail auf einen Link zum Einloggen. Sie können dieses Fenster schließen.","Choose":"Wählen","Choose a Theme":"Wählen Sie ein Thema","Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":"Wählen Sie aus einer Vielzahl von Pfeilformen für die Quelle und das Ziel einer Kante aus. Formen beinhalten Dreieck, Dreieck-Tee, Kreis-Dreieck, Dreieck-Kreuz, Dreieck-Rückbiegung, Vee, Tee, Quadrat, Kreis, Diamant, Chevron und keine.","Choose the plan that\'s right for you and start creating amazing flowcharts with Flowchart Fun Pro":"Wählen Sie den Plan, der am besten zu Ihnen passt, und starten Sie mit Flowchart Fun Pro die Erstellung beeindruckender Flussdiagramme.","Circle":"Kreis","Classes":"Klassen","Clay":"Ton","Clear":"Löschen","Clear text?":"Text löschen?","Click on \\\\nLearn Syntax\\\\nto learn more":"Klicke auf \\\\nSyntax lernen\\\\num mehr zu erfahren.","Click to preview":"Klicken Sie zum Vorschauen","Clone":"Klon","Close":"Schließen","Color":"Farbe","Colors include red, orange, yellow, blue, purple, black, white, and gray.":"Farben beinhalten Rot, Orange, Gelb, Blau, Violett, Schwarz, Weiß und Grau.","Column":"Spalte","Comic Book":"Comic-Buch","Comment":"Kommentar","Concentric":"Konzentrisch","Confirm New Email":"Neue E-Mail bestätigen","Confirm your email address to sign in.":"Bestätigen Sie Ihre E-Mail-Adresse, um sich anzumelden.","Containers":"Behälter","Containers are nodes that contain other nodes. They are declared using curly braces.":"Container sind Knoten, die andere Knoten enthalten. Sie werden mit geschweiften Klammern deklariert.","Continue":"Weiter","Convert this to a hosted chart?":"Konvertiere dies in ein gehostetes Diagramm?","Convert to hosted chart?":"In gehostetes Diagramm konvertieren?","Copy":"Kopieren","Copy PNG Image":"PNG-Bild kopieren","Copy SVG Code":"Kopieren Sie den SVG-Code","Copy your mermaid.js code or open it directly in the mermaid.js live editor.":"Kopieren Sie Ihren mermaid.js-Code oder öffnen Sie ihn direkt im mermaid.js Live-Editor.","Create Flowcharts from a Prompt with AI":"Erstellen Sie Diagramme aus einer Eingabeaufforderung mit KI","Create New Flowchart":"Neuen Flussdiagramm erstellen","Create a New Flowchart":"Erstellen Sie ein neues Diagramm","Create professional-quality visual aids for presentations or training materials":"Erstellen Sie professionelle visuelle Hilfsmittel für Präsentationen oder Trainingsmaterialien","Creating an edge between two nodes is done by indenting the second node below the first":"Eine Kante zwischen zwei Knoten wird erstellt, indem der zweite Knoten unter dem ersten eingerückt wird","Custom Sharing Options":"Benutzerdefinierte Freigabeoptionen","Customer Portal":"Kundenportal","Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":"Passen Sie Ihr Thema an, indem Sie die <0>Cytoscape CSS</0> unten bearbeiten. Unsere Styling-Dokumentation ist bald verfügbar! In der Zwischenzeit ist die beste Ressource die <1>Cytoscape</1>-Dokumentation. Stellen Sie Fragen im <2>Discord</2>, wenn Sie steckenbleiben.","Dark":"Dunkel","Dark Mode":"Dunkelmodus","Date":"Datum","Deleted when browser data is cleared":"Wird gelöscht, wenn Browserdaten gelöscht werden","Describe the flowchart you wish to create":"Beschreiben Sie das Diagramm, das Sie erstellen möchten","Direction":"Richtung","Discard":"Verwerfen","Do you want to cancel your subscription?":"Möchten Sie Ihr Abonnement kündigen?","Do you want to delete this?":"Möchten Sie dies löschen?","Document":"Document","Download":"Herunterladen","Download JPG":"JPG herunterladen","Download PNG":"PNG herunterladen","Download SVG":"SVG herunterladen","Drag and drop a CSV file here, or click to select a file":"Ziehen Sie eine CSV-Datei hierher oder klicken Sie, um eine Datei auszuwählen","Draw an edge from multiple nodes by beginning the line with a reference":"Zeichnen Sie eine Kante von mehreren Knoten, indem Sie die Zeile mit einer Referenz beginnen.","Drop the file here ...":"Datei hier ablegen ...","Easily import existing flowcharts from other software":"Importieren Sie einfach vorhandene Flussdiagramme aus anderer Software","Edge":"Kante","Edge ID, Classes, Attributes":"Kante-ID, Klassen, Attribute","Edge Label":"Kantenbeschriftung","Edge Label Column":"Spalte mit Kantenbeschriftung","Edge Style":"Kantenstil","Edge missing indentation":"Kante fehlende Einrückung","Edges":"Kanten","Edges are declared in the same row as their source node":"Kanten werden in der gleichen Zeile wie ihr Quellknoten deklariert","Edges are declared in the same row as their target node":"Kanten werden in der gleichen Zeile wie ihr Zielknoten deklariert","Edges are declared in their own row":"Kanten werden in ihrer eigenen Zeile deklariert","Edges can also have ID\'s, classes, and attributes before the label":"Kanten können auch ID\'s, Klassen und Attribute vor der Bezeichnung haben","Edges can be styled with dashed, dotted, or solid lines":"Kanten können mit gestrichelten, punktierten oder soliden Linien gestaltet werden","Edges in Separate Rows":"Kanten in separaten Zeilen","Edges in Source Node Row":"Kanten in Quellknotenzeile","Edges in Target Node Row":"Kanten in Zielknotenzeile","Editable":"Editierbar","Editor":"Editor","Eggs":"Eier","Email":"E-Mail","Empty":"Leer","Enter your email":"Geben Sie Ihre E-Mail-Adresse ein","Enter your email address and we\'ll send you a magic link to sign in.":"Geben Sie Ihre E-Mail-Adresse ein, und wir senden Ihnen einen magischen Link, um sich anzumelden.","Enter your email address below and we\'ll send you a link to reset your password.":"Geben Sie unten Ihre E-Mail-Adresse ein, und wir senden Ihnen einen Link, um Ihr Passwort zurückzusetzen.","Equal To":"Gleich","Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":"Essay-Schreibprozess-Flowchart, der Schüler durch Brainstorming, Gliederung, Entwurf und Überarbeitung führt","Everything in Free":"Alles im kostenlosen Plan","Excalidraw":"Excalidraw","Export":"Exportieren","Export High-Resolution Images":"Hochauflösende Bilder exportieren","Export to SVG":"In SVG exportieren","Extra Large":"Extra groß","Extract":"Extrakt","Feedback":"Feedback","Follow Us on Twitter":"Folgen Sie uns auf Twitter","Forgot your password?":"Haben Sie Ihr Passwort vergessen?","Found a bug? Have a feature request? We would love to hear from you!":"Einen Fehler gefunden? Eine Funktionsanfrage? Wir würden uns freuen von Ihnen zu hören!","Free":"Kostenlos","Fullscreen":"Vollbild","Futuristic":"Futuristisch","Get More Layouts":"Mehr Layouts erhalten","Get personalized attention and support from our expert team":"Erhalten Sie personalisierte Aufmerksamkeit und Unterstützung von unserem Experten-Team","Getting Started":"Erste Schritte","Graph":"Diagramm","Grid":"Raster","Have complex questions or issues? We\'re here to help.":"Haben Sie komplexe Fragen oder Probleme? Wir sind hier, um zu helfen.","Help":"Hilfe","History":"Verlauf","Home":"Startseite","How are edges declared in this data?":"Wie werden Kanten in diesen Daten deklariert?","ID\'s":"IDs","If an account with that email exists, we\'ve sent you an email with instructions on how to reset your password.":"Wenn ein Konto mit dieser E-Mail vorhanden ist, haben wir Ihnen eine E-Mail mit Anweisungen zum Zurücksetzen Ihres Passworts gesendet.","If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":"Wenn Sie <0>Flowchart Fun</0> gerne verwenden, überlegen Sie bitte, das Projekt zu unterstützen.","If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\\\:</0>":"Wenn Sie eine Kante erstellen möchten, rücken Sie diese Zeile ein. Wenn nicht, entkomme dem Doppelpunkt mit einem Backslash <0>\\\\:</0>","Image Export":"Bildexport","Images":"Bilder","Import Data":"Daten importieren","Import data from a CSV file.":"Daten aus einer CSV-Datei importieren.","Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":"Daten aus jeder CSV-Datei importieren und auf einem neuen Flussdiagramm abbilden. Dies ist eine großartige Möglichkeit, Daten aus anderen Quellen wie Lucidchart, Google Sheets und Visio zu importieren.","Import from Lucidchart, Visio, and other CSV files":"Importieren Sie aus Lucidchart, Visio und anderen CSV-Dateien","Import your diagram it into Microsoft Visio using one of these CSV files.":"Importieren Sie Ihr Diagramm mithilfe einer dieser CSV-Dateien in Microsoft Visio.","Important Changes are Coming to Flowchart Fun.":"Wichtige Änderungen kommen zu Flowchart Fun.","Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":"Der Datenimport ist eine Profifunktion. Sie können für nur 3 USD pro Monat oder 30 USD pro Jahr auf Flowchart Fun Pro upgraden.","Impress your audience with high-quality images produced by Flowchart Fun":"Beeindrucken Sie Ihr Publikum mit hochwertigen Bildern, die von Flowchart Fun produziert werden","Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":"Fügen Sie einen Titel mit einem <0>title</0>-Attribut ein. Um die Visio-Farbgebung zu verwenden, fügen Sie ein <1>roleType</1>-Attribut gleich einem der folgenden hinzu:","Indenting creates an edge":"Durch das Einrücken entsteht eine Kante","Info":"Info","Instruct":"Anweisen","Is":"Ist","Language":"Sprache","Large":"Groß","Layout":"Layout","Layout Frozen":"Layout eingefroren","Layout is Frozen":"Layout ist eingefroren","Leading References":"Führende Referenzen","Learn More":"Mehr erfahren","Learn Syntax":"Syntax lernen","Learn about Flowchart Fun Pro":"Erfahren Sie mehr über Flowchart Fun Pro","Left to Right":"Von links nach rechts","Let\'s get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":"Los geht\'s! Geben Sie unten Ihre E-Mail-Adresse ein, um Ihr Flowchart Fun-Konto zu erstellen und unsere leistungsstarken Funktionen zu nutzen.","Light":"Hell","Light Mode":"Heller Modus","Link":"Link","Load Theme":"Thema laden","Loading a theme will erase any custom styles you have applied":"Das Laden eines Themas wird alle benutzerdefinierten Stile löschen, die Sie angewendet haben","Lock Zoom to Graph":"Zoom an Graph anpassen","Log In":"Anmelden","Log Out":"Abmelden","Log in to start creating flowcharts.":"Melden Sie sich an, um mit dem Erstellen von Flussdiagrammen zu beginnen.","Make a One-Time Donation":"Machen Sie eine einmalige Spende","Make publicly accessible":"Öffentlich zugänglich machen","Make sure you use the same email you will use to log in.":"Stellen Sie sicher, dass Sie die gleiche E-Mail verwenden, mit der Sie sich anmelden werden.","Map Data":"Daten abbilden","Market understanding and competitive landscape maintenance for SaaS product development":"Marktverständnis und Wettbewerbslandschaftspflege für die Entwicklung von SaaS-Produkten","Medium":"Mittel","Method":"Methode","Monospace":"Monospace","Monthly":"Monatlich","Multiple pointers on same line":"Mehrere Zeiger auf derselben Zeile","Museum":"Museum","Name":"Name","New":"Neues","New Email":"Neue e-mail","Next charge":"Nächste Gebühr","No Edges":"Keine Kanten","Node":"Knoten","Node Border Style":"Knotenrahmenstil","Node Colors":"Knotenfarben","Node ID":"Knoten-ID","Node ID, Classes, Attributes":"Knoten-ID, Klassen, Attribute","Node Label":"Knotenbeschriftung","Node Shapes":"Knotenformen","Node and pointer on same line":"Knoten und Zeiger auf derselben Zeile","Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":"Knoten können mit gestrichelten, gepunkteten oder doppelten Linien gestaltet werden. Grenzen können auch mit border-none entfernt werden.","Not Empty":"Nicht leer","Office Hours":"Geschäftszeiten","Once in a while the magic link will end up in your spam folder. If you don\'t see it after a few minutes, check there or request a new link.":"legentlich landet der magische Link in Ihrem Spam-Ordner. Wenn Sie ihn nach ein paar Minuten nicht sehen, überprüfen Sie dort oder fordern Sie einen neuen Link an.","One-on-One Support":"One-on-One-Support","One-time Share Links":"Einmalige Share-Links","Only available on this device.<0/>Clearing your browser cache will erase them.":"Nur auf diesem Gerät verfügbar.<0/>Wenn Sie Ihren Browser-Cache leeren, werden sie gelöscht.","Open Customer Portal":"Öffnen Sie das Kundenportal","Organization Chart":"Organigramm","Password":"Passwort","Paste the information you wish to convert to a flowchart":"Fügen Sie die Informationen ein, die Sie in ein Flussdiagramm umwandeln möchten","Permalinks":"Permalinks","Permanent Charts are a Pro Feature":"Permanente Diagramme sind eine Pro-Funktion","Permanent Flowcharts":"Permanente Flussdiagramme","Persistent":"Beständig","Persistent Flowcharts":"Beständige Diagramme","Playbook":"Spielbuch","Pointer and container on same line":"Zeiger und Container auf derselben Zeile","Pricing":"Preisgestaltung","Process for corporate social responsibility initiatives development and implementation across company operations":"Prozess zur Entwicklung und Implementierung von Corporate Social Responsibility-Initiativen in Unternehmensoperationen","Processing Data":"Datenverarbeitung","Prompt":"Aufforderung","Public":"Öffentlich","Random":"Zufällig","Read-only":"Schreibgeschützt","Reference by Class":"Referenz nach Klasse","Reference by ID":"Referenz nach ID","Reference by Label":"Referenz nach Label","References":"Referenzen","References are used to create edges between nodes that are created elsewhere in the document":"Referenzen werden verwendet, um Kanten zwischen Knoten zu erstellen, die anderswo im Dokument erstellt werden","Referencing a node by its exact label":"Referenzierung eines Knotens durch sein exaktes Label","Referencing a node by its unique ID":"Referenzierung eines Knotens durch seine eindeutige ID","Referencing multiple nodes with the same assigned class":"Mehrere Knoten mit der selben zugewiesenen Klasse referenzieren","Refresh Page":"Seite aktualisieren","Reload to Update":"Neu laden, um zu aktualisieren","Remove Image Watermark":"Bildwasserzeichen entfernen","Rename":"Umbenennen","Request Magic Link":"Magic-Link anfordern","Request Password Reset":"Passwort zurücksetzen anfordern","Reset":"Zurücksetzen","Reset Password":"Passwort zurücksetzen","Resume Subscription":"Abonnement fortsetzen","Retro":"Retro","Return":"Zurückkehren","Right to Left":"Von rechts nach links","Roadmap":"Fahrplan","Save":"Speichern","Save 20% (2 months free!)":"Sparen Sie 20% (2 Monate kostenlos!)","Save time and automate certain tasks with this cutting-edge feature":"Sparen Sie Zeit und automatisieren Sie bestimmte Aufgaben mit dieser fortschrittlichen Funktion","Save time and streamline your workflow by importing data from other sources":"Sparen Sie Zeit und vereinfachen Sie Ihren Arbeitsablauf, indem Sie Daten von anderen Quellen importieren","Select your plan!":"Wählen Sie Ihren Plan aus!","Send Feedback":"Feedback senden","Settings":"Einstellungen","Shape":"Form","Sign In":"Anmelden","Sign Up":"Registrieren","Sign in with <0>GitHub</0>":"Anmelden mit <0>GitHub</0>","Sign in with <0>Google</0>":"Anmelden mit <0>Google</0>","Size":"Größe","Small":"Klein","Some classes are available to help style your flowchart":"Einige Klassen stehen zur Verfügung, um Ihren Flussdiagramm zu gestalten","Sorry! This page is only available in English.":"Entschuldigung! Diese Seite ist nur auf Englisch verfügbar.","Source Arrow":"Quellpfeil","Source Column":"Quellspalte","Source Delimiter":"Quell-Trennzeichen","Source/Target Arrow Shape":"Quelle/Ziel-Pfeilform","Spacing":"Abstand","Special Attributes":"Spezielle Attribute","Start":"Start","Start Over":"Von vorne anfangen","Status":"Status","Store any data associated to a node":"Speichern Sie alle Daten, die einem Knoten zugeordnet sind","Stored in the cloud":"In der Cloud gespeichert","Stored on this computer":"Auf diesem Computer gespeichert","Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":"Vereinfachen Sie Ihren Arbeitsablauf und vereinfachen Sie die Prozessvisualisierung mit Flowchart Fun. Abonnieren Sie jetzt und nutzen Sie die leistungsstarken Pro-Funktionen!","Style":"Stil","Style Classes":"Stil-Klassen","Submit":"Einsenden","Subscribe":"Abonnieren","Subscription":"Abonnement","Subscription will end":"Abonnement wird beendet","Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":"Supply-Chain-Analyse und -Optimierung: Kostenreduzierung, Effizienzverbesserung und Stakeholder-Kollaboration","Target Arrow":"Zielpfeil","Target Column":"Ziel-Spalte","Target Delimiter":"Ziel-Trennzeichen","Temporary":"Temporäres","Temporary Flowchart Warning":"Vorübergehende Flussdiagramm Warnung","Temporary Flowcharts":"Temporäre Diagramme","Text Sizes":"Textgrößen","Text followed by colon+space creates an edge with the text as the label":"Text gefolgt von Doppelpunkt + Leerzeichen erstellt eine Kante mit dem Text als Label","Text on a line creates a node with the text as the label":"Text in einer Zeile erstellt einen Knoten mit dem Text als Label","Text sizes include small, regular, large, and extra-large.":"Textgrößen umfassen klein, regulär, groß und extra groß.","Thank you for your feedback!":"Danke für Ihr Feedback!","The column that contains the edge label(s)":"Die Spalte, die die Kantenbeschriftung(en) enthält","The column that contains the source node ID(s)":"Die Spalte, die die Quellknoten-ID(en) enthält","The column that contains the target node ID(s)":"Die Spalte, die die Zielknoten-ID(en) enthält","The delimiter used to separate multiple source nodes":"Der Trennzeichen, das verwendet wird, um mehrere Quellknoten zu trennen","The delimiter used to separate multiple target nodes":"Der Trennzeichen, das verwendet wird, um mehrere Zielknoten zu trennen","The possible shapes are:":"Die möglichen Formen sind:","Theme Editor":"Thema-Editor","There are no edges in this data":"Es gibt keine Kanten in diesen Daten","This action cannot be undone.":"Diese Aktion kann nicht rückgängig gemacht werden.","This app works by typing":"Diese App funktioniert durch Tippen","This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":"Dieses Dokument ist nur auf diesem Computer gespeichert. Wenn Sie eine Möglichkeit suchen, Dokumente zu speichern und von überall aus darauf zugreifen zu können, probieren Sie unseren Pro-Plan für nur 3 $/Monat aus!","This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":"Diese Funktion ist nur für Pro-Benutzer verfügbar. <0>Werden Sie Pro-Nutzer</0>, um es freizuschalten.","This is a temporary chart, it will be deleted when you clear your browser cache.":"Dies ist ein temporäres Diagramm, es wird gelöscht, wenn Sie den Browser-Cache löschen.","Tip":"Tipp","To fix this change one of the edge IDs":"Um dies zu beheben, ändern Sie eine der Kanten-IDs","To fix this change one of the node IDs":"Um das zu beheben, ändern Sie eine der Knoten-IDs","To fix this move one pointer to the next line":"Um das zu beheben, verschieben Sie einen Zeiger auf die nächste Zeile","To fix this move the pointer to the next line":"Um das zu beheben, verschieben Sie den Zeiger auf die nächste Zeile","To fix this start the container <0/> on a different line":"Um dies zu beheben, starten Sie den Container <0/> auf einer anderen Zeile.","To learn more about why we require you to log in, please read <0>this blog post</0>.":"Um mehr darüber zu erfahren, warum wir Sie zum Anmelden auffordern, lesen Sie bitte <0>diesen Blog-Beitrag</0>.","Top to Bottom":"Von oben nach unten","Try again":"Erneut versuchen","Two edges have the same ID":"Zwei Kanten haben die gleiche ID","Two nodes have the same ID":"Zwei Knoten haben die gleiche ID","Type":"Typ","Unfreeze":"Freigeben","Unique text value to identify a node":"Einzigartiger Textwert, um einen Knoten zu identifizieren","Unknown Parsing Error":"Unbekannter Parser-Fehler","Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":"Entfesseln Sie Ihre Kreativität und vereinfachen Sie Ihren Arbeitsablauf mit Flowchart Fun Pro – ab nur 3 $/Monat!","Update Email":"E-Mail aktualisieren","Upgrade to Pro":"Auf Pro upgraden","Upload your File":"Laden Sie Ihre Datei hoch","Use AI to create flowcharts from data sets or algorithms":"Verwenden Sie KI, um aus Datensätzen oder Algorithmen Flussdiagramme zu erstellen.","Use AI to generate a flowchart from a prompt. <0>Learn More</0>":"Verwenden Sie AI, um aus einem Prompt ein Flowchart zu erstellen. <0>Erfahren Sie mehr</0>","Use classes to group nodes":"Verwenden Sie Klassen, um Knoten zu gruppieren","Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":"Verwenden Sie das Attribut <0>href</0>, um einem Knoten einen Link zu setzen, der in einem neuen Tab geöffnet wird.","Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":"Verwenden Sie das Attribut <0>src</0>, um das Bild eines Knotens zu setzen. Das Bild wird an den Knoten angepasst, sodass Sie möglicherweise die Breite und Höhe des Knotens anpassen müssen, um das gewünschte Ergebnis zu erzielen. Es werden nur öffentliche Bilder (nicht von CORS blockiert) unterstützt.","Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":"Verwenden Sie die Attribute <0>w</0> und <1>h</1>, um die Breite und Höhe eines Knotens explizit festzulegen.","Use the customer portal to change your billing information.":"Verwenden Sie das Kundenportal, um Ihre Rechnungsinformationen zu ändern.","Use this file for org charts, hierarchies, and other organizational structures.":"Verwenden Sie diese Datei für Organigramme, Hierarchien und andere Organisationsstrukturen.","Use this file for sequences, processes, and workflows.":"Verwenden Sie diese Datei für Sequenzen, Prozesse und Workflows.","User":"Benutzer","View on Github":"Auf Github ansehen","Water evaporates from the Earth\'s surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":"Wasser verdunstet von der Erdoberfläche, steigt in die Atmosphäre auf und fällt als Niederschlag wieder herunter. Dieses Wasser fließt dann in Flüsse, Seen und Ozeane, wo es wieder verdunstet und in die Atmosphäre zurückgeführt wird.","Welcome to Flowchart Fun Pro!":"Willkommen bei Flowchart Fun Pro!","Width and Height":"Breite und Höhe","Would you like to continue?":"Möchten Sie fortfahren?","Yearly":"Jährlich","You already have a flowchart with this name.":"Sie haben bereits ein Diagramm mit diesem Namen.","You are about to add {numNodes} nodes and {numEdges} edges to your graph.":["Sie sind dabei, ",["numNodes"]," Knoten und ",["numEdges"]," Kanten zu Ihrem Graphen hinzuzufügen."],"You can also wrap text in \\"\\\\( \\\\)\\" to create an edge":"Sie können Text auch in \\"\\\\( \\\\)\\" umbrechen, um eine Kante zu erstellen","You must log in to create a persistent flowchart.":"Sie müssen sich anmelden, um einen beständigen Flussdiagramm zu erstellen.","You need to log in to access this page.":"Sie müssen sich anmelden, um auf diese Seite zuzugreifen.","You\'re almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":"Du bist fast da! Nur noch ein Schritt, um das volle Potenzial von <0>Flowchart Fun Pro</0> freizuschalten. Gib deine Zahlungsdetails unten ein, um dein Abonnement abzuschließen und heute mit dem Erstellen beeindruckender Flussdiagramme zu beginnen.","Your Charts":"Ihre Diagramme","Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":"Ihr Abonnement ist nicht mehr aktiv. Wenn Sie dauerhafte Diagramme erstellen und bearbeiten möchten, upgraden Sie auf Pro.","Zoom In":"Vergrößern","Zoom Out":"Verkleinern","any text: before a colon creates a label":"Jeder beliebige Text: vor einem Doppelpunkt wird zu einer Beschriftung","or":"oder"}'
+  ),
 };
