@@ -1,455 +1,5 @@
 /*eslint-disable*/ module.exports = {
-  messages: {
-    "$3 per month": "$3 per month",
-    "$30 per year": "$30 per year",
-    "(before a colon creates a label)": "(before a colon creates a label)",
-    "<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":
-      "<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>",
-    "A new version of the app is available. Please reload to update.":
-      "A new version of the app is available. Please reload to update.",
-    About: "About",
-    "Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":
-      "Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.",
-    "Accessible from any device": "Accessible from any device",
-    Account: "Account",
-    "Activate your account": "Activate your account",
-    Advanced: "Advanced",
-    Amount: "Amount",
-    "An error occurred. Try resubmitting or email {0} directly.": [
-      "An error occurred. Try resubmitting or email ",
-      ["0"],
-      " directly.",
-    ],
-    Appearance: "Appearance",
-    "Apply Style": "Apply Style",
-    Attributes: "Attributes",
-    "Auto-Layouts": "Auto-Layouts",
-    "Back To Editor": "Back To Editor",
-    "Basic Flowchart": "Basic Flowchart",
-    "Become a Github Sponsor": "Become a Github Sponsor",
-    "Become a Pro User": "Become a Pro User",
-    "Begin with a simple example showing how <0>Flowchart Fun</0> works.":
-      "Begin with a simple example showing how <0>Flowchart Fun</0> works.",
-    Blank: "Blank",
-    Blog: "Blog",
-    "Book a Meeting": "Book a Meeting",
-    Border: "Border",
-    "Bottom to Top": "Bottom to Top",
-    Breadthfirst: "Breadthfirst",
-    Cancel: "Cancel",
-    "Cancel your subscription. Your hosted charts will become read-only.":
-      "Cancel your subscription. Your hosted charts will become read-only.",
-    "Certain attributes can be used to customize the appearance or functionality of elements.":
-      "Certain attributes can be used to customize the appearance or functionality of elements.",
-    "Change Email Address": "Change Email Address",
-    Changelog: "Changelog",
-    Charts: "Charts",
-    "Check your email for a link to log in.<0/>You can close this window.":
-      "Check your email for a link to log in.<0/>You can close this window.",
-    "Choose a Theme": "Choose a Theme",
-    "Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":
-      "Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .",
-    "Choose the plan that's right for you and start creating amazing flowcharts with Flowchart Fun Pro":
-      "Choose the plan that's right for you and start creating amazing flowcharts with Flowchart Fun Pro",
-    Circle: "Circle",
-    Classes: "Classes",
-    Clay: "Clay",
-    Clear: "Clear",
-    "Clear text?": "Clear text?",
-    "Click on \\nLearn Syntax\\nto learn more":
-      "Click on \\nLearn Syntax\\nto learn more",
-    "Click to preview": "Click to preview",
-    Clone: "Clone",
-    Close: "Close",
-    Color: "Color",
-    "Colors include red, orange, yellow, blue, purple, black, white, and gray.":
-      "Colors include red, orange, yellow, blue, purple, black, white, and gray.",
-    Column: "Column",
-    "Comic Book": "Comic Book",
-    Comment: "Comment",
-    Concentric: "Concentric",
-    "Confirm New Email": "Confirm New Email",
-    Containers: "Containers",
-    "Containers are nodes that contain other nodes. They are declared using curly braces.":
-      "Containers are nodes that contain other nodes. They are declared using curly braces.",
-    Continue: "Continue",
-    "Convert this to a hosted chart?": "Convert this to a hosted chart?",
-    "Convert to hosted chart?": "Convert to hosted chart?",
-    Copy: "Copy",
-    "Copy PNG Image": "Copy PNG Image",
-    "Copy SVG Code": "Copy SVG Code",
-    "Copy your mermaid.js code or open it directly in the mermaid.js live editor.":
-      "Copy your mermaid.js code or open it directly in the mermaid.js live editor.",
-    "Create Flowcharts from a Prompt with AI":
-      "Create Flowcharts from a Prompt with AI",
-    "Create New Flowchart": "Create New Flowchart",
-    "Create a New Flowchart": "Create a New Flowchart",
-    "Create professional-quality visual aids for presentations or training materials":
-      "Create professional-quality visual aids for presentations or training materials",
-    "Creating an edge between two nodes is done by indenting the second node below the first":
-      "Creating an edge between two nodes is done by indenting the second node below the first",
-    "Custom Sharing Options": "Custom Sharing Options",
-    "Customer Portal": "Customer Portal",
-    "Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":
-      "Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.",
-    Dark: "Dark",
-    "Dark Mode": "Dark Mode",
-    Date: "Date",
-    "Deleted when browser data is cleared":
-      "Deleted when browser data is cleared",
-    "Describe the flowchart you wish to create":
-      "Describe the flowchart you wish to create",
-    Direction: "Direction",
-    Discard: "Discard",
-    "Do you want to cancel your subscription?":
-      "Do you want to cancel your subscription?",
-    "Do you want to delete this?": "Do you want to delete this?",
-    Document: "Document",
-    "Don't have an account?": "Don't have an account?",
-    Download: "Download",
-    "Download JPG": "Download JPG",
-    "Download PNG": "Download PNG",
-    "Download SVG": "Download SVG",
-    "Drag and drop a CSV file here, or click to select a file":
-      "Drag and drop a CSV file here, or click to select a file",
-    "Draw an edge from multiple nodes by beginning the line with a reference":
-      "Draw an edge from multiple nodes by beginning the line with a reference",
-    "Drop the file here ...": "Drop the file here ...",
-    "Easily import existing flowcharts from other software":
-      "Easily import existing flowcharts from other software",
-    Edge: "Edge",
-    "Edge ID, Classes, Attributes": "Edge ID, Classes, Attributes",
-    "Edge Label": "Edge Label",
-    "Edge Label Column": "Edge Label Column",
-    "Edge Style": "Edge Style",
-    "Edge missing indentation": "Edge missing indentation",
-    Edges: "Edges",
-    "Edges are declared in the same row as their source node":
-      "Edges are declared in the same row as their source node",
-    "Edges are declared in the same row as their target node":
-      "Edges are declared in the same row as their target node",
-    "Edges are declared in their own row":
-      "Edges are declared in their own row",
-    "Edges can also have ID's, classes, and attributes before the label":
-      "Edges can also have ID's, classes, and attributes before the label",
-    "Edges can be styled with dashed, dotted, or solid lines":
-      "Edges can be styled with dashed, dotted, or solid lines",
-    "Edges in Separate Rows": "Edges in Separate Rows",
-    "Edges in Source Node Row": "Edges in Source Node Row",
-    "Edges in Target Node Row": "Edges in Target Node Row",
-    Editable: "Editable",
-    Editor: "Editor",
-    Eggs: "Eggs",
-    Email: "Email",
-    Empty: "Empty",
-    "Enter your email": "Enter your email",
-    "Equal To": "Equal To",
-    "Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":
-      "Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages",
-    "Everything in Free": "Everything in Free",
-    Excalidraw: "Excalidraw",
-    Export: "Export",
-    "Export High-Resolution Images": "Export High-Resolution Images",
-    "Export to SVG": "Export to SVG",
-    "Extra Large": "Extra Large",
-    Extract: "Extract",
-    Feedback: "Feedback",
-    "Follow Us on Twitter": "Follow Us on Twitter",
-    "Found a bug? Have a feature request? We would love to hear from you!":
-      "Found a bug? Have a feature request? We would love to hear from you!",
-    Free: "Free",
-    Fullscreen: "Fullscreen",
-    Futuristic: "Futuristic",
-    "Get More Layouts": "Get More Layouts",
-    "Get personalized attention and support from our expert team":
-      "Get personalized attention and support from our expert team",
-    "Getting Started": "Getting Started",
-    Graph: "Graph",
-    Grid: "Grid",
-    "Have complex questions or issues? We're here to help.":
-      "Have complex questions or issues? We're here to help.",
-    Help: "Help",
-    History: "History",
-    Home: "Home",
-    "How are edges declared in this data?":
-      "How are edges declared in this data?",
-    "ID's": "ID's",
-    "If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":
-      "If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project",
-    "If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\:</0>":
-      "If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\:</0>",
-    "Image Export": "Image Export",
-    Images: "Images",
-    "Import Data": "Import Data",
-    "Import data from a CSV file.": "Import data from a CSV file.",
-    "Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":
-      "Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.",
-    "Import from Lucidchart, Visio, and other CSV files":
-      "Import from Lucidchart, Visio, and other CSV files",
-    "Import your diagram it into Microsoft Visio using one of these CSV files.":
-      "Import your diagram it into Microsoft Visio using one of these CSV files.",
-    "Important Changes are Coming to Flowchart Fun.":
-      "Important Changes are Coming to Flowchart Fun.",
-    "Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":
-      "Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.",
-    "Impress your audience with high-quality images produced by Flowchart Fun":
-      "Impress your audience with high-quality images produced by Flowchart Fun",
-    "Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":
-      "Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:",
-    "Indenting creates an edge": "Indenting creates an edge",
-    Info: "Info",
-    Instruct: "Instruct",
-    Is: "Is",
-    Language: "Language",
-    Large: "Large",
-    Layout: "Layout",
-    "Layout Frozen": "Layout Frozen",
-    "Layout is Frozen": "Layout is Frozen",
-    "Leading References": "Leading References",
-    "Learn More": "Learn More",
-    "Learn Syntax": "Learn Syntax",
-    "Learn about Flowchart Fun Pro": "Learn about Flowchart Fun Pro",
-    "Left to Right": "Left to Right",
-    "Let's get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":
-      "Let's get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.",
-    Light: "Light",
-    "Light Mode": "Light Mode",
-    Link: "Link",
-    "Load Theme": "Load Theme",
-    "Loading a theme will erase any custom styles you have applied":
-      "Loading a theme will erase any custom styles you have applied",
-    "Lock Zoom to Graph": "Lock Zoom to Graph",
-    "Log In": "Log In",
-    "Log Out": "Log Out",
-    "Log in to start creating flowcharts.":
-      "Log in to start creating flowcharts.",
-    "Make a One-Time Donation": "Make a One-Time Donation",
-    "Make publicly accessible": "Make publicly accessible",
-    "Make sure you use the same email you will use to log in.":
-      "Make sure you use the same email you will use to log in.",
-    "Map Data": "Map Data",
-    "Market understanding and competitive landscape maintenance for SaaS product development":
-      "Market understanding and competitive landscape maintenance for SaaS product development",
-    Medium: "Medium",
-    Method: "Method",
-    Monospace: "Monospace",
-    Monthly: "Monthly",
-    "Multiple pointers on same line": "Multiple pointers on same line",
-    Museum: "Museum",
-    Name: "Name",
-    New: "New",
-    "New Email": "New Email",
-    "Next charge": "Next charge",
-    "No Edges": "No Edges",
-    Node: "Node",
-    "Node Border Style": "Node Border Style",
-    "Node Colors": "Node Colors",
-    "Node ID": "Node ID",
-    "Node ID, Classes, Attributes": "Node ID, Classes, Attributes",
-    "Node Label": "Node Label",
-    "Node Shapes": "Node Shapes",
-    "Node and pointer on same line": "Node and pointer on same line",
-    "Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":
-      "Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.",
-    "Not Empty": "Not Empty",
-    "Office Hours": "Office Hours",
-    "Once in a while the magic link will end up in your spam folder. If you don't see it after a few minutes, check there or request a new link.":
-      "Once in a while the magic link will end up in your spam folder. If you don't see it after a few minutes, check there or request a new link.",
-    "One-on-One Support": "One-on-One Support",
-    "One-time Share Links": "One-time Share Links",
-    "Only available on this device.<0/>Clearing your browser cache will erase them.":
-      "Only available on this device.<0/>Clearing your browser cache will erase them.",
-    "Open Customer Portal": "Open Customer Portal",
-    "Organization Chart": "Organization Chart",
-    "Paste the information you wish to convert to a flowchart":
-      "Paste the information you wish to convert to a flowchart",
-    Permalinks: "Permalinks",
-    "Permanent Charts are a Pro Feature": "Permanent Charts are a Pro Feature",
-    "Permanent Flowcharts": "Permanent Flowcharts",
-    Persistent: "Persistent",
-    "Persistent Flowcharts": "Persistent Flowcharts",
-    Playbook: "Playbook",
-    "Pointer and container on same line": "Pointer and container on same line",
-    Pricing: "Pricing",
-    "Process for corporate social responsibility initiatives development and implementation across company operations":
-      "Process for corporate social responsibility initiatives development and implementation across company operations",
-    "Processing Data": "Processing Data",
-    Prompt: "Prompt",
-    Public: "Public",
-    Random: "Random",
-    "Read-only": "Read-only",
-    "Reference by Class": "Reference by Class",
-    "Reference by ID": "Reference by ID",
-    "Reference by Label": "Reference by Label",
-    References: "References",
-    "References are used to create edges between nodes that are created elsewhere in the document":
-      "References are used to create edges between nodes that are created elsewhere in the document",
-    "Referencing a node by its exact label":
-      "Referencing a node by its exact label",
-    "Referencing a node by its unique ID":
-      "Referencing a node by its unique ID",
-    "Referencing multiple nodes with the same assigned class":
-      "Referencing multiple nodes with the same assigned class",
-    "Refresh Page": "Refresh Page",
-    "Reload to Update": "Reload to Update",
-    "Remove Image Watermark": "Remove Image Watermark",
-    Rename: "Rename",
-    Reset: "Reset",
-    "Resume Subscription": "Resume Subscription",
-    Retro: "Retro",
-    Return: "Return",
-    "Right to Left": "Right to Left",
-    Roadmap: "Roadmap",
-    Save: "Save",
-    "Save 20% (2 months free!)": "Save 20% (2 months free!)",
-    "Save time and automate certain tasks with this cutting-edge feature":
-      "Save time and automate certain tasks with this cutting-edge feature",
-    "Save time and streamline your workflow by importing data from other sources":
-      "Save time and streamline your workflow by importing data from other sources",
-    "Select your plan!": "Select your plan!",
-    "Send Feedback": "Send Feedback",
-    Settings: "Settings",
-    Shape: "Shape",
-    "Sign Up": "Sign Up",
-    "Sign Up Now": "Sign Up Now",
-    Size: "Size",
-    Small: "Small",
-    "Some classes are available to help style your flowchart":
-      "Some classes are available to help style your flowchart",
-    "Sorry! This page is only available in English.":
-      "Sorry! This page is only available in English.",
-    "Source Arrow": "Source Arrow",
-    "Source Column": "Source Column",
-    "Source Delimiter": "Source Delimiter",
-    "Source/Target Arrow Shape": "Source/Target Arrow Shape",
-    Spacing: "Spacing",
-    "Special Attributes": "Special Attributes",
-    Start: "Start",
-    "Start Over": "Start Over",
-    Status: "Status",
-    "Store any data associated to a node":
-      "Store any data associated to a node",
-    "Stored in the cloud": "Stored in the cloud",
-    "Stored on this computer": "Stored on this computer",
-    "Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":
-      "Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!",
-    Style: "Style",
-    "Style Classes": "Style Classes",
-    Submit: "Submit",
-    Subscribe: "Subscribe",
-    Subscription: "Subscription",
-    "Subscription will end": "Subscription will end",
-    "Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":
-      "Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration",
-    "Target Arrow": "Target Arrow",
-    "Target Column": "Target Column",
-    "Target Delimiter": "Target Delimiter",
-    Temporary: "Temporary",
-    "Temporary Flowchart Warning": "Temporary Flowchart Warning",
-    "Temporary Flowcharts": "Temporary Flowcharts",
-    "Text Sizes": "Text Sizes",
-    "Text followed by colon+space creates an edge with the text as the label":
-      "Text followed by colon+space creates an edge with the text as the label",
-    "Text on a line creates a node with the text as the label":
-      "Text on a line creates a node with the text as the label",
-    "Text sizes include small, regular, large, and extra-large.":
-      "Text sizes include small, regular, large, and extra-large.",
-    "Thank you for your feedback!": "Thank you for your feedback!",
-    "The column that contains the edge label(s)":
-      "The column that contains the edge label(s)",
-    "The column that contains the source node ID(s)":
-      "The column that contains the source node ID(s)",
-    "The column that contains the target node ID(s)":
-      "The column that contains the target node ID(s)",
-    "The delimiter used to separate multiple source nodes":
-      "The delimiter used to separate multiple source nodes",
-    "The delimiter used to separate multiple target nodes":
-      "The delimiter used to separate multiple target nodes",
-    "The possible shapes are:": "The possible shapes are:",
-    "Theme Editor": "Theme Editor",
-    "There are no edges in this data": "There are no edges in this data",
-    "This action cannot be undone.": "This action cannot be undone.",
-    "This app works by typing": "This app works by typing",
-    "This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":
-      "This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!",
-    "This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":
-      "This feature is only available to pro users. <0>Become a pro user</0> to unlock it.",
-    "This is a temporary chart, it will be deleted when you clear your browser cache.":
-      "This is a temporary chart, it will be deleted when you clear your browser cache.",
-    Tip: "Tip",
-    "To fix this change one of the edge IDs":
-      "To fix this change one of the edge IDs",
-    "To fix this change one of the node IDs":
-      "To fix this change one of the node IDs",
-    "To fix this move one pointer to the next line":
-      "To fix this move one pointer to the next line",
-    "To fix this move the pointer to the next line":
-      "To fix this move the pointer to the next line",
-    "To fix this start the container <0/> on a different line":
-      "To fix this start the container <0/> on a different line",
-    "Top to Bottom": "Top to Bottom",
-    "Try again": "Try again",
-    "Two edges have the same ID": "Two edges have the same ID",
-    "Two nodes have the same ID": "Two nodes have the same ID",
-    Type: "Type",
-    Unfreeze: "Unfreeze",
-    "Unique text value to identify a node":
-      "Unique text value to identify a node",
-    "Unknown Parsing Error": "Unknown Parsing Error",
-    "Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":
-      "Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!",
-    "Update Email": "Update Email",
-    "Upgrade to Pro": "Upgrade to Pro",
-    "Upload your File": "Upload your File",
-    "Use AI to create flowcharts from data sets or algorithms":
-      "Use AI to create flowcharts from data sets or algorithms",
-    "Use AI to generate a flowchart from a prompt. <0>Learn More</0>":
-      "Use AI to generate a flowchart from a prompt. <0>Learn More</0>",
-    "Use classes to group nodes": "Use classes to group nodes",
-    "Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":
-      "Use the attribute <0>href</0> to set a link on a node that opens in a new tab.",
-    "Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":
-      "Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.",
-    "Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":
-      "Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.",
-    "Use the customer portal to change your billing information.":
-      "Use the customer portal to change your billing information.",
-    "Use this file for org charts, hierarchies, and other organizational structures.":
-      "Use this file for org charts, hierarchies, and other organizational structures.",
-    "Use this file for sequences, processes, and workflows.":
-      "Use this file for sequences, processes, and workflows.",
-    User: "User",
-    "View on Github": "View on Github",
-    "Water evaporates from the Earth's surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":
-      "Water evaporates from the Earth's surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.",
-    "We use magic links to log you in. Enter your email below to get started.":
-      "We use magic links to log you in. Enter your email below to get started.",
-    "Welcome to Flowchart Fun Pro!": "Welcome to Flowchart Fun Pro!",
-    "Width and Height": "Width and Height",
-    "Would you like to continue?": "Would you like to continue?",
-    Yearly: "Yearly",
-    "You already have a flowchart with this name.":
-      "You already have a flowchart with this name.",
-    "You are about to add {numNodes} nodes and {numEdges} edges to your graph.":
-      [
-        "You are about to add ",
-        ["numNodes"],
-        " nodes and ",
-        ["numEdges"],
-        " edges to your graph.",
-      ],
-    'You can also wrap text in "\\( \\)" to create an edge':
-      'You can also wrap text in "\\( \\)" to create an edge',
-    "You must log in to create a persistent flowchart.":
-      "You must log in to create a persistent flowchart.",
-    "You're almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":
-      "You're almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.",
-    "Your Charts": "Your Charts",
-    "Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":
-      "Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.",
-    "Zoom In": "Zoom In",
-    "Zoom Out": "Zoom Out",
-    "any text: before a colon creates a label":
-      "any text: before a colon creates a label",
-  },
+  messages: JSON.parse(
+    '{"$3 per month":"$3 per month","$30 per year":"$30 per year","(before a colon creates a label)":"(before a colon creates a label)","<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":"<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>","<0>Sign In</0> / <1>Sign Up</1> with email and password":"<0>Sign In</0> / <1>Sign Up</1> with email and password","<0>Starting August 28th you will need a subscription to create and edit charts. <1>Learn more</1>.</0>":"<0>Starting August 28th you will need a subscription to create and edit charts. <1>Learn more</1>.</0>","<0>You currently have a free account.<1/><2>Learn about our Pro Features and subscribe on our pricing page</2>.</0>":"<0>You currently have a free account.<1/><2>Learn about our Pro Features and subscribe on our pricing page</2>.</0>","A new version of the app is available. Please reload to update.":"A new version of the app is available. Please reload to update.","About":"About","Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":"Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.","Accessible from any device":"Accessible from any device","Account":"Account","Activate your account":"Activate your account","Advanced":"Advanced","Amount":"Amount","An error occurred. Try resubmitting or email {0} directly.":["An error occurred. Try resubmitting or email ",["0"]," directly."],"Appearance":"Appearance","Apply Style":"Apply Style","Attributes":"Attributes","Auto-Layouts":"Auto-Layouts","Back To Editor":"Back To Editor","Basic Flowchart":"Basic Flowchart","Become a Github Sponsor":"Become a Github Sponsor","Become a Pro User":"Become a Pro User","Begin with a simple example showing how <0>Flowchart Fun</0> works.":"Begin with a simple example showing how <0>Flowchart Fun</0> works.","Blank":"Blank","Blog":"Blog","Book a Meeting":"Book a Meeting","Border":"Border","Bottom to Top":"Bottom to Top","Breadthfirst":"Breadthfirst","Cancel":"Cancel","Cancel your subscription. Your hosted charts will become read-only.":"Cancel your subscription. Your hosted charts will become read-only.","Certain attributes can be used to customize the appearance or functionality of elements.":"Certain attributes can be used to customize the appearance or functionality of elements.","Change Email Address":"Change Email Address","Changelog":"Changelog","Charts":"Charts","Check your email for a link to log in.<0/>You can close this window.":"Check your email for a link to log in.<0/>You can close this window.","Choose":"Choose","Choose a Theme":"Choose a Theme","Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":"Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .","Choose the plan that\'s right for you and start creating amazing flowcharts with Flowchart Fun Pro":"Choose the plan that\'s right for you and start creating amazing flowcharts with Flowchart Fun Pro","Circle":"Circle","Classes":"Classes","Clay":"Clay","Clear":"Clear","Clear text?":"Clear text?","Click on \\\\nLearn Syntax\\\\nto learn more":"Click on \\\\nLearn Syntax\\\\nto learn more","Click to preview":"Click to preview","Clone":"Clone","Close":"Close","Color":"Color","Colors include red, orange, yellow, blue, purple, black, white, and gray.":"Colors include red, orange, yellow, blue, purple, black, white, and gray.","Column":"Column","Comic Book":"Comic Book","Comment":"Comment","Concentric":"Concentric","Confirm New Email":"Confirm New Email","Confirm your email address to sign in.":"Confirm your email address to sign in.","Containers":"Containers","Containers are nodes that contain other nodes. They are declared using curly braces.":"Containers are nodes that contain other nodes. They are declared using curly braces.","Continue":"Continue","Convert this to a hosted chart?":"Convert this to a hosted chart?","Convert to hosted chart?":"Convert to hosted chart?","Copy":"Copy","Copy PNG Image":"Copy PNG Image","Copy SVG Code":"Copy SVG Code","Copy your mermaid.js code or open it directly in the mermaid.js live editor.":"Copy your mermaid.js code or open it directly in the mermaid.js live editor.","Create Flowcharts from a Prompt with AI":"Create Flowcharts from a Prompt with AI","Create New Flowchart":"Create New Flowchart","Create a New Flowchart":"Create a New Flowchart","Create professional-quality visual aids for presentations or training materials":"Create professional-quality visual aids for presentations or training materials","Creating an edge between two nodes is done by indenting the second node below the first":"Creating an edge between two nodes is done by indenting the second node below the first","Custom Sharing Options":"Custom Sharing Options","Customer Portal":"Customer Portal","Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":"Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.","Dark":"Dark","Dark Mode":"Dark Mode","Date":"Date","Deleted when browser data is cleared":"Deleted when browser data is cleared","Describe the flowchart you wish to create":"Describe the flowchart you wish to create","Direction":"Direction","Discard":"Discard","Do you want to cancel your subscription?":"Do you want to cancel your subscription?","Do you want to delete this?":"Do you want to delete this?","Document":"Document","Download":"Download","Download JPG":"Download JPG","Download PNG":"Download PNG","Download SVG":"Download SVG","Drag and drop a CSV file here, or click to select a file":"Drag and drop a CSV file here, or click to select a file","Draw an edge from multiple nodes by beginning the line with a reference":"Draw an edge from multiple nodes by beginning the line with a reference","Drop the file here ...":"Drop the file here ...","Easily import existing flowcharts from other software":"Easily import existing flowcharts from other software","Edge":"Edge","Edge ID, Classes, Attributes":"Edge ID, Classes, Attributes","Edge Label":"Edge Label","Edge Label Column":"Edge Label Column","Edge Style":"Edge Style","Edge missing indentation":"Edge missing indentation","Edges":"Edges","Edges are declared in the same row as their source node":"Edges are declared in the same row as their source node","Edges are declared in the same row as their target node":"Edges are declared in the same row as their target node","Edges are declared in their own row":"Edges are declared in their own row","Edges can also have ID\'s, classes, and attributes before the label":"Edges can also have ID\'s, classes, and attributes before the label","Edges can be styled with dashed, dotted, or solid lines":"Edges can be styled with dashed, dotted, or solid lines","Edges in Separate Rows":"Edges in Separate Rows","Edges in Source Node Row":"Edges in Source Node Row","Edges in Target Node Row":"Edges in Target Node Row","Editable":"Editable","Editor":"Editor","Eggs":"Eggs","Email":"Email","Empty":"Empty","Enter your email":"Enter your email","Enter your email address and we\'ll send you a magic link to sign in.":"Enter your email address and we\'ll send you a magic link to sign in.","Enter your email address below and we\'ll send you a link to reset your password.":"Enter your email address below and we\'ll send you a link to reset your password.","Equal To":"Equal To","Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":"Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages","Everything in Free":"Everything in Free","Excalidraw":"Excalidraw","Export":"Export","Export High-Resolution Images":"Export High-Resolution Images","Export to SVG":"Export to SVG","Extra Large":"Extra Large","Extract":"Extract","Feedback":"Feedback","Follow Us on Twitter":"Follow Us on Twitter","Forgot your password?":"Forgot your password?","Found a bug? Have a feature request? We would love to hear from you!":"Found a bug? Have a feature request? We would love to hear from you!","Free":"Free","Fullscreen":"Fullscreen","Futuristic":"Futuristic","Get More Layouts":"Get More Layouts","Get personalized attention and support from our expert team":"Get personalized attention and support from our expert team","Getting Started":"Getting Started","Graph":"Graph","Grid":"Grid","Have complex questions or issues? We\'re here to help.":"Have complex questions or issues? We\'re here to help.","Help":"Help","History":"History","Home":"Home","How are edges declared in this data?":"How are edges declared in this data?","ID\'s":"ID\'s","If an account with that email exists, we\'ve sent you an email with instructions on how to reset your password.":"If an account with that email exists, we\'ve sent you an email with instructions on how to reset your password.","If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":"If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project","If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\\\:</0>":"If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\\\:</0>","Image Export":"Image Export","Images":"Images","Import Data":"Import Data","Import data from a CSV file.":"Import data from a CSV file.","Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":"Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.","Import from Lucidchart, Visio, and other CSV files":"Import from Lucidchart, Visio, and other CSV files","Import your diagram it into Microsoft Visio using one of these CSV files.":"Import your diagram it into Microsoft Visio using one of these CSV files.","Important Changes are Coming to Flowchart Fun.":"Important Changes are Coming to Flowchart Fun.","Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":"Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.","Impress your audience with high-quality images produced by Flowchart Fun":"Impress your audience with high-quality images produced by Flowchart Fun","Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":"Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:","Indenting creates an edge":"Indenting creates an edge","Info":"Info","Instruct":"Instruct","Is":"Is","Language":"Language","Large":"Large","Layout":"Layout","Layout Frozen":"Layout Frozen","Layout is Frozen":"Layout is Frozen","Leading References":"Leading References","Learn More":"Learn More","Learn Syntax":"Learn Syntax","Learn about Flowchart Fun Pro":"Learn about Flowchart Fun Pro","Left to Right":"Left to Right","Let\'s get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":"Let\'s get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.","Light":"Light","Light Mode":"Light Mode","Link":"Link","Load Theme":"Load Theme","Loading a theme will erase any custom styles you have applied":"Loading a theme will erase any custom styles you have applied","Lock Zoom to Graph":"Lock Zoom to Graph","Log In":"Log In","Log Out":"Log Out","Log in to start creating flowcharts.":"Log in to start creating flowcharts.","Make a One-Time Donation":"Make a One-Time Donation","Make publicly accessible":"Make publicly accessible","Make sure you use the same email you will use to log in.":"Make sure you use the same email you will use to log in.","Map Data":"Map Data","Market understanding and competitive landscape maintenance for SaaS product development":"Market understanding and competitive landscape maintenance for SaaS product development","Medium":"Medium","Method":"Method","Monospace":"Monospace","Monthly":"Monthly","Multiple pointers on same line":"Multiple pointers on same line","Museum":"Museum","Name":"Name","New":"New","New Email":"New Email","Next charge":"Next charge","No Edges":"No Edges","Node":"Node","Node Border Style":"Node Border Style","Node Colors":"Node Colors","Node ID":"Node ID","Node ID, Classes, Attributes":"Node ID, Classes, Attributes","Node Label":"Node Label","Node Shapes":"Node Shapes","Node and pointer on same line":"Node and pointer on same line","Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":"Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.","Not Empty":"Not Empty","Office Hours":"Office Hours","Once in a while the magic link will end up in your spam folder. If you don\'t see it after a few minutes, check there or request a new link.":"Once in a while the magic link will end up in your spam folder. If you don\'t see it after a few minutes, check there or request a new link.","One-on-One Support":"One-on-One Support","One-time Share Links":"One-time Share Links","Only available on this device.<0/>Clearing your browser cache will erase them.":"Only available on this device.<0/>Clearing your browser cache will erase them.","Open Customer Portal":"Open Customer Portal","Organization Chart":"Organization Chart","Password":"Password","Paste the information you wish to convert to a flowchart":"Paste the information you wish to convert to a flowchart","Permalinks":"Permalinks","Permanent Charts are a Pro Feature":"Permanent Charts are a Pro Feature","Permanent Flowcharts":"Permanent Flowcharts","Persistent":"Persistent","Persistent Flowcharts":"Persistent Flowcharts","Playbook":"Playbook","Pointer and container on same line":"Pointer and container on same line","Pricing":"Pricing","Process for corporate social responsibility initiatives development and implementation across company operations":"Process for corporate social responsibility initiatives development and implementation across company operations","Processing Data":"Processing Data","Prompt":"Prompt","Public":"Public","Random":"Random","Read-only":"Read-only","Reference by Class":"Reference by Class","Reference by ID":"Reference by ID","Reference by Label":"Reference by Label","References":"References","References are used to create edges between nodes that are created elsewhere in the document":"References are used to create edges between nodes that are created elsewhere in the document","Referencing a node by its exact label":"Referencing a node by its exact label","Referencing a node by its unique ID":"Referencing a node by its unique ID","Referencing multiple nodes with the same assigned class":"Referencing multiple nodes with the same assigned class","Refresh Page":"Refresh Page","Reload to Update":"Reload to Update","Remove Image Watermark":"Remove Image Watermark","Rename":"Rename","Request Magic Link":"Request Magic Link","Request Password Reset":"Request Password Reset","Reset":"Reset","Reset Password":"Reset Password","Resume Subscription":"Resume Subscription","Retro":"Retro","Return":"Return","Right to Left":"Right to Left","Roadmap":"Roadmap","Save":"Save","Save 20% (2 months free!)":"Save 20% (2 months free!)","Save time and automate certain tasks with this cutting-edge feature":"Save time and automate certain tasks with this cutting-edge feature","Save time and streamline your workflow by importing data from other sources":"Save time and streamline your workflow by importing data from other sources","Select your plan!":"Select your plan!","Send Feedback":"Send Feedback","Settings":"Settings","Shape":"Shape","Sign In":"Sign In","Sign Up":"Sign Up","Sign in with <0>GitHub</0>":"Sign in with <0>GitHub</0>","Sign in with <0>Google</0>":"Sign in with <0>Google</0>","Size":"Size","Small":"Small","Some classes are available to help style your flowchart":"Some classes are available to help style your flowchart","Sorry! This page is only available in English.":"Sorry! This page is only available in English.","Source Arrow":"Source Arrow","Source Column":"Source Column","Source Delimiter":"Source Delimiter","Source/Target Arrow Shape":"Source/Target Arrow Shape","Spacing":"Spacing","Special Attributes":"Special Attributes","Start":"Start","Start Over":"Start Over","Status":"Status","Store any data associated to a node":"Store any data associated to a node","Stored in the cloud":"Stored in the cloud","Stored on this computer":"Stored on this computer","Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":"Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!","Style":"Style","Style Classes":"Style Classes","Submit":"Submit","Subscribe":"Subscribe","Subscription":"Subscription","Subscription will end":"Subscription will end","Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":"Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration","Target Arrow":"Target Arrow","Target Column":"Target Column","Target Delimiter":"Target Delimiter","Temporary":"Temporary","Temporary Flowchart Warning":"Temporary Flowchart Warning","Temporary Flowcharts":"Temporary Flowcharts","Text Sizes":"Text Sizes","Text followed by colon+space creates an edge with the text as the label":"Text followed by colon+space creates an edge with the text as the label","Text on a line creates a node with the text as the label":"Text on a line creates a node with the text as the label","Text sizes include small, regular, large, and extra-large.":"Text sizes include small, regular, large, and extra-large.","Thank you for your feedback!":"Thank you for your feedback!","The column that contains the edge label(s)":"The column that contains the edge label(s)","The column that contains the source node ID(s)":"The column that contains the source node ID(s)","The column that contains the target node ID(s)":"The column that contains the target node ID(s)","The delimiter used to separate multiple source nodes":"The delimiter used to separate multiple source nodes","The delimiter used to separate multiple target nodes":"The delimiter used to separate multiple target nodes","The possible shapes are:":"The possible shapes are:","Theme Editor":"Theme Editor","There are no edges in this data":"There are no edges in this data","This action cannot be undone.":"This action cannot be undone.","This app works by typing":"This app works by typing","This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":"This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!","This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":"This feature is only available to pro users. <0>Become a pro user</0> to unlock it.","This is a temporary chart, it will be deleted when you clear your browser cache.":"This is a temporary chart, it will be deleted when you clear your browser cache.","Tip":"Tip","To fix this change one of the edge IDs":"To fix this change one of the edge IDs","To fix this change one of the node IDs":"To fix this change one of the node IDs","To fix this move one pointer to the next line":"To fix this move one pointer to the next line","To fix this move the pointer to the next line":"To fix this move the pointer to the next line","To fix this start the container <0/> on a different line":"To fix this start the container <0/> on a different line","To learn more about why we require you to log in, please read <0>this blog post</0>.":"To learn more about why we require you to log in, please read <0>this blog post</0>.","Top to Bottom":"Top to Bottom","Try again":"Try again","Two edges have the same ID":"Two edges have the same ID","Two nodes have the same ID":"Two nodes have the same ID","Type":"Type","Unfreeze":"Unfreeze","Unique text value to identify a node":"Unique text value to identify a node","Unknown Parsing Error":"Unknown Parsing Error","Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":"Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!","Update Email":"Update Email","Upgrade to Pro":"Upgrade to Pro","Upload your File":"Upload your File","Use AI to create flowcharts from data sets or algorithms":"Use AI to create flowcharts from data sets or algorithms","Use AI to generate a flowchart from a prompt. <0>Learn More</0>":"Use AI to generate a flowchart from a prompt. <0>Learn More</0>","Use classes to group nodes":"Use classes to group nodes","Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":"Use the attribute <0>href</0> to set a link on a node that opens in a new tab.","Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":"Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.","Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":"Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.","Use the customer portal to change your billing information.":"Use the customer portal to change your billing information.","Use this file for org charts, hierarchies, and other organizational structures.":"Use this file for org charts, hierarchies, and other organizational structures.","Use this file for sequences, processes, and workflows.":"Use this file for sequences, processes, and workflows.","User":"User","View on Github":"View on Github","Water evaporates from the Earth\'s surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":"Water evaporates from the Earth\'s surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.","Welcome to Flowchart Fun Pro!":"Welcome to Flowchart Fun Pro!","Width and Height":"Width and Height","Would you like to continue?":"Would you like to continue?","Yearly":"Yearly","You already have a flowchart with this name.":"You already have a flowchart with this name.","You are about to add {numNodes} nodes and {numEdges} edges to your graph.":["You are about to add ",["numNodes"]," nodes and ",["numEdges"]," edges to your graph."],"You can also wrap text in \\"\\\\( \\\\)\\" to create an edge":"You can also wrap text in \\"\\\\( \\\\)\\" to create an edge","You must log in to create a persistent flowchart.":"You must log in to create a persistent flowchart.","You need to log in to access this page.":"You need to log in to access this page.","You\'re almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":"You\'re almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.","Your Charts":"Your Charts","Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":"Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.","Zoom In":"Zoom In","Zoom Out":"Zoom Out","any text: before a colon creates a label":"any text: before a colon creates a label","or":"or"}'
+  ),
 };
