@@ -1,461 +1,5 @@
 /*eslint-disable*/ module.exports = {
-  messages: {
-    "$3 per month": "3 $ par mois",
-    "$30 per year": "30 $ par an",
-    "(before a colon creates a label)":
-      "(avant que deux-points ne créent une étiquette)",
-    "<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":
-      "<0>Flowchart Fun</0> est un projet open source réalisé par <1>Tone Row</1>",
-    "A new version of the app is available. Please reload to update.":
-      "Une nouvelle version de l'application est disponible. Veuillez recharger pour mettre à jour.",
-    About: "À propos",
-    "Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":
-      "Accédez à ces diagrammes depuis n'importe où.<0/>Partagez et intégrez des diagrammes de flux qui restent synchronisés.",
-    "Accessible from any device": "Accessible depuis n'importe quel appareil",
-    Account: "Compte",
-    "Activate your account": "Activez votre compte",
-    Advanced: "Avancé",
-    Amount: "Montant",
-    "An error occurred. Try resubmitting or email {0} directly.": [
-      "Une erreur s'est produite. Essayez de l'envoyer à nouveau ou bien envoyez un e-mail à l'adresse ",
-      ["0"],
-      ".",
-    ],
-    Appearance: "Thème",
-    "Apply Style": "Appliquer un style",
-    Attributes: "Attributs",
-    "Auto-Layouts": "Mises en page automatiques",
-    "Back To Editor": "Retour à l'éditeur",
-    "Basic Flowchart": "Diagramme de flux de base",
-    "Become a Github Sponsor": "Devenez un sponsor Github",
-    "Become a Pro User": "Devenez un utilisateur Pro",
-    "Begin with a simple example showing how <0>Flowchart Fun</0> works.":
-      "Commencez par un exemple simple montrant comment <0>Flowchart Fun</0> fonctionne.",
-    Blank: "Vide",
-    Blog: "Blog",
-    "Book a Meeting": "Réserver une réunion",
-    Border: "Frontière",
-    "Bottom to Top": "De bas en haut",
-    Breadthfirst: "Parcours en largeur",
-    Cancel: "Annuler",
-    "Cancel your subscription. Your hosted charts will become read-only.":
-      "Résilier votre abonnement. Vos graphiques hébergés seront en lecture seule.",
-    "Certain attributes can be used to customize the appearance or functionality of elements.":
-      "Certains attributs peuvent être utilisés pour personnaliser l'apparence ou la fonctionnalité des éléments.",
-    "Change Email Address": "Changer l'adresse email",
-    Changelog: "Journal des modifications",
-    Charts: "Graphiques",
-    "Check your email for a link to log in.<0/>You can close this window.":
-      "Vérifiez votre e-mail pour un lien de connexion. Vous pouvez fermer cette fenêtre.",
-    "Choose a Theme": "Choisir un thème",
-    "Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":
-      "Choisissez parmi une variété de formes de flèches pour la source et la cible d'un bord. Les formes incluent triangle, triangle-tee, cercle-triangle, triangle-croix, triangle-backcurve, vee, tee, carré, cercle, diamant, chevron, aucun.",
-    "Choose the plan that's right for you and start creating amazing flowcharts with Flowchart Fun Pro":
-      "Choisissez le plan qui vous convient le mieux et commencez à créer des diagrammes incroyables avec Flowchart Fun Pro",
-    Circle: "Cercle",
-    Classes: "Classes",
-    Clay: "Argile",
-    Clear: "Effacer",
-    "Clear text?": "Effacer le texte?",
-    "Click on \\nLearn Syntax\\nto learn more":
-      "Cliquez sur \\nApprendre la syntaxe\\npour en savoir plus.",
-    "Click to preview": "Cliquez pour prévisualiser",
-    Clone: "Cloner",
-    Close: "Fermer",
-    Color: "Couleur",
-    "Colors include red, orange, yellow, blue, purple, black, white, and gray.":
-      "Les couleurs incluent le rouge, l'orange, le jaune, le bleu, le pourpre, le noir, le blanc et le gris.",
-    Column: "Colonne",
-    "Comic Book": "Bande dessinée",
-    Comment: "Commenter",
-    Concentric: "Concentrique",
-    "Confirm New Email": "Confirmer le nouveau Email",
-    Containers: "Conteneurs",
-    "Containers are nodes that contain other nodes. They are declared using curly braces.":
-      "Les conteneurs sont des nœuds qui contiennent d'autres nœuds. Ils sont déclarés à l'aide de accolades.",
-    Continue: "Continuer",
-    "Convert this to a hosted chart?": "Convertir ceci en graphique hébergé?",
-    "Convert to hosted chart?": "Convertir en graphique hébergé ?",
-    Copy: "Copier",
-    "Copy PNG Image": "Copier l'image PNG",
-    "Copy SVG Code": "Copier le code SVG",
-    "Copy your mermaid.js code or open it directly in the mermaid.js live editor.":
-      "Copiez votre code mermaid.js ou ouvrez-le directement dans l'éditeur en direct mermaid.js.",
-    "Create Flowcharts from a Prompt with AI":
-      "Créer des diagrammes à partir d'une invite avec l'IA",
-    "Create New Flowchart": "Créer un nouveau diagramme",
-    "Create a New Flowchart": "Créer un nouveau diagramme",
-    "Create professional-quality visual aids for presentations or training materials":
-      "Créez des aides visuelles de qualité professionnelle pour des présentations ou des matériels de formation",
-    "Creating an edge between two nodes is done by indenting the second node below the first":
-      "Créer une arête entre deux nœuds est fait en indentant le second nœud sous le premier",
-    "Custom Sharing Options": "Options de partage personnalisées",
-    "Customer Portal": "Portail Clients",
-    "Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":
-      "Personnalisez votre thème en modifiant le <0>Cytoscape CSS</0> ci-dessous. Notre documentation de mise en forme arrive bientôt ! Entre-temps, la meilleure ressource est la documentation <1>Cytoscape</1>. Venez poser des questions sur le <2>Discord</2> si vous êtes coincé.",
-    Dark: "Sombre",
-    "Dark Mode": "Mode sombre",
-    Date: "Date",
-    "Deleted when browser data is cleared":
-      "Supprimé lorsque les données du navigateur sont effacées",
-    "Describe the flowchart you wish to create":
-      "Décrivez le schéma que vous souhaitez créer",
-    Direction: "Direction",
-    Discard: "Jeter",
-    "Do you want to cancel your subscription?":
-      "Souhaitez-vous résilier votre abonnement ?",
-    "Do you want to delete this?": "Souhaitez-vous supprimer ceci ?",
-    Document: "Document",
-    "Don't have an account?": "Vous n'avez pas de compte ?",
-    Download: "Télécharger",
-    "Download JPG": "Télécharger JPG",
-    "Download PNG": "Télécharger PNG",
-    "Download SVG": "Télécharger SVG",
-    "Drag and drop a CSV file here, or click to select a file":
-      "Glissez-déposez un fichier CSV ici, ou cliquez pour sélectionner un fichier",
-    "Draw an edge from multiple nodes by beginning the line with a reference":
-      "Dessinez une arête à partir de plusieurs nœuds en commençant la ligne par une référence",
-    "Drop the file here ...": "Déposez le fichier ici ...",
-    "Easily import existing flowcharts from other software":
-      "Importez facilement des diagrammes existants à partir d'autres logiciels",
-    Edge: "Bord",
-    "Edge ID, Classes, Attributes": "ID Edge, Classes, Attributs",
-    "Edge Label": "Étiquette Edge",
-    "Edge Label Column": "Colonne d'étiquette Edge",
-    "Edge Style": "Style Edge",
-    "Edge missing indentation": "Indentation manquante du bord",
-    Edges: "Bords",
-    "Edges are declared in the same row as their source node":
-      "Les bords sont déclarés dans la même ligne que leur nœud source",
-    "Edges are declared in the same row as their target node":
-      "Les bords sont déclarés dans la même ligne que leur nœud cible",
-    "Edges are declared in their own row":
-      "Les bords sont déclarés dans leur propre ligne",
-    "Edges can also have ID's, classes, and attributes before the label":
-      "Les bords peuvent également avoir des ID, des classes et des attributs avant l'étiquette",
-    "Edges can be styled with dashed, dotted, or solid lines":
-      "Les bords peuvent être stylisés avec des lignes en pointillés, en pointillés ou en lignes continues",
-    "Edges in Separate Rows": "Bordures en Rangs Séparés",
-    "Edges in Source Node Row": "Bordures dans la Ligne du Nœud Source",
-    "Edges in Target Node Row": "Bordures dans la Ligne du Nœud Cible",
-    Editable: "Modifiable",
-    Editor: "Éditeur",
-    Eggs: "Oeufs",
-    Email: "E-mail",
-    Empty: "Vide",
-    "Enter your email": "Entrez votre adresse e-mail ",
-    "Equal To": "Égal à",
-    "Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":
-      "Diagramme de flux de l'écriture d'essai, guidant les étudiants à travers la réflexion, le schématisme, le brouillon et les étapes de révision",
-    "Everything in Free": "Tout dans le plan gratuit",
-    Excalidraw: "Excalidraw",
-    Export: "Exporter",
-    "Export High-Resolution Images": "Exporter des images haute résolution",
-    "Export to SVG": "Exporter en SVG",
-    "Extra Large": "Extra Large",
-    Extract: "Extrait",
-    Feedback: "Commentaire",
-    "Follow Us on Twitter": "Suivez-nous sur Twitter",
-    "Found a bug? Have a feature request? We would love to hear from you!":
-      "Vous avez trouvé un bug ? Vous souhaitez faire une demande de fonctionnalité ? Nous aimerions beaucoup vous entendre !",
-    Free: "Gratuit ",
-    Fullscreen: "Plein écran",
-    Futuristic: "Futuriste",
-    "Get More Layouts": "Obtenez plus de mises en page",
-    "Get personalized attention and support from our expert team":
-      "Obtenez une attention et un soutien personnalisés de notre équipe d'experts ",
-    "Getting Started": "Commencer",
-    Graph: "Graphique",
-    Grid: "Quadrillage",
-    "Have complex questions or issues? We're here to help.":
-      "Des questions ou des problèmes complexes ? Nous sommes là pour vous aider. ",
-    Help: "Aide",
-    History: "Historique",
-    Home: "Accueil",
-    "How are edges declared in this data?":
-      "Comment les bords sont-ils déclarés dans ces données?",
-    "ID's": "ID",
-    "If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":
-      "Si vous appréciez l'utilisation de <0>Flowchart Fun</0>, veuillez envisager de soutenir le projet",
-    "If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\:</0>":
-      "Si vous souhaitez créer un bord, faites une indentation de cette ligne. Sinon, échappez le deux-points avec une barre oblique <0>\\:</0>",
-    "Image Export": "Exportation d'images",
-    Images: "Images",
-    "Import Data": "Importer des données",
-    "Import data from a CSV file.":
-      "Importer des données à partir d'un fichier CSV.",
-    "Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":
-      "Importer des données à partir de n'importe quel fichier CSV et les mapper à un nouveau diagramme. C'est une excellente façon d'importer des données à partir d'autres sources telles que Lucidchart, Google Sheets et Visio.",
-    "Import from Lucidchart, Visio, and other CSV files":
-      "Importer à partir de Lucidchart, Visio et autres fichiers CSV",
-    "Import your diagram it into Microsoft Visio using one of these CSV files.":
-      "Importez votre diagramme dans Microsoft Visio à l'aide de l'un de ces fichiers CSV.",
-    "Important Changes are Coming to Flowchart Fun.":
-      "Des changements importants arrivent à Flowchart Fun.",
-    "Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":
-      "L'importation de données est une fonctionnalité professionnelle. Vous pouvez passer à Flowchart Fun Pro pour seulement 3 $/mois ou 30 $/an.",
-    "Impress your audience with high-quality images produced by Flowchart Fun":
-      "Impressionnez votre public avec des images de haute qualité produites par Flowchart Fun ",
-    "Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":
-      "Incluez un titre en utilisant un attribut <0>title</0>. Pour utiliser la coloration Visio, ajoutez un attribut <1>roleType</1> égal à l'un des éléments suivants:",
-    "Indenting creates an edge": "L'indentation crée un bord",
-    Info: "Info",
-    Instruct: "Instruire",
-    Is: "Est",
-    Language: "Langue",
-    Large: "Grand",
-    Layout: "Disposition",
-    "Layout Frozen": "Mise en page gelée",
-    "Layout is Frozen": "La mise en page est gelée",
-    "Leading References": "Principales références",
-    "Learn More": "En savoir plus",
-    "Learn Syntax": "Apprendre la syntaxe",
-    "Learn about Flowchart Fun Pro": "En savoir plus sur Flowchart Fun Pro",
-    "Left to Right": "De gauche à droite",
-    "Let's get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":
-      "Commençons ! Entrez votre adresse e-mail ci-dessous pour créer votre compte Flowchart Fun et commencez à utiliser nos fonctionnalités puissantes.",
-    Light: "Lumineux",
-    "Light Mode": "Mode lumineux",
-    Link: "Lien",
-    "Load Theme": "Charger le thème",
-    "Loading a theme will erase any custom styles you have applied":
-      "Charger un thème effacera tous les styles personnalisés que vous avez appliqués",
-    "Lock Zoom to Graph": "Verrouiller le Zoom sur le Graphique",
-    "Log In": "Connexion",
-    "Log Out": "Déconnexion",
-    "Log in to start creating flowcharts.":
-      "Connectez-vous pour commencer à créer des diagrammes.",
-    "Make a One-Time Donation": "Faites un don unique",
-    "Make publicly accessible": "Rendre accessible au public",
-    "Make sure you use the same email you will use to log in.":
-      "Assurez-vous d'utiliser le même e-mail que vous utiliserez pour vous connecter.",
-    "Map Data": "Cartographier les données",
-    "Market understanding and competitive landscape maintenance for SaaS product development":
-      "Comprendre le marché et maintenir le paysage concurrentiel pour le développement de produits SaaS",
-    Medium: "Moyen",
-    Method: "Méthode",
-    Monospace: "Monospace",
-    Monthly: "Mensuel",
-    "Multiple pointers on same line": "Plusieurs pointeurs sur la même ligne",
-    Museum: "Musée",
-    Name: "Nom",
-    New: "Nouveau",
-    "New Email": "Nouveau courriel",
-    "Next charge": "Prochain paiement",
-    "No Edges": "Pas de bords",
-    Node: "Nœud",
-    "Node Border Style": "Style de bordure de nœud",
-    "Node Colors": "Couleurs de nœud",
-    "Node ID": "Identifiant de nœud",
-    "Node ID, Classes, Attributes": "Identifiant de nœud, classes, attributs",
-    "Node Label": "Étiquette de nœud",
-    "Node Shapes": "Formes de nœud",
-    "Node and pointer on same line": "Nœud et pointeur sur la même ligne",
-    "Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":
-      "Les nœuds peuvent être stylisés avec des traits, des points ou des doubles. Les bordures peuvent également être supprimées avec border-none.",
-    "Not Empty": "Pas vide",
-    "Office Hours": "Heures de travail",
-    "Once in a while the magic link will end up in your spam folder. If you don't see it after a few minutes, check there or request a new link.":
-      " temps en temps, le lien magique finira par atterrir dans votre dossier de pourriel. Si vous ne le voyez pas après quelques minutes, vérifiez-y ou demandez un nouveau lien.",
-    "One-on-One Support": "Assistance individuelle",
-    "One-time Share Links": "Liens de partage à usage unique",
-    "Only available on this device.<0/>Clearing your browser cache will erase them.":
-      "Disponible uniquement sur cet appareil.<0/>La suppression de votre cache de navigateur les effacera.",
-    "Open Customer Portal": "Ouvrir le portail client",
-    "Organization Chart": "Organigramme",
-    "Paste the information you wish to convert to a flowchart":
-      "Collez les informations que vous souhaitez convertir en organigramme",
-    Permalinks: "Permalinks",
-    "Permanent Charts are a Pro Feature":
-      "Les diagrammes permanents sont une fonctionnalité Pro",
-    "Permanent Flowcharts": "Diagrammes permanents",
-    Persistent: "Persistant",
-    "Persistent Flowcharts": "Diagrammes persistants",
-    Playbook: "Livre-jeu",
-    "Pointer and container on same line":
-      "Pointeur et conteneur sur la même ligne",
-    Pricing: "Tarification",
-    "Process for corporate social responsibility initiatives development and implementation across company operations":
-      "Processus de développement et d'implémentation d'initiatives de responsabilité sociale des entreprises dans l'ensemble des opérations de l'entreprise",
-    "Processing Data": "Traitement des données",
-    Prompt: "Invite",
-    Public: "Public",
-    Random: "Aléatoire",
-    "Read-only": "Lecture seulement",
-    "Reference by Class": "Référence par classe",
-    "Reference by ID": "Référence par ID",
-    "Reference by Label": "Référence par étiquette",
-    References: "Références",
-    "References are used to create edges between nodes that are created elsewhere in the document":
-      "Les références sont utilisées pour créer des arêtes entre les nœuds créés ailleurs dans le document",
-    "Referencing a node by its exact label":
-      "Référencer un nœud par sa étiquette exacte",
-    "Referencing a node by its unique ID":
-      "Référencer un nœud par son ID unique",
-    "Referencing multiple nodes with the same assigned class":
-      "Référencement de multiples nœuds avec la même classe assignée",
-    "Refresh Page": "Rafraîchir la page",
-    "Reload to Update": "Recharger pour mettre à jour",
-    "Remove Image Watermark": "Supprimer le filigrane de l'image",
-    Rename: "Renommer",
-    Reset: "Réinitialiser",
-    "Resume Subscription": "Reprendre l'abonnement",
-    Retro: "Rétro",
-    Return: "Retour",
-    "Right to Left": "De droite à gauche",
-    Roadmap: "Roadmap",
-    Save: "Sauver",
-    "Save 20% (2 months free!)": "Économisez 20% (2 mois gratuits!)",
-    "Save time and automate certain tasks with this cutting-edge feature":
-      "Économisez du temps et automatisez certaines tâches avec cette fonctionnalité de pointe",
-    "Save time and streamline your workflow by importing data from other sources":
-      "Économisez du temps et simplifiez votre flux de travail en important des données à partir d'autres sources",
-    "Select your plan!": "Sélectionnez votre plan !",
-    "Send Feedback": "Envoyer des commentaires",
-    Settings: "Paramètres",
-    Shape: "Forme",
-    "Sign Up": "Inscription",
-    "Sign Up Now": "Inscrivez-vous maintenant",
-    Size: "Taille",
-    Small: "Petit",
-    "Some classes are available to help style your flowchart":
-      "Certaines classes sont disponibles pour aider à styliser votre diagramme de flux",
-    "Sorry! This page is only available in English.":
-      "Désolé ! Cette page n'est disponible qu'en anglais.",
-    "Source Arrow": "Flèche de Source",
-    "Source Column": "Colonne source",
-    "Source Delimiter": "Délimiteur source",
-    "Source/Target Arrow Shape": "Forme de flèche source / cible",
-    Spacing: "Espacement",
-    "Special Attributes": "Attributs spéciaux",
-    Start: "Début",
-    "Start Over": "Recommencer",
-    Status: "État",
-    "Store any data associated to a node":
-      "Stocker toutes les données associées à un nœud",
-    "Stored in the cloud": "Stocké dans le cloud",
-    "Stored on this computer": "Stocké sur cet ordinateur",
-    "Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":
-      "Simplifiez votre flux de travail et simplifiez votre visualisation de processus avec Flowchart Fun. Abonnez-vous maintenant et profitez de nos puissantes fonctionnalités Pro !",
-    Style: "Style",
-    "Style Classes": "Classes de style",
-    Submit: "Soumettre",
-    Subscribe: "S'abonner",
-    Subscription: "Abonnement",
-    "Subscription will end": "L'abonnement prendra fin",
-    "Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":
-      "Analyse et optimisation de la chaîne d'approvisionnement : réduction des coûts, amélioration de l'efficacité et collaboration des parties prenantes",
-    "Target Arrow": "Flèche de Cible",
-    "Target Column": "Colonne cible",
-    "Target Delimiter": "Délimiteur cible",
-    Temporary: "Temporaire",
-    "Temporary Flowchart Warning":
-      "Avertissement de diagramme de flux temporaire",
-    "Temporary Flowcharts": "Diagrammes temporaires",
-    "Text Sizes": "Tailles de texte",
-    "Text followed by colon+space creates an edge with the text as the label":
-      "Texte suivi d'un deux-points + espace crée un bord avec le texte comme étiquette",
-    "Text on a line creates a node with the text as the label":
-      "Texte sur une ligne crée un nœud avec le texte comme étiquette",
-    "Text sizes include small, regular, large, and extra-large.":
-      "Les tailles de texte comprennent petite, régulière, grande et extra-grande.",
-    "Thank you for your feedback!": "Merci pour votre commentaire !",
-    "The column that contains the edge label(s)":
-      "La colonne qui contient les étiquettes de bord",
-    "The column that contains the source node ID(s)":
-      "La colonne qui contient les ID de nœud source",
-    "The column that contains the target node ID(s)":
-      "La colonne qui contient les ID de nœud cible",
-    "The delimiter used to separate multiple source nodes":
-      "Le délimiteur utilisé pour séparer plusieurs nœuds source",
-    "The delimiter used to separate multiple target nodes":
-      "Le délimiteur utilisé pour séparer plusieurs nœuds cibles",
-    "The possible shapes are:": "Les formes possibles sont :",
-    "Theme Editor": "Éditeur de thème",
-    "There are no edges in this data": "Il n'y a pas d'arêtes dans ces données",
-    "This action cannot be undone.": "Cette action ne peut pas être annulée.",
-    "This app works by typing": "Chaque ligne creé une nouvelle boite",
-    "This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":
-      "Ce document n'est sauvegardé que sur cet ordinateur. Si vous voulez un moyen de stocker des documents et de les consulter où que vous soyez, essayez notre plan Pro pour seulement 3 $/mois !",
-    "This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":
-      "Cette fonctionnalité est uniquement disponible pour les utilisateurs pro. <0>Devenez un utilisateur pro</0> pour la débloquer.",
-    "This is a temporary chart, it will be deleted when you clear your browser cache.":
-      "Il s'agit d'un graphique temporaire, il sera supprimé lorsque vous effacerez le cache de votre navigateur.",
-    Tip: "Astuce",
-    "To fix this change one of the edge IDs":
-      "Pour corriger cela, changez l'un des ID de bord",
-    "To fix this change one of the node IDs":
-      "Pour corriger ceci, changez l'un des ID de nœud",
-    "To fix this move one pointer to the next line":
-      "Pour corriger ceci, déplacez un pointeur vers la ligne suivante",
-    "To fix this move the pointer to the next line":
-      "Pour corriger ceci, déplacez le pointeur vers la ligne suivante",
-    "To fix this start the container <0/> on a different line":
-      "Pour corriger cela, commencez le conteneur <0/> sur une ligne différente.",
-    "Top to Bottom": "De haut en bas",
-    "Try again": "Réessayer",
-    "Two edges have the same ID": "Deux arêtes ont le même ID",
-    "Two nodes have the same ID": "Deux nœuds ont le même ID",
-    Type: "Type",
-    Unfreeze: "Débloquer",
-    "Unique text value to identify a node":
-      "Valeur de texte unique pour identifier un nœud",
-    "Unknown Parsing Error": "Erreur d'analyse inconnue",
-    "Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":
-      "Libérez votre créativité et simplifiez votre flux de travail avec Flowchart Fun Pro - à partir de seulement 3 $ / mois !",
-    "Update Email": "Mettre à jour l'e-mail",
-    "Upgrade to Pro": "Passer à Pro",
-    "Upload your File": "Téléchargez votre fichier",
-    "Use AI to create flowcharts from data sets or algorithms":
-      "Utilisez l'IA pour créer des diagrammes à partir de jeux de données ou d'algorithmes",
-    "Use AI to generate a flowchart from a prompt. <0>Learn More</0>":
-      "Utilisez l'IA pour générer un diagramme de flux à partir d'une invite. <0>En savoir plus</0>",
-    "Use classes to group nodes":
-      "Utilisez des classes pour regrouper les nœuds",
-    "Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":
-      "Utilisez l'attribut <0>href</0> pour créer un lien sur un nœud qui s'ouvre dans un nouvel onglet.",
-    "Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":
-      "Utilisez l'attribut <0>src</0> pour définir l'image d'un nœud. L'image sera mise à l'échelle pour s'adapter au nœud, vous devrez donc peut-être ajuster la largeur et la hauteur du nœud pour obtenir le résultat souhaité. Seules les images publiques (non bloquées par CORS) sont prises en charge.",
-    "Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":
-      "Utilisez les attributs <0>w</0> et <1>h</1> pour définir explicitement la largeur et la hauteur d'un nœud.",
-    "Use the customer portal to change your billing information.":
-      "Utilisez le portail client pour modifier vos informations de facturation.",
-    "Use this file for org charts, hierarchies, and other organizational structures.":
-      "Utilisez ce fichier pour les organigrammes, les hiérarchies et autres structures organisationnelles.",
-    "Use this file for sequences, processes, and workflows.":
-      "Utilisez ce fichier pour les séquences, les processus et les workflows.",
-    User: "Utilisateur",
-    "View on Github": "Voir sur Github",
-    "Water evaporates from the Earth's surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":
-      "L'eau s'évapore de la surface de la Terre, monte dans l'atmosphère et retombe sous forme de précipitations. Cette eau s'écoule alors dans les rivières, les lacs et les océans, où elle s'évapore à nouveau et est recyclée dans l'atmosphère.",
-    "We use magic links to log you in. Enter your email below to get started.":
-      "Nous utilisons des liens magiques pour vous connecter. Entrez votre email ci-dessous pour commencer.",
-    "Welcome to Flowchart Fun Pro!": "Bienvenue à Flowchart Fun Pro !",
-    "Width and Height": "Largeur et hauteur",
-    "Would you like to continue?": "Voulez-vous continuer ?",
-    Yearly: "Annuellement",
-    "You already have a flowchart with this name.":
-      "Vous avez déjà un diagramme avec ce nom.",
-    "You are about to add {numNodes} nodes and {numEdges} edges to your graph.":
-      [
-        "Vous êtes sur le point d'ajouter ",
-        ["numNodes"],
-        " nœuds et ",
-        ["numEdges"],
-        " arêtes à votre graphe.",
-      ],
-    'You can also wrap text in "\\( \\)" to create an edge':
-      'Vous pouvez également envelopper le texte dans "\\( \\)" pour créer un bord',
-    "You must log in to create a persistent flowchart.":
-      "Vous devez vous connecter pour créer un diagramme de flux persistant.",
-    "You're almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":
-      "Vous y êtes presque ! Une dernière étape pour débloquer tout le potentiel de <0>Flowchart Fun Pro</0>. Entrez vos informations de paiement ci-dessous pour compléter votre abonnement et commencez à créer des diagrammes de flux incroyables dès aujourd'hui.",
-    "Your Charts": "Vos diagrammes",
-    "Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":
-      "Votre abonnement n'est plus actif. Si vous souhaitez créer et modifier des graphiques permanents, passez à Pro.",
-    "Zoom In": "Zoomer",
-    "Zoom Out": "Zoomer vers l'extérieur",
-    "any text: before a colon creates a label":
-      "n'importe quel texte: avant que deux-points ne créent une étiquette",
-  },
+  messages: JSON.parse(
+    '{"$3 per month":"3 $ par mois","$30 per year":"30 $ par an","(before a colon creates a label)":"(avant que deux-points ne créent une étiquette)","<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":"<0>Flowchart Fun</0> est un projet open source réalisé par <1>Tone Row</1>","<0>Sign In</0> / <1>Sign Up</1> with email and password":"<0>Se connecter</0> / <1>S\'inscrire</1> avec email et mot de passe","<0>Starting August 28th you will need a subscription to create and edit charts. <1>Learn more</1>.</0>":"<0>À partir du 28 août, vous aurez besoin d\'un abonnement pour créer et modifier des graphiques. <1>En savoir plus</1>.</0>","<0>You currently have a free account.<1/><2>Learn about our Pro Features and subscribe on our pricing page</2>.</0>":"<0>Vous avez actuellement un compte gratuit.<1/><2>Apprenez-en plus sur nos fonctionnalités Pro et abonnez-vous sur notre page tarifaire</2>.</0>","A new version of the app is available. Please reload to update.":"Une nouvelle version de l\'application est disponible. Veuillez recharger pour mettre à jour.","About":"À propos","Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":"Accédez à ces diagrammes depuis n\'importe où.<0/>Partagez et intégrez des diagrammes de flux qui restent synchronisés.","Accessible from any device":"Accessible depuis n\'importe quel appareil","Account":"Compte","Activate your account":"Activez votre compte","Advanced":"Avancé","Amount":"Montant","An error occurred. Try resubmitting or email {0} directly.":["Une erreur s\'est produite. Essayez de l\'envoyer à nouveau ou bien envoyez un e-mail à l\'adresse ",["0"],"."],"Appearance":"Thème","Apply Style":"Appliquer un style","Attributes":"Attributs","Auto-Layouts":"Mises en page automatiques","Back To Editor":"Retour à l\'éditeur","Basic Flowchart":"Diagramme de flux de base","Become a Github Sponsor":"Devenez un sponsor Github","Become a Pro User":"Devenez un utilisateur Pro","Begin with a simple example showing how <0>Flowchart Fun</0> works.":"Commencez par un exemple simple montrant comment <0>Flowchart Fun</0> fonctionne.","Blank":"Vide","Blog":"Blog","Book a Meeting":"Réserver une réunion","Border":"Frontière","Bottom to Top":"De bas en haut","Breadthfirst":"Parcours en largeur","Cancel":"Annuler","Cancel your subscription. Your hosted charts will become read-only.":"Résilier votre abonnement. Vos graphiques hébergés seront en lecture seule.","Certain attributes can be used to customize the appearance or functionality of elements.":"Certains attributs peuvent être utilisés pour personnaliser l\'apparence ou la fonctionnalité des éléments.","Change Email Address":"Changer l\'adresse email","Changelog":"Journal des modifications","Charts":"Graphiques","Check your email for a link to log in.<0/>You can close this window.":"Vérifiez votre e-mail pour un lien de connexion. Vous pouvez fermer cette fenêtre.","Choose":"Choisir","Choose a Theme":"Choisir un thème","Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":"Choisissez parmi une variété de formes de flèches pour la source et la cible d\'un bord. Les formes incluent triangle, triangle-tee, cercle-triangle, triangle-croix, triangle-backcurve, vee, tee, carré, cercle, diamant, chevron, aucun.","Choose the plan that\'s right for you and start creating amazing flowcharts with Flowchart Fun Pro":"Choisissez le plan qui vous convient le mieux et commencez à créer des diagrammes incroyables avec Flowchart Fun Pro","Circle":"Cercle","Classes":"Classes","Clay":"Argile","Clear":"Effacer","Clear text?":"Effacer le texte?","Click on \\\\nLearn Syntax\\\\nto learn more":"Cliquez sur \\\\nApprendre la syntaxe\\\\npour en savoir plus.","Click to preview":"Cliquez pour prévisualiser","Clone":"Cloner","Close":"Fermer","Color":"Couleur","Colors include red, orange, yellow, blue, purple, black, white, and gray.":"Les couleurs incluent le rouge, l\'orange, le jaune, le bleu, le pourpre, le noir, le blanc et le gris.","Column":"Colonne","Comic Book":"Bande dessinée","Comment":"Commenter","Concentric":"Concentrique","Confirm New Email":"Confirmer le nouveau Email","Confirm your email address to sign in.":"Confirmez votre adresse e-mail pour vous connecter.","Containers":"Conteneurs","Containers are nodes that contain other nodes. They are declared using curly braces.":"Les conteneurs sont des nœuds qui contiennent d\'autres nœuds. Ils sont déclarés à l\'aide de accolades.","Continue":"Continuer","Convert this to a hosted chart?":"Convertir ceci en graphique hébergé?","Convert to hosted chart?":"Convertir en graphique hébergé ?","Copy":"Copier","Copy PNG Image":"Copier l\'image PNG","Copy SVG Code":"Copier le code SVG","Copy your mermaid.js code or open it directly in the mermaid.js live editor.":"Copiez votre code mermaid.js ou ouvrez-le directement dans l\'éditeur en direct mermaid.js.","Create Flowcharts from a Prompt with AI":"Créer des diagrammes à partir d\'une invite avec l\'IA","Create New Flowchart":"Créer un nouveau diagramme","Create a New Flowchart":"Créer un nouveau diagramme","Create professional-quality visual aids for presentations or training materials":"Créez des aides visuelles de qualité professionnelle pour des présentations ou des matériels de formation","Creating an edge between two nodes is done by indenting the second node below the first":"Créer une arête entre deux nœuds est fait en indentant le second nœud sous le premier","Custom Sharing Options":"Options de partage personnalisées","Customer Portal":"Portail Clients","Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":"Personnalisez votre thème en modifiant le <0>Cytoscape CSS</0> ci-dessous. Notre documentation de mise en forme arrive bientôt ! Entre-temps, la meilleure ressource est la documentation <1>Cytoscape</1>. Venez poser des questions sur le <2>Discord</2> si vous êtes coincé.","Dark":"Sombre","Dark Mode":"Mode sombre","Date":"Date","Deleted when browser data is cleared":"Supprimé lorsque les données du navigateur sont effacées","Describe the flowchart you wish to create":"Décrivez le schéma que vous souhaitez créer","Direction":"Direction","Discard":"Jeter","Do you want to cancel your subscription?":"Souhaitez-vous résilier votre abonnement ?","Do you want to delete this?":"Souhaitez-vous supprimer ceci ?","Document":"Document","Download":"Télécharger","Download JPG":"Télécharger JPG","Download PNG":"Télécharger PNG","Download SVG":"Télécharger SVG","Drag and drop a CSV file here, or click to select a file":"Glissez-déposez un fichier CSV ici, ou cliquez pour sélectionner un fichier","Draw an edge from multiple nodes by beginning the line with a reference":"Dessinez une arête à partir de plusieurs nœuds en commençant la ligne par une référence","Drop the file here ...":"Déposez le fichier ici ...","Easily import existing flowcharts from other software":"Importez facilement des diagrammes existants à partir d\'autres logiciels","Edge":"Bord","Edge ID, Classes, Attributes":"ID Edge, Classes, Attributs","Edge Label":"Étiquette Edge","Edge Label Column":"Colonne d\'étiquette Edge","Edge Style":"Style Edge","Edge missing indentation":"Indentation manquante du bord","Edges":"Bords","Edges are declared in the same row as their source node":"Les bords sont déclarés dans la même ligne que leur nœud source","Edges are declared in the same row as their target node":"Les bords sont déclarés dans la même ligne que leur nœud cible","Edges are declared in their own row":"Les bords sont déclarés dans leur propre ligne","Edges can also have ID\'s, classes, and attributes before the label":"Les bords peuvent également avoir des ID, des classes et des attributs avant l\'étiquette","Edges can be styled with dashed, dotted, or solid lines":"Les bords peuvent être stylisés avec des lignes en pointillés, en pointillés ou en lignes continues","Edges in Separate Rows":"Bordures en Rangs Séparés","Edges in Source Node Row":"Bordures dans la Ligne du Nœud Source","Edges in Target Node Row":"Bordures dans la Ligne du Nœud Cible","Editable":"Modifiable","Editor":"Éditeur","Eggs":"Oeufs","Email":"E-mail","Empty":"Vide","Enter your email":"Entrez votre adresse e-mail ","Enter your email address and we\'ll send you a magic link to sign in.":"Entrez votre adresse e-mail et nous vous enverrons un lien magique pour vous connecter.","Enter your email address below and we\'ll send you a link to reset your password.":"Entrez votre adresse e-mail ci-dessous et nous vous enverrons un lien pour réinitialiser votre mot de passe.","Equal To":"Égal à","Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":"Diagramme de flux de l\'écriture d\'essai, guidant les étudiants à travers la réflexion, le schématisme, le brouillon et les étapes de révision","Everything in Free":"Tout dans le plan gratuit","Excalidraw":"Excalidraw","Export":"Exporter","Export High-Resolution Images":"Exporter des images haute résolution","Export to SVG":"Exporter en SVG","Extra Large":"Extra Large","Extract":"Extrait","Feedback":"Commentaire","Follow Us on Twitter":"Suivez-nous sur Twitter","Forgot your password?":"Mot de passe oublié ?","Found a bug? Have a feature request? We would love to hear from you!":"Vous avez trouvé un bug ? Vous souhaitez faire une demande de fonctionnalité ? Nous aimerions beaucoup vous entendre !","Free":"Gratuit ","Fullscreen":"Plein écran","Futuristic":"Futuriste","Get More Layouts":"Obtenez plus de mises en page","Get personalized attention and support from our expert team":"Obtenez une attention et un soutien personnalisés de notre équipe d\'experts ","Getting Started":"Commencer","Graph":"Graphique","Grid":"Quadrillage","Have complex questions or issues? We\'re here to help.":"Des questions ou des problèmes complexes ? Nous sommes là pour vous aider. ","Help":"Aide","History":"Historique","Home":"Accueil","How are edges declared in this data?":"Comment les bords sont-ils déclarés dans ces données?","ID\'s":"ID","If an account with that email exists, we\'ve sent you an email with instructions on how to reset your password.":"Si un compte avec cet e-mail existe, nous vous avons envoyé un e-mail avec des instructions sur la façon de réinitialiser votre mot de passe.","If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":"Si vous appréciez l\'utilisation de <0>Flowchart Fun</0>, veuillez envisager de soutenir le projet","If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\\\:</0>":"Si vous souhaitez créer un bord, faites une indentation de cette ligne. Sinon, échappez le deux-points avec une barre oblique <0>\\\\:</0>","Image Export":"Exportation d\'images","Images":"Images","Import Data":"Importer des données","Import data from a CSV file.":"Importer des données à partir d\'un fichier CSV.","Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":"Importer des données à partir de n\'importe quel fichier CSV et les mapper à un nouveau diagramme. C\'est une excellente façon d\'importer des données à partir d\'autres sources telles que Lucidchart, Google Sheets et Visio.","Import from Lucidchart, Visio, and other CSV files":"Importer à partir de Lucidchart, Visio et autres fichiers CSV","Import your diagram it into Microsoft Visio using one of these CSV files.":"Importez votre diagramme dans Microsoft Visio à l\'aide de l\'un de ces fichiers CSV.","Important Changes are Coming to Flowchart Fun.":"Des changements importants arrivent à Flowchart Fun.","Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":"L\'importation de données est une fonctionnalité professionnelle. Vous pouvez passer à Flowchart Fun Pro pour seulement 3 $/mois ou 30 $/an.","Impress your audience with high-quality images produced by Flowchart Fun":"Impressionnez votre public avec des images de haute qualité produites par Flowchart Fun ","Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":"Incluez un titre en utilisant un attribut <0>title</0>. Pour utiliser la coloration Visio, ajoutez un attribut <1>roleType</1> égal à l\'un des éléments suivants:","Indenting creates an edge":"L\'indentation crée un bord","Info":"Info","Instruct":"Instruire","Is":"Est","Language":"Langue","Large":"Grand","Layout":"Disposition","Layout Frozen":"Mise en page gelée","Layout is Frozen":"La mise en page est gelée","Leading References":"Principales références","Learn More":"En savoir plus","Learn Syntax":"Apprendre la syntaxe","Learn about Flowchart Fun Pro":"En savoir plus sur Flowchart Fun Pro","Left to Right":"De gauche à droite","Let\'s get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":"Commençons ! Entrez votre adresse e-mail ci-dessous pour créer votre compte Flowchart Fun et commencez à utiliser nos fonctionnalités puissantes.","Light":"Lumineux","Light Mode":"Mode lumineux","Link":"Lien","Load Theme":"Charger le thème","Loading a theme will erase any custom styles you have applied":"Charger un thème effacera tous les styles personnalisés que vous avez appliqués","Lock Zoom to Graph":"Verrouiller le Zoom sur le Graphique","Log In":"Connexion","Log Out":"Déconnexion","Log in to start creating flowcharts.":"Connectez-vous pour commencer à créer des diagrammes.","Make a One-Time Donation":"Faites un don unique","Make publicly accessible":"Rendre accessible au public","Make sure you use the same email you will use to log in.":"Assurez-vous d\'utiliser le même e-mail que vous utiliserez pour vous connecter.","Map Data":"Cartographier les données","Market understanding and competitive landscape maintenance for SaaS product development":"Comprendre le marché et maintenir le paysage concurrentiel pour le développement de produits SaaS","Medium":"Moyen","Method":"Méthode","Monospace":"Monospace","Monthly":"Mensuel","Multiple pointers on same line":"Plusieurs pointeurs sur la même ligne","Museum":"Musée","Name":"Nom","New":"Nouveau","New Email":"Nouveau courriel","Next charge":"Prochain paiement","No Edges":"Pas de bords","Node":"Nœud","Node Border Style":"Style de bordure de nœud","Node Colors":"Couleurs de nœud","Node ID":"Identifiant de nœud","Node ID, Classes, Attributes":"Identifiant de nœud, classes, attributs","Node Label":"Étiquette de nœud","Node Shapes":"Formes de nœud","Node and pointer on same line":"Nœud et pointeur sur la même ligne","Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":"Les nœuds peuvent être stylisés avec des traits, des points ou des doubles. Les bordures peuvent également être supprimées avec border-none.","Not Empty":"Pas vide","Office Hours":"Heures de travail","Once in a while the magic link will end up in your spam folder. If you don\'t see it after a few minutes, check there or request a new link.":" temps en temps, le lien magique finira par atterrir dans votre dossier de pourriel. Si vous ne le voyez pas après quelques minutes, vérifiez-y ou demandez un nouveau lien.","One-on-One Support":"Assistance individuelle","One-time Share Links":"Liens de partage à usage unique","Only available on this device.<0/>Clearing your browser cache will erase them.":"Disponible uniquement sur cet appareil.<0/>La suppression de votre cache de navigateur les effacera.","Open Customer Portal":"Ouvrir le portail client","Organization Chart":"Organigramme","Password":"Mot de passe ","Paste the information you wish to convert to a flowchart":"Collez les informations que vous souhaitez convertir en organigramme","Permalinks":"Permalinks","Permanent Charts are a Pro Feature":"Les diagrammes permanents sont une fonctionnalité Pro","Permanent Flowcharts":"Diagrammes permanents","Persistent":"Persistant","Persistent Flowcharts":"Diagrammes persistants","Playbook":"Livre-jeu","Pointer and container on same line":"Pointeur et conteneur sur la même ligne","Pricing":"Tarification","Process for corporate social responsibility initiatives development and implementation across company operations":"Processus de développement et d\'implémentation d\'initiatives de responsabilité sociale des entreprises dans l\'ensemble des opérations de l\'entreprise","Processing Data":"Traitement des données","Prompt":"Invite","Public":"Public","Random":"Aléatoire","Read-only":"Lecture seulement","Reference by Class":"Référence par classe","Reference by ID":"Référence par ID","Reference by Label":"Référence par étiquette","References":"Références","References are used to create edges between nodes that are created elsewhere in the document":"Les références sont utilisées pour créer des arêtes entre les nœuds créés ailleurs dans le document","Referencing a node by its exact label":"Référencer un nœud par sa étiquette exacte","Referencing a node by its unique ID":"Référencer un nœud par son ID unique","Referencing multiple nodes with the same assigned class":"Référencement de multiples nœuds avec la même classe assignée","Refresh Page":"Rafraîchir la page","Reload to Update":"Recharger pour mettre à jour","Remove Image Watermark":"Supprimer le filigrane de l\'image","Rename":"Renommer","Request Magic Link":"Demandez un lien magique","Request Password Reset":"Demandez une réinitialisation du mot de passe ","Reset":"Réinitialiser","Reset Password":"Réinitialiser le mot de passe","Resume Subscription":"Reprendre l\'abonnement","Retro":"Rétro","Return":"Retour","Right to Left":"De droite à gauche","Roadmap":"Roadmap","Save":"Sauver","Save 20% (2 months free!)":"Économisez 20% (2 mois gratuits!)","Save time and automate certain tasks with this cutting-edge feature":"Économisez du temps et automatisez certaines tâches avec cette fonctionnalité de pointe","Save time and streamline your workflow by importing data from other sources":"Économisez du temps et simplifiez votre flux de travail en important des données à partir d\'autres sources","Select your plan!":"Sélectionnez votre plan !","Send Feedback":"Envoyer des commentaires","Settings":"Paramètres","Shape":"Forme","Sign In":"Se connecter ","Sign Up":"Inscription","Sign in with <0>GitHub</0>":"Se connecter avec <0>GitHub</0>","Sign in with <0>Google</0>":"Se connecter avec <0>Google</0>","Size":"Taille","Small":"Petit","Some classes are available to help style your flowchart":"Certaines classes sont disponibles pour aider à styliser votre diagramme de flux","Sorry! This page is only available in English.":"Désolé ! Cette page n\'est disponible qu\'en anglais.","Source Arrow":"Flèche de Source","Source Column":"Colonne source","Source Delimiter":"Délimiteur source","Source/Target Arrow Shape":"Forme de flèche source / cible","Spacing":"Espacement","Special Attributes":"Attributs spéciaux","Start":"Début","Start Over":"Recommencer","Status":"État","Store any data associated to a node":"Stocker toutes les données associées à un nœud","Stored in the cloud":"Stocké dans le cloud","Stored on this computer":"Stocké sur cet ordinateur","Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":"Simplifiez votre flux de travail et simplifiez votre visualisation de processus avec Flowchart Fun. Abonnez-vous maintenant et profitez de nos puissantes fonctionnalités Pro !","Style":"Style","Style Classes":"Classes de style","Submit":"Soumettre","Subscribe":"S\'abonner","Subscription":"Abonnement","Subscription will end":"L\'abonnement prendra fin","Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":"Analyse et optimisation de la chaîne d\'approvisionnement : réduction des coûts, amélioration de l\'efficacité et collaboration des parties prenantes","Target Arrow":"Flèche de Cible","Target Column":"Colonne cible","Target Delimiter":"Délimiteur cible","Temporary":"Temporaire","Temporary Flowchart Warning":"Avertissement de diagramme de flux temporaire","Temporary Flowcharts":"Diagrammes temporaires","Text Sizes":"Tailles de texte","Text followed by colon+space creates an edge with the text as the label":"Texte suivi d\'un deux-points + espace crée un bord avec le texte comme étiquette","Text on a line creates a node with the text as the label":"Texte sur une ligne crée un nœud avec le texte comme étiquette","Text sizes include small, regular, large, and extra-large.":"Les tailles de texte comprennent petite, régulière, grande et extra-grande.","Thank you for your feedback!":"Merci pour votre commentaire !","The column that contains the edge label(s)":"La colonne qui contient les étiquettes de bord","The column that contains the source node ID(s)":"La colonne qui contient les ID de nœud source","The column that contains the target node ID(s)":"La colonne qui contient les ID de nœud cible","The delimiter used to separate multiple source nodes":"Le délimiteur utilisé pour séparer plusieurs nœuds source","The delimiter used to separate multiple target nodes":"Le délimiteur utilisé pour séparer plusieurs nœuds cibles","The possible shapes are:":"Les formes possibles sont :","Theme Editor":"Éditeur de thème","There are no edges in this data":"Il n\'y a pas d\'arêtes dans ces données","This action cannot be undone.":"Cette action ne peut pas être annulée.","This app works by typing":"Chaque ligne creé une nouvelle boite","This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":"Ce document n\'est sauvegardé que sur cet ordinateur. Si vous voulez un moyen de stocker des documents et de les consulter où que vous soyez, essayez notre plan Pro pour seulement 3 $/mois !","This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":"Cette fonctionnalité est uniquement disponible pour les utilisateurs pro. <0>Devenez un utilisateur pro</0> pour la débloquer.","This is a temporary chart, it will be deleted when you clear your browser cache.":"Il s\'agit d\'un graphique temporaire, il sera supprimé lorsque vous effacerez le cache de votre navigateur.","Tip":"Astuce","To fix this change one of the edge IDs":"Pour corriger cela, changez l\'un des ID de bord","To fix this change one of the node IDs":"Pour corriger ceci, changez l\'un des ID de nœud","To fix this move one pointer to the next line":"Pour corriger ceci, déplacez un pointeur vers la ligne suivante","To fix this move the pointer to the next line":"Pour corriger ceci, déplacez le pointeur vers la ligne suivante","To fix this start the container <0/> on a different line":"Pour corriger cela, commencez le conteneur <0/> sur une ligne différente.","To learn more about why we require you to log in, please read <0>this blog post</0>.":"Pour en savoir plus sur la raison pour laquelle nous vous demandons de vous connecter, veuillez lire <0>ce message de blog</0>.","Top to Bottom":"De haut en bas","Try again":"Réessayer","Two edges have the same ID":"Deux arêtes ont le même ID","Two nodes have the same ID":"Deux nœuds ont le même ID","Type":"Type","Unfreeze":"Débloquer","Unique text value to identify a node":"Valeur de texte unique pour identifier un nœud","Unknown Parsing Error":"Erreur d\'analyse inconnue","Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":"Libérez votre créativité et simplifiez votre flux de travail avec Flowchart Fun Pro - à partir de seulement 3 $ / mois !","Update Email":"Mettre à jour l\'e-mail","Upgrade to Pro":"Passer à Pro","Upload your File":"Téléchargez votre fichier","Use AI to create flowcharts from data sets or algorithms":"Utilisez l\'IA pour créer des diagrammes à partir de jeux de données ou d\'algorithmes","Use AI to generate a flowchart from a prompt. <0>Learn More</0>":"Utilisez l\'IA pour générer un diagramme de flux à partir d\'une invite. <0>En savoir plus</0>","Use classes to group nodes":"Utilisez des classes pour regrouper les nœuds","Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":"Utilisez l\'attribut <0>href</0> pour créer un lien sur un nœud qui s\'ouvre dans un nouvel onglet.","Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":"Utilisez l\'attribut <0>src</0> pour définir l\'image d\'un nœud. L\'image sera mise à l\'échelle pour s\'adapter au nœud, vous devrez donc peut-être ajuster la largeur et la hauteur du nœud pour obtenir le résultat souhaité. Seules les images publiques (non bloquées par CORS) sont prises en charge.","Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":"Utilisez les attributs <0>w</0> et <1>h</1> pour définir explicitement la largeur et la hauteur d\'un nœud.","Use the customer portal to change your billing information.":"Utilisez le portail client pour modifier vos informations de facturation.","Use this file for org charts, hierarchies, and other organizational structures.":"Utilisez ce fichier pour les organigrammes, les hiérarchies et autres structures organisationnelles.","Use this file for sequences, processes, and workflows.":"Utilisez ce fichier pour les séquences, les processus et les workflows.","User":"Utilisateur","View on Github":"Voir sur Github","Water evaporates from the Earth\'s surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":"L\'eau s\'évapore de la surface de la Terre, monte dans l\'atmosphère et retombe sous forme de précipitations. Cette eau s\'écoule alors dans les rivières, les lacs et les océans, où elle s\'évapore à nouveau et est recyclée dans l\'atmosphère.","Welcome to Flowchart Fun Pro!":"Bienvenue à Flowchart Fun Pro !","Width and Height":"Largeur et hauteur","Would you like to continue?":"Voulez-vous continuer ?","Yearly":"Annuellement","You already have a flowchart with this name.":"Vous avez déjà un diagramme avec ce nom.","You are about to add {numNodes} nodes and {numEdges} edges to your graph.":["Vous êtes sur le point d\'ajouter ",["numNodes"]," nœuds et ",["numEdges"]," arêtes à votre graphe."],"You can also wrap text in \\"\\\\( \\\\)\\" to create an edge":"Vous pouvez également envelopper le texte dans \\"\\\\( \\\\)\\" pour créer un bord","You must log in to create a persistent flowchart.":"Vous devez vous connecter pour créer un diagramme de flux persistant.","You need to log in to access this page.":"Vous devez vous connecter pour accéder à cette page.","You\'re almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":"Vous y êtes presque ! Une dernière étape pour débloquer tout le potentiel de <0>Flowchart Fun Pro</0>. Entrez vos informations de paiement ci-dessous pour compléter votre abonnement et commencez à créer des diagrammes de flux incroyables dès aujourd\'hui.","Your Charts":"Vos diagrammes","Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":"Votre abonnement n\'est plus actif. Si vous souhaitez créer et modifier des graphiques permanents, passez à Pro.","Zoom In":"Zoomer","Zoom Out":"Zoomer vers l\'extérieur","any text: before a colon creates a label":"n\'importe quel texte: avant que deux-points ne créent une étiquette","or":"ou"}'
+  ),
 };
