@@ -1,457 +1,5 @@
 /*eslint-disable*/ module.exports = {
-  messages: {
-    "$3 per month": "मासिक $3",
-    "$30 per year": "वार्षिक $30",
-    "(before a colon creates a label)": "(कॉलन से पहले एक लेबल बनाता है)",
-    "<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":
-      "<0>Flowchart Fun</0> <1>Tone Row</1> द्वारा बनाई गई एक खुला स्रोत परियोजना है ",
-    "A new version of the app is available. Please reload to update.":
-      "एप्प का एक नया वर्जन उपलब्ध है। अपडेट करने के लिए रीलोड करें।",
-    About: "के बारे में",
-    "Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":
-      "इन चार्ट्स को कहीं से भी एक्सेस करें।<0/>सिंक्रोनाइज़्ड रहने वाले फ्लोचार्ट को साझा करें और एम्बेड करें।",
-    "Accessible from any device": "किसी भी उपकरण से पहुँचा जा सकता है",
-    Account: "खाता",
-    "Activate your account": "अपना खाता सक्रिय करें",
-    Advanced: "विकसित",
-    Amount: "रकम",
-    "An error occurred. Try resubmitting or email {0} directly.": [
-      "एक एरर हो गया. फिर से सबमिट करने की कोशिश करें या सीधे ",
-      ["0"],
-      " ईमेल करें.",
-    ],
-    Appearance: "दिखावट",
-    "Apply Style": "शैली लागू करें",
-    Attributes: "गुण",
-    "Auto-Layouts": "ऑटो लेआउट",
-    "Back To Editor": "संपादक पर वापस जाएं",
-    "Basic Flowchart": "बेसिक फ्लोचार्ट",
-    "Become a Github Sponsor": "गिटहब स्पॉन्सर बनें",
-    "Become a Pro User": "प्रो उपयोगकर्ता बनें",
-    "Begin with a simple example showing how <0>Flowchart Fun</0> works.":
-      "<0>फ्लोचार्ट फन</0> काम कैसे करता है इसके लिए एक आसान उदाहरण से शुरू करें।",
-    Blank: "खाली",
-    Blog: "ब्लॉग",
-    "Book a Meeting": "एक बैठक बुक करें",
-    Border: "सीमा",
-    "Bottom to Top": "नीचे से शीर्ष तक",
-    Breadthfirst: "चौड़ाई पहले",
-    Cancel: "रद्द करें",
-    "Cancel your subscription. Your hosted charts will become read-only.":
-      "अपनी सदस्यता रद्द करें. आपके होस्ट किये गए चार्ट सिर्फ़ पढ़े जा सकेंगे.",
-    "Certain attributes can be used to customize the appearance or functionality of elements.":
-      "कुछ गुण तत्वों की दिखता या कार्यक्षमता को अनुकूलित करने के लिए उपयोग किए जा सकते हैं।",
-    "Change Email Address": "ईमेल पता बदलें",
-    Changelog: "बदलाव का",
-    Charts: "चार्ट",
-    "Check your email for a link to log in.<0/>You can close this window.":
-      "लॉग इन करने के लिए अपने ईमेल की जाँच करें। आप इस विंडो को बंद कर सकते हैं।",
-    "Choose a Theme": "एक थीम चुनें",
-    "Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":
-      "एक किनारे के स्रोत और लक्ष्य के लिए विभिन्न तीर आकारों से चुनें। आकार शामिल हैं त्रिकोण, त्रिकोण-टी, सर्कल-त्रिकोण, त्रिकोण-क्रॉस, त्रिकोण-बैककर्व, वी, टी, चौकोर, हीरा, चेवरॉन और कोई नहीं।",
-    "Choose the plan that's right for you and start creating amazing flowcharts with Flowchart Fun Pro":
-      "आपके लिए सही योजना चुनें और Flowchart Fun Pro के साथ अद्भुत फ्लोचार्ट बनाना शुरू करें",
-    Circle: "परिधि",
-    Classes: "वर्ग",
-    Clay: "मिट्टी",
-    Clear: "साफ़",
-    "Clear text?": "पाठ साफ़ करें?",
-    "Click on \\nLearn Syntax\\nto learn more":
-      "\\nसिंटैक्स सीखने के लिए \\nपर क्लिक करें\\n",
-    "Click to preview": "पूर्वावलोकन के लिए क्लिक करें",
-    Clone: "क्लोन करें",
-    Close: "बंद करें",
-    Color: "रंग",
-    "Colors include red, orange, yellow, blue, purple, black, white, and gray.":
-      "रंग में लाल, नारंगी, पीला, नीला, बैंगनी, काला, सफेद, और ग्रे शामिल हैं।",
-    Column: "स्तंभ",
-    "Comic Book": "कॉमिक बुक",
-    Comment: "टिप्पणी",
-    Concentric: "गाढ़ा",
-    "Confirm New Email": "नई ईमेल की पुष्टि करें",
-    Containers: "कंटेनर",
-    "Containers are nodes that contain other nodes. They are declared using curly braces.":
-      "कंटेनर उन नोड्स हैं जो अन्य नोड्स को सम्मिलित करते हैं। वे कर्ली ब्रेस का उपयोग करके घोषित किया जाता है।",
-    Continue: "जारी रखें",
-    "Convert this to a hosted chart?":
-      "यह एक होस्ट की गई चार्ट में परिवर्तित करें?",
-    "Convert to hosted chart?": "होस्टेड चार्ट में कनवर्ट करें?",
-    Copy: "कॉपी करें",
-    "Copy PNG Image": "PNG छवि कॉपी करें",
-    "Copy SVG Code": "SVG कोड कॉपी करें",
-    "Copy your mermaid.js code or open it directly in the mermaid.js live editor.":
-      "अपना मर्मेड.जेएस कोड कॉपी करें या मर्मेड.जेएस लाइव एडिटर में सीधे खोलें।",
-    "Create Flowcharts from a Prompt with AI":
-      "एआई के साथ प्रॉम्प्ट से डायग्राम बनाएं",
-    "Create New Flowchart": "नया फ्लोचार्ट बनाएं",
-    "Create a New Flowchart": "एक नया आरेख बनाएँ",
-    "Create professional-quality visual aids for presentations or training materials":
-      "प्रदर्शन या प्रशिक्षण सामग्री के लिए पेशेवर गुणवत्ता के दृष्टिकोण बनाएं",
-    "Creating an edge between two nodes is done by indenting the second node below the first":
-      "दो नोड्स के बीच एक एज बनाने के लिए दूसरा नोड पहले वाले के नीचे इंडेंट करना होता है",
-    "Custom Sharing Options": "कस्टम शेयरिंग विकल्प",
-    "Customer Portal": "ग्राहक पोर्टल",
-    "Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":
-      "नीचे दिए गए <0>Cytoscape CSS</0> को संशोधित करके अपनी थीम को अनुकूलित करें। हमारी स्टाइलिंग दस्तावेज़ की तैयारी जल्द ही होगी! इस बीच, सर्वश्रेष्ठ स्रोत <1>Cytoscape</1> दस्तावेज़ है। अगर आप संकट में पड़ जाते हैं तो डिस्कोर्ड में सवाल पूछने आएं।",
-    Dark: "डार्क",
-    "Dark Mode": "डार्क मोड",
-    Date: "तारीख़",
-    "Deleted when browser data is cleared":
-      "ब्राउज़र डेटा साफ़ होने पर हटा दिया गया",
-    "Describe the flowchart you wish to create":
-      "उस आरेख का वर्णन करें जिसे आप बनाना चाहते हैं",
-    Direction: "दिशा",
-    Discard: "रद्द करें",
-    "Do you want to cancel your subscription?":
-      "क्या आप अपनी सदस्यता रद्द करना चाहते हैं?",
-    "Do you want to delete this?": "क्या आप इसे डिलीट करना चाहते हैं?",
-    Document: "दस्तावेज़",
-    "Don't have an account?": "खाता नहीं है?",
-    Download: "डाउनलोड",
-    "Download JPG": "JPG डाउनलोड करें",
-    "Download PNG": "PNG डाउनलोड करें",
-    "Download SVG": "SVG डाउनलोड करें",
-    "Drag and drop a CSV file here, or click to select a file":
-      "CSV फ़ाइल यहां ड्रैग और ड्रॉप करें, या फ़ाइल का चयन करने के लिए क्लिक करें",
-    "Draw an edge from multiple nodes by beginning the line with a reference":
-      "एक से अधिक नोड्स से एज ड्रा करने के लिए लाइन को एक संदर्भ के साथ शुरू करें",
-    "Drop the file here ...": "फाइल यहाँ ड्रॉप करें ...",
-    "Easily import existing flowcharts from other software":
-      "अन्य सॉफ्टवेयर से मौजूदा फ्लोचार्ट आसानी से आयात करें",
-    Edge: "किनारा",
-    "Edge ID, Classes, Attributes": "किनारे आईडी, वर्ग, गुण",
-    "Edge Label": "किनारे लेबल",
-    "Edge Label Column": "किनारे लेबल कॉलम",
-    "Edge Style": "किनारे शैली",
-    "Edge missing indentation": "एड्ज अंतराल गुम है",
-    Edges: "किन्तुओं",
-    "Edges are declared in the same row as their source node":
-      "उनके स्रोत नोड के ही पंक्ति में किन्तुओं की घोषणा की जाती है",
-    "Edges are declared in the same row as their target node":
-      "उनके लक्ष्य नोड के ही पंक्ति में किन्तुओं की घोषणा की जाती है",
-    "Edges are declared in their own row":
-      "किन्तुओं को अपनी खुद की पंक्ति में घोषणा की जाती है",
-    "Edges can also have ID's, classes, and attributes before the label":
-      "लेबल से पहले, किन्तुओं को आईडीज़, क्लासेज़ और गुण देने की अनुमति होती है",
-    "Edges can be styled with dashed, dotted, or solid lines":
-      "किन्तुओं को डैश्ड, डॉटेड या सॉलिड लाइन्स के साथ स्टाइल किया जा सकता है",
-    "Edges in Separate Rows": "अलग पंक्तियों में किन्हें",
-    "Edges in Source Node Row": "स्रोत नोड पंक्ति में किन्हें",
-    "Edges in Target Node Row": "लक्ष्य नोड पंक्ति में किन्हें",
-    Editable: "संपादन योग्य",
-    Editor: "संपादक",
-    Eggs: "अंडे",
-    Email: "ईमेल",
-    Empty: "खाली",
-    "Enter your email": "अपना ईमेल दर्ज करें",
-    "Equal To": "बराबर",
-    "Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":
-      "निबंध लेखन प्रक्रिया फ्लोचार्ट, छात्रों को ब्रैनस्टॉर्मिंग, आउटलाइनिंग, ड्राफ्टिंग और संशोधन चरणों में मार्गदर्शित करना।",
-    "Everything in Free": "फ्री प्लान में सब कुछ",
-    Excalidraw: "Excalidraw",
-    Export: "एक्सपोर्ट करें",
-    "Export High-Resolution Images": "उच्च-रिज़ॉल्यूशन छवियां निर्यात करें",
-    "Export to SVG": "SVG को निर्यात करें",
-    "Extra Large": "अति बड़ा",
-    Extract: "निकालना",
-    Feedback: "फ़ीडबैक",
-    "Follow Us on Twitter": "हमारे साथ ट्विटर पर फॉलो करें",
-    "Found a bug? Have a feature request? We would love to hear from you!":
-      "एक बग पाया? एक सुविधा अनुरोध है? हम आपसे सुनने के लिए प्रसन्न होंगे!",
-    Free: "मुफ्त",
-    Fullscreen: "फ़ुलस्क्रीन",
-    Futuristic: "भविष्य का",
-    "Get More Layouts": "ज़्यादा लेआउट पाएं",
-    "Get personalized attention and support from our expert team":
-      "हमारे विशेषज्ञ टीम से व्यक्तिगत ध्यान और समर्थन प्राप्त करें",
-    "Getting Started": "शुरू करना",
-    Graph: "ग्राफ़",
-    Grid: "ग्रिड",
-    "Have complex questions or issues? We're here to help.":
-      "जटिल सवाल या समस्याएं हैं? हम यहां आपकी मदद के लिए हैं।",
-    Help: "मदद",
-    History: "हिस्ट्री",
-    Home: "होम",
-    "How are edges declared in this data?":
-      "इस डेटा में कैसे किन्हीं कड़ियाँ घोषित होती हैं?",
-    "ID's": "आईडीज़",
-    "If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":
-      "यदि आप <0>Flowchart Fun</0> का उपयोग करना आनंदित हैं, तो कृपया परियोजना को समर्थित करने की सोच करें",
-    "If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\:</0>":
-      "यदि आप एक एड्ज बनाने के लिए मतलब है, तो इस लाइन को इंडेंट करें। यदि नहीं, तो कॉलन को बैक स्लैश के साथ भागो <0> \\: </0>",
-    "Image Export": "छवि निर्यात",
-    Images: "छवियाँ",
-    "Import Data": "डेटा आयात करें",
-    "Import data from a CSV file.": "CSV फ़ाइल से डेटा आयात करें।",
-    "Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":
-      "किसी भी CSV फ़ाइल से डेटा आयात करें और इसे एक नए प्रक्रिया चार्ट में मैप करें। यह अन्य स्रोतों जैसे लुसिडचार्ट, गूगल शीट्स और विसिओ से डेटा आयात करने के लिए एक अच्छा तरीका है।",
-    "Import from Lucidchart, Visio, and other CSV files":
-      "लुसिडचार्ट, विसिओ और अन्य CSV फ़ाइलों से आयात करें",
-    "Import your diagram it into Microsoft Visio using one of these CSV files.":
-      "इन CSV फ़ाइलों में से एक का उपयोग करके अपनी आरेख Microsoft Visio में आयात करें।",
-    "Important Changes are Coming to Flowchart Fun.":
-      "फ्लोचार्ट फन के लिए महत्वपूर्ण बदलाव आ रहे हैं।",
-    "Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":
-      "डेटा आयात एक प्रो फीचर है। आप केवल $ 3/महीने या $ 30/वर्ष के लिए फ्लोचार्ट फन प्रो में अपग्रेड कर सकते हैं।",
-    "Impress your audience with high-quality images produced by Flowchart Fun":
-      "फ्लोचार्ट फन द्वारा उत्पादित उच्च गुणवत्ता की छवियों से अपने दर्शकों को आकर्षित करें",
-    "Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":
-      "एक <0>title</0> विशेषता का उपयोग करके एक शीर्षक शामिल करें। Visio रंगीनी का उपयोग करने के लिए, निम्नलिखित में से किसी एक के बराबर एक <1>roleType</1> विशेषता जोड़ें:",
-    "Indenting creates an edge": "इंडेंटिंग एक बढ़त बनाता है",
-    Info: "जानकारी",
-    Instruct: "हिदायत",
-    Is: "हाँ",
-    Language: "भाषा",
-    Large: "बड़ा",
-    Layout: "ख़ाका",
-    "Layout Frozen": "लेआउट फ्रोज़न",
-    "Layout is Frozen": "लेआउट जम गया है",
-    "Leading References": "प्रमुख संदर्भ",
-    "Learn More": "और अधिक जानें",
-    "Learn Syntax": "सिंटैक्स सीखें",
-    "Learn about Flowchart Fun Pro": "Flowchart Fun Pro के बारे में जानें",
-    "Left to Right": "बाएं से दाएं",
-    "Let's get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":
-      "चलो शुरू करें! अपना फ्लोचार्ट फन खाता बनाने और हमारे शक्तिशाली विशेषताओं का उपयोग शुरू करने के लिए नीचे अपना ईमेल पता दर्ज करें।",
-    Light: "लाइट",
-    "Light Mode": "लाइट मोड",
-    Link: "लिंक",
-    "Load Theme": "थीम लोड करें",
-    "Loading a theme will erase any custom styles you have applied":
-      "थीम लोड करने से आपने अपने कस्टम स्टाइल को मिटा दिया होगा",
-    "Lock Zoom to Graph": "ग्राफ पर जूम लॉक करें",
-    "Log In": "लॉग इन करें",
-    "Log Out": "लॉग आउट",
-    "Log in to start creating flowcharts.":
-      "फ्लोचार्ट बनाने शुरू करने के लिए लॉग इन करें।",
-    "Make a One-Time Donation": "एक बार दान करें",
-    "Make publicly accessible": "सार्वजनिक रूप से एक्सेस दें",
-    "Make sure you use the same email you will use to log in.":
-      "उसी ईमेल का उपयोग करें जिसे आप लॉग इन करने के लिए उपयोग करेंगे।",
-    "Map Data": "डेटा मैप",
-    "Market understanding and competitive landscape maintenance for SaaS product development":
-      "SaaS उत्पाद विकास के लिए बाजार की समझ और प्रतिस्पर्धात्मक भूमिका बनाए रखना।",
-    Medium: "मध्यम",
-    Method: "विधि",
-    Monospace: "Monospace",
-    Monthly: "मासिक",
-    "Multiple pointers on same line": "एक ही लाइन पर कई प्रतीक",
-    Museum: "म्यूज़ियम",
-    Name: "नाम",
-    New: "नया",
-    "New Email": "नई ईमेल",
-    "Next charge": "अगला चार्ज",
-    "No Edges": "कोई किनारे नहीं",
-    Node: "नोड",
-    "Node Border Style": "नोड सीमा शैली",
-    "Node Colors": "नोड रंग",
-    "Node ID": "नोड आईडी",
-    "Node ID, Classes, Attributes": "नोड आईडी, वर्ग, गुण",
-    "Node Label": "नोड लेबल",
-    "Node Shapes": "नोड आकार",
-    "Node and pointer on same line": "नोड और प्रतीक एक ही लाइन पर",
-    "Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":
-      "नोड्स को डैश्ड, डॉट्ड या डबल के साथ स्टाइल किया जा सकता है। बॉर्डर-नोन के साथ भी बॉर्डर हटाए जा सकते हैं।",
-    "Not Empty": "खाली नहीं",
-    "Office Hours": "कार्यालय अवधि",
-    "Once in a while the magic link will end up in your spam folder. If you don't see it after a few minutes, check there or request a new link.":
-      "कभी-कभी मैजिक लिंक आपके स्पैम फोल्डर में खत्म हो जाता है। अगर आप कुछ मिनट बाद उसे नहीं देख रहे हैं, तो वहां जाकर देखें या एक नया लिंक अनुरोध करें।",
-    "One-on-One Support": "एक-पर-एक समर्थन",
-    "One-time Share Links": "वन-टाइम शेयर लिंक",
-    "Only available on this device.<0/>Clearing your browser cache will erase them.":
-      "केवल इस उपकरण पर उपलब्ध है।<0/>अपने ब्राउज़र कैश साफ़ करने से उन्हें मिटा दिया जाएगा।",
-    "Open Customer Portal": "ग्राहक पोर्टल खोलें",
-    "Organization Chart": "संगठन चार्ट",
-    "Paste the information you wish to convert to a flowchart":
-      "उस जानकारी को पेस्ट करें जिसे आप फ़्लोचार्ट में बदलना चाहते हैं",
-    Permalinks: "Permalinks",
-    "Permanent Charts are a Pro Feature": "स्थायी चार्ट एक प्रो सुविधा हैं",
-    "Permanent Flowcharts": "लगातार चार्ट",
-    Persistent: "अविरल",
-    "Persistent Flowcharts": "लगातार आरेख",
-    Playbook: "प्लेबुक",
-    "Pointer and container on same line": "प्रतीक और कंटेनर एक ही लाइन पर",
-    Pricing: "मूल्य निर्धारण",
-    "Process for corporate social responsibility initiatives development and implementation across company operations":
-      "कंपनी की सार्वजनिक दायित्व परियोजनाओं के विकास और लागू करने की प्रक्रिया।",
-    "Processing Data": "डेटा प्रसंस्करण",
-    Prompt: "प्रॉम्प्ट",
-    Public: "पब्लिक",
-    Random: "अनियमित",
-    "Read-only": "केवल पढ़ने के लिए",
-    "Reference by Class": "क्लास के द्वारा संदर्भ",
-    "Reference by ID": "आईडी द्वारा संदर्भ",
-    "Reference by Label": "लेबल द्वारा संदर्भ",
-    References: "संदर्भ",
-    "References are used to create edges between nodes that are created elsewhere in the document":
-      "संदर्भ दस्तावेज के अन्य भागों में बनाए गए नोड्स के बीच कड़ियाँ बनाने के लिए उपयोग किए जाते हैं",
-    "Referencing a node by its exact label":
-      "सटीक लेबल द्वारा नोड का संदर्भ करना",
-    "Referencing a node by its unique ID":
-      "अद्वितीय आईडी द्वारा नोड का संदर्भ करना",
-    "Referencing multiple nodes with the same assigned class":
-      "एक ही निर्धारित कक्षा के साथ कई नोड का उल्लेख करना",
-    "Refresh Page": "पृष्ठ को ताज़ा करें",
-    "Reload to Update": "अपडेट करने के लिए रीलोड करें",
-    "Remove Image Watermark": "छवि वॉटरमार्क हटाएं",
-    Rename: "फिर से नाम बदलें",
-    Reset: "रीसेट करें",
-    "Resume Subscription": "सदस्यता फिर से शुरू करें",
-    Retro: "रेट्रो",
-    Return: "वापसी",
-    "Right to Left": "दाएं से बाएं",
-    Roadmap: "रोडमैप",
-    Save: "सहेजें",
-    "Save 20% (2 months free!)": "20% बचत (2 महीने मुफ्त!)",
-    "Save time and automate certain tasks with this cutting-edge feature":
-      "इस उन्नत विशेषता के साथ समय बचाएं और कुछ कार्यों को स्वचालित करें",
-    "Save time and streamline your workflow by importing data from other sources":
-      "अन्य स्रोतों से डेटा आयात करके अपनी व्यवस्था को समय बचाएं और स्ट्रीमलाइन करें",
-    "Select your plan!": "अपना योजना चुनें!",
-    "Send Feedback": "प्रतिक्रिया भेजें",
-    Settings: "सेटिंग",
-    Shape: "आकृति",
-    "Sign Up": "साइन अप",
-    "Sign Up Now": "अभी साइन अप करें",
-    Size: "आकार",
-    Small: "छोटा",
-    "Some classes are available to help style your flowchart":
-      "कुछ क्लास आपकी फ्लोचार्ट स्टाइल करने में मदद करने के लिए उपलब्ध हैं",
-    "Sorry! This page is only available in English.":
-      "माफ़ करें! यह पेज केवल अंग्रेजी में उपलब्ध है।",
-    "Source Arrow": "स्रोत तीर ",
-    "Source Column": "स्रोत स्तंभ",
-    "Source Delimiter": "स्रोत डिलिमिटर",
-    "Source/Target Arrow Shape": "स्रोत / लक्ष्य तीर आकार",
-    Spacing: "अंतर",
-    "Special Attributes": "विशेष गुण",
-    Start: "शुरू",
-    "Start Over": "शुरू करें",
-    Status: "स्टेटस",
-    "Store any data associated to a node":
-      "किसी नोड से संबंधित किसी भी डेटा स्टोर करें",
-    "Stored in the cloud": "बादल में संग्रहित",
-    "Stored on this computer": "इस कंप्यूटर पर संग्रहीत",
-    "Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":
-      "फ्लोचार्ट फन के साथ अपनी व्यवस्था को स्ट्रीमलाइन करें और प्रक्रिया विजुअलाइजेशन को सरल बनाएं। अब सदस्यता लें और हमारे शक्तिशाली प्रो विशेषताओं का लाभ उठाएं!",
-    Style: "शैली",
-    "Style Classes": "शैली क्लासेज",
-    Submit: "भेजना",
-    Subscribe: "सदस्यता लें",
-    Subscription: "सदस्यता",
-    "Subscription will end": "सदस्यता समाप्त हो जाएगी",
-    "Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":
-      "आपूर्ति श्रृंखला विश्लेषण और अनुकूलन: लागत कमी, कार्यक्षमता में सुधार और स्टेकहोल्डर सहयोग।",
-    "Target Arrow": "लक्ष्य तीर",
-    "Target Column": "लक्ष्य कॉलम",
-    "Target Delimiter": "लक्ष्य डेलीमीटर",
-    Temporary: "अस्थायी",
-    "Temporary Flowchart Warning": "अस्थायी पारिस्थितिकी चेतावनी",
-    "Temporary Flowcharts": "अस्थायी आरेख",
-    "Text Sizes": "पाठ आकार",
-    "Text followed by colon+space creates an edge with the text as the label":
-      "स्क्लीन के बाद कोलन + स्पेस एक किरदार बनाता है जिसका पाठ लेबल है",
-    "Text on a line creates a node with the text as the label":
-      "एक पंक्ति पर पाठ एक नोड बनाता है जिसका पाठ लेबल है",
-    "Text sizes include small, regular, large, and extra-large.":
-      "पाठ के आकार छोटे, सामान्य, बड़े और अतिरिक्त बड़े शामिल हैं।",
-    "Thank you for your feedback!": "आपके फ़ीडबैक के लिए धन्यवाद!",
-    "The column that contains the edge label(s)":
-      "कॉलम जो एड्ज लेबल (ओं) को शामिल करता है",
-    "The column that contains the source node ID(s)":
-      "कॉलम जो स्रोत नोड आईडी (ओं) को शामिल करता है",
-    "The column that contains the target node ID(s)":
-      "कॉलम जो लक्ष्य नोड आईडी (ओं) को शामिल करता है",
-    "The delimiter used to separate multiple source nodes":
-      "कई स्रोत नोड्स को अलग करने के लिए उपयोग किया गया डिलिमिटर",
-    "The delimiter used to separate multiple target nodes":
-      "कई लक्ष्य नोड्स को अलग करने के लिए उपयोग किया गया डिलिमिटर",
-    "The possible shapes are:": "संभव आकृतियाँ हैं:",
-    "Theme Editor": "थीम एडिटर",
-    "There are no edges in this data": "इस डेटा में कोई एड्ज नहीं है",
-    "This action cannot be undone.": "इस क्रिया को पूर्ववत नहीं किया जा सकता।",
-    "This app works by typing": "यह ऐप टाइप द्वारा काम करता है",
-    "This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":
-      "यह दस्तावेज़ केवल इस कंप्यूटर पर सहेजी गई है। यदि आप दस्तावेज़ सहेजने और कहीं भी उपयोग करने के लिए एक तरीका चाहते हैं तो हमारे प्रो प्लान के लिए केवल $ 3 / महीने में उपयोग करें!",
-    "This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":
-      "यह सुविधा केवल प्रो उपयोगकर्ताओं के लिए उपलब्ध है। <0>प्रो उपयोगकर्ता बनें</0> इसे अनलॉक करने के लिए।",
-    "This is a temporary chart, it will be deleted when you clear your browser cache.":
-      "यह एक अस्थायी चार्ट है, जब आप अपने ब्राउज़र कैश को साफ़ करेंगे तो यह हटा दिया जाएगा।",
-    Tip: "सुझाव",
-    "To fix this change one of the edge IDs":
-      "इसे ठीक करने के लिए एड्ज आईडी के एक को बदलें",
-    "To fix this change one of the node IDs":
-      "इसे ठीक करने के लिए नोड आईडी के एक को बदलें",
-    "To fix this move one pointer to the next line":
-      "इसे ठीक करने के लिए प्रतीक को अगली पंक्ति पर ले जाएं",
-    "To fix this move the pointer to the next line":
-      "इसे ठीक करने के लिए प्रतीक को अगली पंक्ति पर ले जाएं",
-    "To fix this start the container <0/> on a different line":
-      "इसे ठीक करने के लिए कंटेनर <0/> को एक अलग पंक्ति पर शुरू करें",
-    "Top to Bottom": "ऊपर से नीचे",
-    "Try again": "फिर से कोशिश करें",
-    "Two edges have the same ID": "दो किनारे उसी आईडी के हैं",
-    "Two nodes have the same ID": "दो नोड उसी आईडी के हैं",
-    Type: "प्रकार",
-    Unfreeze: "अनफ़्रीज़",
-    "Unique text value to identify a node":
-      "एक नोड को पहचानने के लिए अद्वितीय पाठ मूल्य",
-    "Unknown Parsing Error": "अज्ञात पार्सिंग त्रुटि",
-    "Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":
-      "अपनी क्रेयता को उतारें और फ्लोचार्ट फन प्रो के साथ अपनी व्यवस्था को स्ट्रीमलाइन करें – सिर्फ $ 3 / महीने से शुरू होते हुए!",
-    "Update Email": "ईमेल अपडेट करें",
-    "Upgrade to Pro": "प्रो में अपग्रेड करें",
-    "Upload your File": "अपनी फाइल अपलोड करें",
-    "Use AI to create flowcharts from data sets or algorithms":
-      "डेटा सेट्स या एल्गोरिदम्स से फ्लोचार्ट बनाने के लिए एआई का उपयोग करें",
-    "Use AI to generate a flowchart from a prompt. <0>Learn More</0>":
-      "एक प्रॉम्प्ट से एक फ्लोचार्ट उत्पन्न करने के लिए AI का उपयोग करें। <0>और अधिक जानें</0>",
-    "Use classes to group nodes":
-      "नोड्स को समूहों में सम्मिलित करने के लिए वर्ग उपयोग करें",
-    "Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":
-      "नोड पर एक लिंक सेट करने के लिए गुण <0>href</0> उपयोग करें जो एक नये टैब में खुलेगा।",
-    "Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":
-      "नोड की छवि सेट करने के लिए <0>src</0> गुण का उपयोग करें। छवि नोड के आकार में स्केल की जाएगी, इसलिए आपको आवश्यक प्रतिस्थापित करने के लिए नोड की चौड़ाई और ऊँचाई को समायोजित करना होगा। केवल सार्वजनिक छवियाँ (CORS द्वारा ब्लॉक नहीं की गई) समर्थित हैं।",
-    "Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":
-      "नोड की विशिष्ट चौड़ाई और ऊँचाई सेट करने के लिए <0>w</0> और <1>h</1> गुण का उपयोग करें।",
-    "Use the customer portal to change your billing information.":
-      "अपनी बिलिंग जानकारी बदलने के लिए ग्राहक पोर्टल का उपयोग करें।",
-    "Use this file for org charts, hierarchies, and other organizational structures.":
-      "संगठनात्मक चार्ट, पदानुक्रमों और अन्य संगठनात्मक संरचनाओं के लिए इस फ़ाइल का उपयोग करें।",
-    "Use this file for sequences, processes, and workflows.":
-      "अनुक्रमों, प्रक्रियाओं और कार्यप्रवाहों के लिए इस फ़ाइल का उपयोग करें।",
-    User: "यूज़र",
-    "View on Github": "Github पर देखें",
-    "Water evaporates from the Earth's surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":
-      "पानी पृथ्वी के सतह से उष्णता होकर आत्मान में उठता है और वर्षा के रूप में फिर नीचे गिरता है। इस पानी को तब नदियों, तिब्बती तालों और महासागरों में जाता है, जहां यह फिर से उष्णता होता है और आत्मान में पुनर्स्थापित किया जाता है।",
-    "We use magic links to log you in. Enter your email below to get started.":
-      "हम आपको लॉग इन करने के लिए जादू लिंक का उपयोग करते हैं। नीचे अपना ईमेल दर्ज करें और शुरू करें।",
-    "Welcome to Flowchart Fun Pro!": "फ्लोचार्ट फन प्रो के लिए स्वागत है!",
-    "Width and Height": "चौड़ाई और ऊँचाई",
-    "Would you like to continue?": "क्या आप जारी रखना चाहते हैं?",
-    Yearly: "वार्षिक",
-    "You already have a flowchart with this name.":
-      "आपके पास इस नाम का आरेख पहले से है।",
-    "You are about to add {numNodes} nodes and {numEdges} edges to your graph.":
-      [
-        "आप अपने ग्राफ में ",
-        ["numNodes"],
-        " नोड्स और ",
-        ["numEdges"],
-        " एड्ज़ जोड़ने वाले हैं।",
-      ],
-    'You can also wrap text in "\\( \\)" to create an edge':
-      'किनारे बनाने के लिए आप टेक्स्ट को "\\( \\)" में भी रैप कर सकते हैं',
-    "You must log in to create a persistent flowchart.":
-      "आपको एक अविरल फ्लोचार्ट बनाने के लिए लॉग इन करना होगा।",
-    "You're almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":
-      "आप बस एक ही कदम दूर हैं! <0>Flowchart Fun Pro</0> की पूरी प्रत्याशित क्षमता को अनलॉक करने के लिए नीचे अपने भुगतान विवरण दर्ज करें और आज ही अद्भुत फ्लोचार्ट बनाने शुरू करें।",
-    "Your Charts": "आपके चार्ट",
-    "Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":
-      "आपका सदस्यता अब निष्क्रिय है। यदि आप स्थायी चार्ट बनाना और संपादित करना चाहते हैं, तो प्रो अपग्रेड करें।",
-    "Zoom In": "ज़ूम इन करें",
-    "Zoom Out": "आउट ज़ूम करें",
-    "any text: before a colon creates a label":
-      "कोई भी टेक्स्ट: कॉलन से पहले एक लेबल बनाता है",
-  },
+  messages: JSON.parse(
+    '{"$3 per month":"मासिक $3","$30 per year":"वार्षिक $30","(before a colon creates a label)":"(कॉलन से पहले एक लेबल बनाता है)","<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":"<0>Flowchart Fun</0> <1>Tone Row</1> द्वारा बनाई गई एक खुला स्रोत परियोजना है ","<0>Sign In</0> / <1>Sign Up</1> with email and password":"<0>साइन इन</0> / <1>साइन अप</1> ईमेल और पासवर्ड के साथ","<0>Starting August 28th you will need a subscription to create and edit charts. <1>Learn more</1>.</0>":"<0>28 अगस्त से आपको चार्ट बनाने और संपादित करने के लिए एक सदस्यता की आवश्यकता होगी। <1>और अधिक जानें</1>।</0>","<0>You currently have a free account.<1/><2>Learn about our Pro Features and subscribe on our pricing page</2>.</0>":"<0>आपके पास वर्तमान में एक मुफ्त खाता है। <1/><2>हमारी प्रो फीचर्स के बारे में जानें और हमारी मुल्यों के पृष्ठ पर सदस्यता लें</2>।</0>","A new version of the app is available. Please reload to update.":"एप्प का एक नया वर्जन उपलब्ध है। अपडेट करने के लिए रीलोड करें।","About":"के बारे में","Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":"इन चार्ट्स को कहीं से भी एक्सेस करें।<0/>सिंक्रोनाइज़्ड रहने वाले फ्लोचार्ट को साझा करें और एम्बेड करें।","Accessible from any device":"किसी भी उपकरण से पहुँचा जा सकता है","Account":"खाता","Activate your account":"अपना खाता सक्रिय करें","Advanced":"विकसित","Amount":"रकम","An error occurred. Try resubmitting or email {0} directly.":["एक एरर हो गया. फिर से सबमिट करने की कोशिश करें या सीधे ",["0"]," ईमेल करें."],"Appearance":"दिखावट","Apply Style":"शैली लागू करें","Attributes":"गुण","Auto-Layouts":"ऑटो लेआउट","Back To Editor":"संपादक पर वापस जाएं","Basic Flowchart":"बेसिक फ्लोचार्ट","Become a Github Sponsor":"गिटहब स्पॉन्सर बनें","Become a Pro User":"प्रो उपयोगकर्ता बनें","Begin with a simple example showing how <0>Flowchart Fun</0> works.":"<0>फ्लोचार्ट फन</0> काम कैसे करता है इसके लिए एक आसान उदाहरण से शुरू करें।","Blank":"खाली","Blog":"ब्लॉग","Book a Meeting":"एक बैठक बुक करें","Border":"सीमा","Bottom to Top":"नीचे से शीर्ष तक","Breadthfirst":"चौड़ाई पहले","Cancel":"रद्द करें","Cancel your subscription. Your hosted charts will become read-only.":"अपनी सदस्यता रद्द करें. आपके होस्ट किये गए चार्ट सिर्फ़ पढ़े जा सकेंगे.","Certain attributes can be used to customize the appearance or functionality of elements.":"कुछ गुण तत्वों की दिखता या कार्यक्षमता को अनुकूलित करने के लिए उपयोग किए जा सकते हैं।","Change Email Address":"ईमेल पता बदलें","Changelog":"बदलाव का","Charts":"चार्ट","Check your email for a link to log in.<0/>You can close this window.":"लॉग इन करने के लिए अपने ईमेल की जाँच करें। आप इस विंडो को बंद कर सकते हैं।","Choose":"चुनें","Choose a Theme":"एक थीम चुनें","Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":"एक किनारे के स्रोत और लक्ष्य के लिए विभिन्न तीर आकारों से चुनें। आकार शामिल हैं त्रिकोण, त्रिकोण-टी, सर्कल-त्रिकोण, त्रिकोण-क्रॉस, त्रिकोण-बैककर्व, वी, टी, चौकोर, हीरा, चेवरॉन और कोई नहीं।","Choose the plan that\'s right for you and start creating amazing flowcharts with Flowchart Fun Pro":"आपके लिए सही योजना चुनें और Flowchart Fun Pro के साथ अद्भुत फ्लोचार्ट बनाना शुरू करें","Circle":"परिधि","Classes":"वर्ग","Clay":"मिट्टी","Clear":"साफ़","Clear text?":"पाठ साफ़ करें?","Click on \\\\nLearn Syntax\\\\nto learn more":"\\\\nसिंटैक्स सीखने के लिए \\\\nपर क्लिक करें\\\\n","Click to preview":"पूर्वावलोकन के लिए क्लिक करें","Clone":"क्लोन करें","Close":"बंद करें","Color":"रंग","Colors include red, orange, yellow, blue, purple, black, white, and gray.":"रंग में लाल, नारंगी, पीला, नीला, बैंगनी, काला, सफेद, और ग्रे शामिल हैं।","Column":"स्तंभ","Comic Book":"कॉमिक बुक","Comment":"टिप्पणी","Concentric":"गाढ़ा","Confirm New Email":"नई ईमेल की पुष्टि करें","Confirm your email address to sign in.":"साइन इन करने के लिए अपना ईमेल पता पुष्टि करें।","Containers":"कंटेनर","Containers are nodes that contain other nodes. They are declared using curly braces.":"कंटेनर उन नोड्स हैं जो अन्य नोड्स को सम्मिलित करते हैं। वे कर्ली ब्रेस का उपयोग करके घोषित किया जाता है।","Continue":"जारी रखें","Convert this to a hosted chart?":"यह एक होस्ट की गई चार्ट में परिवर्तित करें?","Convert to hosted chart?":"होस्टेड चार्ट में कनवर्ट करें?","Copy":"कॉपी करें","Copy PNG Image":"PNG छवि कॉपी करें","Copy SVG Code":"SVG कोड कॉपी करें","Copy your mermaid.js code or open it directly in the mermaid.js live editor.":"अपना मर्मेड.जेएस कोड कॉपी करें या मर्मेड.जेएस लाइव एडिटर में सीधे खोलें।","Create Flowcharts from a Prompt with AI":"एआई के साथ प्रॉम्प्ट से डायग्राम बनाएं","Create New Flowchart":"नया फ्लोचार्ट बनाएं","Create a New Flowchart":"एक नया आरेख बनाएँ","Create professional-quality visual aids for presentations or training materials":"प्रदर्शन या प्रशिक्षण सामग्री के लिए पेशेवर गुणवत्ता के दृष्टिकोण बनाएं","Creating an edge between two nodes is done by indenting the second node below the first":"दो नोड्स के बीच एक एज बनाने के लिए दूसरा नोड पहले वाले के नीचे इंडेंट करना होता है","Custom Sharing Options":"कस्टम शेयरिंग विकल्प","Customer Portal":"ग्राहक पोर्टल","Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":"नीचे दिए गए <0>Cytoscape CSS</0> को संशोधित करके अपनी थीम को अनुकूलित करें। हमारी स्टाइलिंग दस्तावेज़ की तैयारी जल्द ही होगी! इस बीच, सर्वश्रेष्ठ स्रोत <1>Cytoscape</1> दस्तावेज़ है। अगर आप संकट में पड़ जाते हैं तो डिस्कोर्ड में सवाल पूछने आएं।","Dark":"डार्क","Dark Mode":"डार्क मोड","Date":"तारीख़","Deleted when browser data is cleared":"ब्राउज़र डेटा साफ़ होने पर हटा दिया गया","Describe the flowchart you wish to create":"उस आरेख का वर्णन करें जिसे आप बनाना चाहते हैं","Direction":"दिशा","Discard":"रद्द करें","Do you want to cancel your subscription?":"क्या आप अपनी सदस्यता रद्द करना चाहते हैं?","Do you want to delete this?":"क्या आप इसे डिलीट करना चाहते हैं?","Document":"दस्तावेज़","Download":"डाउनलोड","Download JPG":"JPG डाउनलोड करें","Download PNG":"PNG डाउनलोड करें","Download SVG":"SVG डाउनलोड करें","Drag and drop a CSV file here, or click to select a file":"CSV फ़ाइल यहां ड्रैग और ड्रॉप करें, या फ़ाइल का चयन करने के लिए क्लिक करें","Draw an edge from multiple nodes by beginning the line with a reference":"एक से अधिक नोड्स से एज ड्रा करने के लिए लाइन को एक संदर्भ के साथ शुरू करें","Drop the file here ...":"फाइल यहाँ ड्रॉप करें ...","Easily import existing flowcharts from other software":"अन्य सॉफ्टवेयर से मौजूदा फ्लोचार्ट आसानी से आयात करें","Edge":"किनारा","Edge ID, Classes, Attributes":"किनारे आईडी, वर्ग, गुण","Edge Label":"किनारे लेबल","Edge Label Column":"किनारे लेबल कॉलम","Edge Style":"किनारे शैली","Edge missing indentation":"एड्ज अंतराल गुम है","Edges":"किन्तुओं","Edges are declared in the same row as their source node":"उनके स्रोत नोड के ही पंक्ति में किन्तुओं की घोषणा की जाती है","Edges are declared in the same row as their target node":"उनके लक्ष्य नोड के ही पंक्ति में किन्तुओं की घोषणा की जाती है","Edges are declared in their own row":"किन्तुओं को अपनी खुद की पंक्ति में घोषणा की जाती है","Edges can also have ID\'s, classes, and attributes before the label":"लेबल से पहले, किन्तुओं को आईडीज़, क्लासेज़ और गुण देने की अनुमति होती है","Edges can be styled with dashed, dotted, or solid lines":"किन्तुओं को डैश्ड, डॉटेड या सॉलिड लाइन्स के साथ स्टाइल किया जा सकता है","Edges in Separate Rows":"अलग पंक्तियों में किन्हें","Edges in Source Node Row":"स्रोत नोड पंक्ति में किन्हें","Edges in Target Node Row":"लक्ष्य नोड पंक्ति में किन्हें","Editable":"संपादन योग्य","Editor":"संपादक","Eggs":"अंडे","Email":"ईमेल","Empty":"खाली","Enter your email":"अपना ईमेल दर्ज करें","Enter your email address and we\'ll send you a magic link to sign in.":"अपना ईमेल पता दर्ज करें और हम आपको साइन इन करने के लिए एक जादूगरी लिंक भेजेंगे।","Enter your email address below and we\'ll send you a link to reset your password.":"नीचे अपना ईमेल पता दर्ज करें और हम आपको अपना पासवर्ड रीसेट करने के लिए एक लिंक भेजेंगे।","Equal To":"बराबर","Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":"निबंध लेखन प्रक्रिया फ्लोचार्ट, छात्रों को ब्रैनस्टॉर्मिंग, आउटलाइनिंग, ड्राफ्टिंग और संशोधन चरणों में मार्गदर्शित करना।","Everything in Free":"फ्री प्लान में सब कुछ","Excalidraw":"Excalidraw","Export":"एक्सपोर्ट करें","Export High-Resolution Images":"उच्च-रिज़ॉल्यूशन छवियां निर्यात करें","Export to SVG":"SVG को निर्यात करें","Extra Large":"अति बड़ा","Extract":"निकालना","Feedback":"फ़ीडबैक","Follow Us on Twitter":"हमारे साथ ट्विटर पर फॉलो करें","Forgot your password?":"अपना पासवर्ड भूल गए?","Found a bug? Have a feature request? We would love to hear from you!":"एक बग पाया? एक सुविधा अनुरोध है? हम आपसे सुनने के लिए प्रसन्न होंगे!","Free":"मुफ्त","Fullscreen":"फ़ुलस्क्रीन","Futuristic":"भविष्य का","Get More Layouts":"ज़्यादा लेआउट पाएं","Get personalized attention and support from our expert team":"हमारे विशेषज्ञ टीम से व्यक्तिगत ध्यान और समर्थन प्राप्त करें","Getting Started":"शुरू करना","Graph":"ग्राफ़","Grid":"ग्रिड","Have complex questions or issues? We\'re here to help.":"जटिल सवाल या समस्याएं हैं? हम यहां आपकी मदद के लिए हैं।","Help":"मदद","History":"हिस्ट्री","Home":"होम","How are edges declared in this data?":"इस डेटा में कैसे किन्हीं कड़ियाँ घोषित होती हैं?","ID\'s":"आईडीज़","If an account with that email exists, we\'ve sent you an email with instructions on how to reset your password.":"अगर उस ईमेल के साथ एक खाता मौजूद है, तो हमने आपको पासवर्ड रीसेट करने के लिए निर्देशों के साथ एक ईमेल भेजा है।","If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":"यदि आप <0>Flowchart Fun</0> का उपयोग करना आनंदित हैं, तो कृपया परियोजना को समर्थित करने की सोच करें","If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\\\:</0>":"यदि आप एक एड्ज बनाने के लिए मतलब है, तो इस लाइन को इंडेंट करें। यदि नहीं, तो कॉलन को बैक स्लैश के साथ भागो <0> \\\\: </0>","Image Export":"छवि निर्यात","Images":"छवियाँ","Import Data":"डेटा आयात करें","Import data from a CSV file.":"CSV फ़ाइल से डेटा आयात करें।","Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":"किसी भी CSV फ़ाइल से डेटा आयात करें और इसे एक नए प्रक्रिया चार्ट में मैप करें। यह अन्य स्रोतों जैसे लुसिडचार्ट, गूगल शीट्स और विसिओ से डेटा आयात करने के लिए एक अच्छा तरीका है।","Import from Lucidchart, Visio, and other CSV files":"लुसिडचार्ट, विसिओ और अन्य CSV फ़ाइलों से आयात करें","Import your diagram it into Microsoft Visio using one of these CSV files.":"इन CSV फ़ाइलों में से एक का उपयोग करके अपनी आरेख Microsoft Visio में आयात करें।","Important Changes are Coming to Flowchart Fun.":"फ्लोचार्ट फन के लिए महत्वपूर्ण बदलाव आ रहे हैं।","Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":"डेटा आयात एक प्रो फीचर है। आप केवल $ 3/महीने या $ 30/वर्ष के लिए फ्लोचार्ट फन प्रो में अपग्रेड कर सकते हैं।","Impress your audience with high-quality images produced by Flowchart Fun":"फ्लोचार्ट फन द्वारा उत्पादित उच्च गुणवत्ता की छवियों से अपने दर्शकों को आकर्षित करें","Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":"एक <0>title</0> विशेषता का उपयोग करके एक शीर्षक शामिल करें। Visio रंगीनी का उपयोग करने के लिए, निम्नलिखित में से किसी एक के बराबर एक <1>roleType</1> विशेषता जोड़ें:","Indenting creates an edge":"इंडेंटिंग एक बढ़त बनाता है","Info":"जानकारी","Instruct":"हिदायत","Is":"हाँ","Language":"भाषा","Large":"बड़ा","Layout":"ख़ाका","Layout Frozen":"लेआउट फ्रोज़न","Layout is Frozen":"लेआउट जम गया है","Leading References":"प्रमुख संदर्भ","Learn More":"और अधिक जानें","Learn Syntax":"सिंटैक्स सीखें","Learn about Flowchart Fun Pro":"Flowchart Fun Pro के बारे में जानें","Left to Right":"बाएं से दाएं","Let\'s get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":"चलो शुरू करें! अपना फ्लोचार्ट फन खाता बनाने और हमारे शक्तिशाली विशेषताओं का उपयोग शुरू करने के लिए नीचे अपना ईमेल पता दर्ज करें।","Light":"लाइट","Light Mode":"लाइट मोड","Link":"लिंक","Load Theme":"थीम लोड करें","Loading a theme will erase any custom styles you have applied":"थीम लोड करने से आपने अपने कस्टम स्टाइल को मिटा दिया होगा","Lock Zoom to Graph":"ग्राफ पर जूम लॉक करें","Log In":"लॉग इन करें","Log Out":"लॉग आउट","Log in to start creating flowcharts.":"फ्लोचार्ट बनाने शुरू करने के लिए लॉग इन करें।","Make a One-Time Donation":"एक बार दान करें","Make publicly accessible":"सार्वजनिक रूप से एक्सेस दें","Make sure you use the same email you will use to log in.":"उसी ईमेल का उपयोग करें जिसे आप लॉग इन करने के लिए उपयोग करेंगे।","Map Data":"डेटा मैप","Market understanding and competitive landscape maintenance for SaaS product development":"SaaS उत्पाद विकास के लिए बाजार की समझ और प्रतिस्पर्धात्मक भूमिका बनाए रखना।","Medium":"मध्यम","Method":"विधि","Monospace":"Monospace","Monthly":"मासिक","Multiple pointers on same line":"एक ही लाइन पर कई प्रतीक","Museum":"म्यूज़ियम","Name":"नाम","New":"नया","New Email":"नई ईमेल","Next charge":"अगला चार्ज","No Edges":"कोई किनारे नहीं","Node":"नोड","Node Border Style":"नोड सीमा शैली","Node Colors":"नोड रंग","Node ID":"नोड आईडी","Node ID, Classes, Attributes":"नोड आईडी, वर्ग, गुण","Node Label":"नोड लेबल","Node Shapes":"नोड आकार","Node and pointer on same line":"नोड और प्रतीक एक ही लाइन पर","Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":"नोड्स को डैश्ड, डॉट्ड या डबल के साथ स्टाइल किया जा सकता है। बॉर्डर-नोन के साथ भी बॉर्डर हटाए जा सकते हैं।","Not Empty":"खाली नहीं","Office Hours":"कार्यालय अवधि","Once in a while the magic link will end up in your spam folder. If you don\'t see it after a few minutes, check there or request a new link.":"कभी-कभी मैजिक लिंक आपके स्पैम फोल्डर में खत्म हो जाता है। अगर आप कुछ मिनट बाद उसे नहीं देख रहे हैं, तो वहां जाकर देखें या एक नया लिंक अनुरोध करें।","One-on-One Support":"एक-पर-एक समर्थन","One-time Share Links":"वन-टाइम शेयर लिंक","Only available on this device.<0/>Clearing your browser cache will erase them.":"केवल इस उपकरण पर उपलब्ध है।<0/>अपने ब्राउज़र कैश साफ़ करने से उन्हें मिटा दिया जाएगा।","Open Customer Portal":"ग्राहक पोर्टल खोलें","Organization Chart":"संगठन चार्ट","Password":"पासवर्ड","Paste the information you wish to convert to a flowchart":"उस जानकारी को पेस्ट करें जिसे आप फ़्लोचार्ट में बदलना चाहते हैं","Permalinks":"Permalinks","Permanent Charts are a Pro Feature":"स्थायी चार्ट एक प्रो सुविधा हैं","Permanent Flowcharts":"लगातार चार्ट","Persistent":"अविरल","Persistent Flowcharts":"लगातार आरेख","Playbook":"प्लेबुक","Pointer and container on same line":"प्रतीक और कंटेनर एक ही लाइन पर","Pricing":"मूल्य निर्धारण","Process for corporate social responsibility initiatives development and implementation across company operations":"कंपनी की सार्वजनिक दायित्व परियोजनाओं के विकास और लागू करने की प्रक्रिया।","Processing Data":"डेटा प्रसंस्करण","Prompt":"प्रॉम्प्ट","Public":"पब्लिक","Random":"अनियमित","Read-only":"केवल पढ़ने के लिए","Reference by Class":"क्लास के द्वारा संदर्भ","Reference by ID":"आईडी द्वारा संदर्भ","Reference by Label":"लेबल द्वारा संदर्भ","References":"संदर्भ","References are used to create edges between nodes that are created elsewhere in the document":"संदर्भ दस्तावेज के अन्य भागों में बनाए गए नोड्स के बीच कड़ियाँ बनाने के लिए उपयोग किए जाते हैं","Referencing a node by its exact label":"सटीक लेबल द्वारा नोड का संदर्भ करना","Referencing a node by its unique ID":"अद्वितीय आईडी द्वारा नोड का संदर्भ करना","Referencing multiple nodes with the same assigned class":"एक ही निर्धारित कक्षा के साथ कई नोड का उल्लेख करना","Refresh Page":"पृष्ठ को ताज़ा करें","Reload to Update":"अपडेट करने के लिए रीलोड करें","Remove Image Watermark":"छवि वॉटरमार्क हटाएं","Rename":"फिर से नाम बदलें","Request Magic Link":"अनुरोध जादू लिंक","Request Password Reset":"पासवर्ड रीसेट का अनुरोध करें","Reset":"रीसेट करें","Reset Password":"पासवर्ड रीसेट करें","Resume Subscription":"सदस्यता फिर से शुरू करें","Retro":"रेट्रो","Return":"वापसी","Right to Left":"दाएं से बाएं","Roadmap":"रोडमैप","Save":"सहेजें","Save 20% (2 months free!)":"20% बचत (2 महीने मुफ्त!)","Save time and automate certain tasks with this cutting-edge feature":"इस उन्नत विशेषता के साथ समय बचाएं और कुछ कार्यों को स्वचालित करें","Save time and streamline your workflow by importing data from other sources":"अन्य स्रोतों से डेटा आयात करके अपनी व्यवस्था को समय बचाएं और स्ट्रीमलाइन करें","Select your plan!":"अपना योजना चुनें!","Send Feedback":"प्रतिक्रिया भेजें","Settings":"सेटिंग","Shape":"आकृति","Sign In":"साइन इन करें","Sign Up":"साइन अप","Sign in with <0>GitHub</0>":"<0>GitHub</0> के साथ साइन इन करें","Sign in with <0>Google</0>":"<0>Google</0> के साथ साइन इन करें","Size":"आकार","Small":"छोटा","Some classes are available to help style your flowchart":"कुछ क्लास आपकी फ्लोचार्ट स्टाइल करने में मदद करने के लिए उपलब्ध हैं","Sorry! This page is only available in English.":"माफ़ करें! यह पेज केवल अंग्रेजी में उपलब्ध है।","Source Arrow":"स्रोत तीर ","Source Column":"स्रोत स्तंभ","Source Delimiter":"स्रोत डिलिमिटर","Source/Target Arrow Shape":"स्रोत / लक्ष्य तीर आकार","Spacing":"अंतर","Special Attributes":"विशेष गुण","Start":"शुरू","Start Over":"शुरू करें","Status":"स्टेटस","Store any data associated to a node":"किसी नोड से संबंधित किसी भी डेटा स्टोर करें","Stored in the cloud":"बादल में संग्रहित","Stored on this computer":"इस कंप्यूटर पर संग्रहीत","Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":"फ्लोचार्ट फन के साथ अपनी व्यवस्था को स्ट्रीमलाइन करें और प्रक्रिया विजुअलाइजेशन को सरल बनाएं। अब सदस्यता लें और हमारे शक्तिशाली प्रो विशेषताओं का लाभ उठाएं!","Style":"शैली","Style Classes":"शैली क्लासेज","Submit":"भेजना","Subscribe":"सदस्यता लें","Subscription":"सदस्यता","Subscription will end":"सदस्यता समाप्त हो जाएगी","Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":"आपूर्ति श्रृंखला विश्लेषण और अनुकूलन: लागत कमी, कार्यक्षमता में सुधार और स्टेकहोल्डर सहयोग।","Target Arrow":"लक्ष्य तीर","Target Column":"लक्ष्य कॉलम","Target Delimiter":"लक्ष्य डेलीमीटर","Temporary":"अस्थायी","Temporary Flowchart Warning":"अस्थायी पारिस्थितिकी चेतावनी","Temporary Flowcharts":"अस्थायी आरेख","Text Sizes":"पाठ आकार","Text followed by colon+space creates an edge with the text as the label":"स्क्लीन के बाद कोलन + स्पेस एक किरदार बनाता है जिसका पाठ लेबल है","Text on a line creates a node with the text as the label":"एक पंक्ति पर पाठ एक नोड बनाता है जिसका पाठ लेबल है","Text sizes include small, regular, large, and extra-large.":"पाठ के आकार छोटे, सामान्य, बड़े और अतिरिक्त बड़े शामिल हैं।","Thank you for your feedback!":"आपके फ़ीडबैक के लिए धन्यवाद!","The column that contains the edge label(s)":"कॉलम जो एड्ज लेबल (ओं) को शामिल करता है","The column that contains the source node ID(s)":"कॉलम जो स्रोत नोड आईडी (ओं) को शामिल करता है","The column that contains the target node ID(s)":"कॉलम जो लक्ष्य नोड आईडी (ओं) को शामिल करता है","The delimiter used to separate multiple source nodes":"कई स्रोत नोड्स को अलग करने के लिए उपयोग किया गया डिलिमिटर","The delimiter used to separate multiple target nodes":"कई लक्ष्य नोड्स को अलग करने के लिए उपयोग किया गया डिलिमिटर","The possible shapes are:":"संभव आकृतियाँ हैं:","Theme Editor":"थीम एडिटर","There are no edges in this data":"इस डेटा में कोई एड्ज नहीं है","This action cannot be undone.":"इस क्रिया को पूर्ववत नहीं किया जा सकता।","This app works by typing":"यह ऐप टाइप द्वारा काम करता है","This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":"यह दस्तावेज़ केवल इस कंप्यूटर पर सहेजी गई है। यदि आप दस्तावेज़ सहेजने और कहीं भी उपयोग करने के लिए एक तरीका चाहते हैं तो हमारे प्रो प्लान के लिए केवल $ 3 / महीने में उपयोग करें!","This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":"यह सुविधा केवल प्रो उपयोगकर्ताओं के लिए उपलब्ध है। <0>प्रो उपयोगकर्ता बनें</0> इसे अनलॉक करने के लिए।","This is a temporary chart, it will be deleted when you clear your browser cache.":"यह एक अस्थायी चार्ट है, जब आप अपने ब्राउज़र कैश को साफ़ करेंगे तो यह हटा दिया जाएगा।","Tip":"सुझाव","To fix this change one of the edge IDs":"इसे ठीक करने के लिए एड्ज आईडी के एक को बदलें","To fix this change one of the node IDs":"इसे ठीक करने के लिए नोड आईडी के एक को बदलें","To fix this move one pointer to the next line":"इसे ठीक करने के लिए प्रतीक को अगली पंक्ति पर ले जाएं","To fix this move the pointer to the next line":"इसे ठीक करने के लिए प्रतीक को अगली पंक्ति पर ले जाएं","To fix this start the container <0/> on a different line":"इसे ठीक करने के लिए कंटेनर <0/> को एक अलग पंक्ति पर शुरू करें","To learn more about why we require you to log in, please read <0>this blog post</0>.":"हमें आपको लॉगिन करने के लिए आवश्यक क्यों है, कृपया <0>यह ब्लॉग पोस्ट</0> पढ़ें।","Top to Bottom":"ऊपर से नीचे","Try again":"फिर से कोशिश करें","Two edges have the same ID":"दो किनारे उसी आईडी के हैं","Two nodes have the same ID":"दो नोड उसी आईडी के हैं","Type":"प्रकार","Unfreeze":"अनफ़्रीज़","Unique text value to identify a node":"एक नोड को पहचानने के लिए अद्वितीय पाठ मूल्य","Unknown Parsing Error":"अज्ञात पार्सिंग त्रुटि","Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":"अपनी क्रेयता को उतारें और फ्लोचार्ट फन प्रो के साथ अपनी व्यवस्था को स्ट्रीमलाइन करें – सिर्फ $ 3 / महीने से शुरू होते हुए!","Update Email":"ईमेल अपडेट करें","Upgrade to Pro":"प्रो में अपग्रेड करें","Upload your File":"अपनी फाइल अपलोड करें","Use AI to create flowcharts from data sets or algorithms":"डेटा सेट्स या एल्गोरिदम्स से फ्लोचार्ट बनाने के लिए एआई का उपयोग करें","Use AI to generate a flowchart from a prompt. <0>Learn More</0>":"एक प्रॉम्प्ट से एक फ्लोचार्ट उत्पन्न करने के लिए AI का उपयोग करें। <0>और अधिक जानें</0>","Use classes to group nodes":"नोड्स को समूहों में सम्मिलित करने के लिए वर्ग उपयोग करें","Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":"नोड पर एक लिंक सेट करने के लिए गुण <0>href</0> उपयोग करें जो एक नये टैब में खुलेगा।","Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":"नोड की छवि सेट करने के लिए <0>src</0> गुण का उपयोग करें। छवि नोड के आकार में स्केल की जाएगी, इसलिए आपको आवश्यक प्रतिस्थापित करने के लिए नोड की चौड़ाई और ऊँचाई को समायोजित करना होगा। केवल सार्वजनिक छवियाँ (CORS द्वारा ब्लॉक नहीं की गई) समर्थित हैं।","Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":"नोड की विशिष्ट चौड़ाई और ऊँचाई सेट करने के लिए <0>w</0> और <1>h</1> गुण का उपयोग करें।","Use the customer portal to change your billing information.":"अपनी बिलिंग जानकारी बदलने के लिए ग्राहक पोर्टल का उपयोग करें।","Use this file for org charts, hierarchies, and other organizational structures.":"संगठनात्मक चार्ट, पदानुक्रमों और अन्य संगठनात्मक संरचनाओं के लिए इस फ़ाइल का उपयोग करें।","Use this file for sequences, processes, and workflows.":"अनुक्रमों, प्रक्रियाओं और कार्यप्रवाहों के लिए इस फ़ाइल का उपयोग करें।","User":"यूज़र","View on Github":"Github पर देखें","Water evaporates from the Earth\'s surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":"पानी पृथ्वी के सतह से उष्णता होकर आत्मान में उठता है और वर्षा के रूप में फिर नीचे गिरता है। इस पानी को तब नदियों, तिब्बती तालों और महासागरों में जाता है, जहां यह फिर से उष्णता होता है और आत्मान में पुनर्स्थापित किया जाता है।","Welcome to Flowchart Fun Pro!":"फ्लोचार्ट फन प्रो के लिए स्वागत है!","Width and Height":"चौड़ाई और ऊँचाई","Would you like to continue?":"क्या आप जारी रखना चाहते हैं?","Yearly":"वार्षिक","You already have a flowchart with this name.":"आपके पास इस नाम का आरेख पहले से है।","You are about to add {numNodes} nodes and {numEdges} edges to your graph.":["आप अपने ग्राफ में ",["numNodes"]," नोड्स और ",["numEdges"]," एड्ज़ जोड़ने वाले हैं।"],"You can also wrap text in \\"\\\\( \\\\)\\" to create an edge":"किनारे बनाने के लिए आप टेक्स्ट को \\"\\\\( \\\\)\\" में भी रैप कर सकते हैं","You must log in to create a persistent flowchart.":"आपको एक अविरल फ्लोचार्ट बनाने के लिए लॉग इन करना होगा।","You need to log in to access this page.":"इस पृष्ठ तक पहुँचने के लिए आपको लॉग इन करना होगा।","You\'re almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":"आप बस एक ही कदम दूर हैं! <0>Flowchart Fun Pro</0> की पूरी प्रत्याशित क्षमता को अनलॉक करने के लिए नीचे अपने भुगतान विवरण दर्ज करें और आज ही अद्भुत फ्लोचार्ट बनाने शुरू करें।","Your Charts":"आपके चार्ट","Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":"आपका सदस्यता अब निष्क्रिय है। यदि आप स्थायी चार्ट बनाना और संपादित करना चाहते हैं, तो प्रो अपग्रेड करें।","Zoom In":"ज़ूम इन करें","Zoom Out":"आउट ज़ूम करें","any text: before a colon creates a label":"कोई भी टेक्स्ट: कॉलन से पहले एक लेबल बनाता है","or":"या"}'
+  ),
 };
