@@ -1,449 +1,5 @@
 /*eslint-disable*/ module.exports = {
-  messages: {
-    "$3 per month": "월 3 달러",
-    "$30 per year": "연 30 달러",
-    "(before a colon creates a label)": "(레이블을 생성합니다)",
-    "<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":
-      "<0>Flowchart Fun</0>은 <1>Tone Row</1>가 만든 오픈 소스 프로젝트입니다.",
-    "A new version of the app is available. Please reload to update.":
-      "새로운 버전의 앱이 사용 가능합니다. 업데이트하려면 다시로드하십시오.",
-    About: "소개",
-    "Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":
-      "어디서든 이 차트에 액세스하세요.<0/>동기화되어 유지되는 플로우차트를 공유하고 삽입하세요.",
-    "Accessible from any device": "모든 장치에서 액세스 가능",
-    Account: "계정",
-    "Activate your account": "계정을 활성화하세요",
-    Advanced: "고급의",
-    Amount: "금액",
-    "An error occurred. Try resubmitting or email {0} directly.": [
-      "오류가 발생하였습니다. 다시 제출하거나 ",
-      ["0"],
-      "으로 직접 이메일을 보내주십시오.",
-    ],
-    Appearance: "외관",
-    "Apply Style": "스타일 적용",
-    Attributes: "속성",
-    "Auto-Layouts": "자동 레이아웃",
-    "Back To Editor": "편집기로 돌아가기",
-    "Basic Flowchart": "기본 플로우 차트",
-    "Become a Github Sponsor": "깃허브 스폰서가 되기",
-    "Become a Pro User": "프로 사용자가 되기",
-    "Begin with a simple example showing how <0>Flowchart Fun</0> works.":
-      "<0>Flowchart Fun</0>이 어떻게 작동하는지를 보여주는 간단한 예제로 시작하세요.",
-    Blank: "빈칸",
-    Blog: "블로그",
-    "Book a Meeting": "미팅 예약",
-    Border: "경계",
-    "Bottom to Top": "아래에서 위로",
-    Breadthfirst: "폭 우선",
-    Cancel: "취소",
-    "Cancel your subscription. Your hosted charts will become read-only.":
-      "구독을 취소하십시오. 귀하의 호스트 차트가 읽기 전용이 됩니다.",
-    "Certain attributes can be used to customize the appearance or functionality of elements.":
-      "일부 속성은 요소의 모양 또는 기능을 사용자 정의하기 위해 사용할 수 있습니다.",
-    "Change Email Address": "이메일 주소 변경",
-    Changelog: "변경 로그",
-    Charts: "차트",
-    "Check your email for a link to log in.<0/>You can close this window.":
-      "로그인할 링크가 들어있는 이메일을 확인하세요. 이 창은 닫을 수 있습니다.",
-    "Choose a Theme": "테마를 선택하세요",
-    "Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":
-      "에지의 소스와 목적지를 위한 다양한 화살표 모양을 선택합니다. 모양에는 삼각형, 삼각형-티, 원-삼각형, 삼각형-가위형, 삼각형-뒤곡선, 베이, 티, 사각형, 원, 다이아몬드, 쉐브론, 없음이 포함됩니다.",
-    "Choose the plan that's right for you and start creating amazing flowcharts with Flowchart Fun Pro":
-      "당신에게 맞는 계획을 선택하고 Flowchart Fun Pro을 사용하여 놀라운 플로우차트를 만들어보세요.",
-    Circle: "원",
-    Classes: "클래스",
-    Clay: "점토",
-    Clear: "지우다",
-    "Clear text?": "텍스트를 지우시겠습니까?",
-    "Click on \\nLearn Syntax\\nto learn more":
-      "\\n문법 습득\\n을 클릭하여 더 배우십시오",
-    "Click to preview": "미리보기를 클릭하세요",
-    Clone: "클론",
-    Close: "닫기",
-    Color: "색상",
-    "Colors include red, orange, yellow, blue, purple, black, white, and gray.":
-      "색상에는 빨강, 주황, 노랑, 파랑, 보라, 검정, 흰색, 회색이 포함됩니다.",
-    Column: "열",
-    "Comic Book": "만화책",
-    Comment: "댓글 달기",
-    Concentric: "동심",
-    "Confirm New Email": "새 이메일 확인",
-    Containers: "컨테이너",
-    "Containers are nodes that contain other nodes. They are declared using curly braces.":
-      "컨테이너는 다른 노드를 포함하는 노드입니다. 중괄호를 사용하여 선언됩니다.",
-    Continue: "계속하기",
-    "Convert this to a hosted chart?": "이것을 호스팅 된 차트로 변환할까요?",
-    "Convert to hosted chart?": "호스팅 차트로 변환하시겠습니까?",
-    Copy: "복사",
-    "Copy PNG Image": "PNG 이미지 복사",
-    "Copy SVG Code": "SVG 코드 복사",
-    "Copy your mermaid.js code or open it directly in the mermaid.js live editor.":
-      "mermaid.js 코드를 복사하거나 mermaid.js 라이브 편집기에서 직접 열어주세요.",
-    "Create Flowcharts from a Prompt with AI":
-      "AI 프롬프트를 사용하여 다이어그램 생성",
-    "Create New Flowchart": "새로운 플로우차트 만들기",
-    "Create a New Flowchart": "새 다이어그램 만들기",
-    "Create professional-quality visual aids for presentations or training materials":
-      "프레젠테이션 또는 교육 자료를 위한 전문 품질의 시각적 지원 작성",
-    "Creating an edge between two nodes is done by indenting the second node below the first":
-      "두 노드 사이에 엣지를 만드는 것은 첫 번째 노드 아래에 두 번째 노드를 들여 쓰는 것으로 합니다.",
-    "Custom Sharing Options": "커스텀 공유 옵션",
-    "Customer Portal": "고객 포털",
-    "Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":
-      "아래의 <0>Cytoscape CSS</0>를 편집하여 테마를 사용자 정의합니다. 스타일 문서는 곧 준비될 예정입니다! 일단은 최고의 자료는 <1>Cytoscape</1> 문서입니다. 막히면 <2>Discord</2>에 문의하세요.",
-    Dark: "다크",
-    "Dark Mode": "다크 모드",
-    Date: "날짜",
-    "Deleted when browser data is cleared": "브라우저 데이터가 지워지면 삭제됨",
-    "Describe the flowchart you wish to create": "만들려는 다이어그램 설명",
-    Direction: "방향",
-    Discard: "버리다",
-    "Do you want to cancel your subscription?": "구독을 취소하시겠습니까?",
-    "Do you want to delete this?": "본 항목을 삭제하시겠습니까?",
-    Document: "문서",
-    "Don't have an account?": "계정이 없습니까?",
-    Download: "다운로드",
-    "Download JPG": "JPG 다운로드",
-    "Download PNG": "PNG 다운로드",
-    "Download SVG": "SVG 다운로드",
-    "Drag and drop a CSV file here, or click to select a file":
-      "CSV 파일을 여기에 드래그 앤 드롭하거나 파일을 선택하려면 클릭하세요.",
-    "Draw an edge from multiple nodes by beginning the line with a reference":
-      "참조를 시작하여 여러 노드로부터 엣지를 그립니다.",
-    "Drop the file here ...": "파일을 여기에 드롭하세요 ...",
-    "Easily import existing flowcharts from other software":
-      "다른 소프트웨어에서 기존 플로우 차트를 쉽게 가져오기",
-    Edge: "가장자리",
-    "Edge ID, Classes, Attributes": "가장자리 ID, 클래스, 속성",
-    "Edge Label": "가장자리 라벨",
-    "Edge Label Column": "가장자리 라벨 열",
-    "Edge Style": "가장자리 스타일",
-    "Edge missing indentation": "들여쓰기가 누락된 가장자리",
-    Edges: "가장자리",
-    "Edges are declared in the same row as their source node":
-      "가장자리는 소스 노드가 있는 같은 행에 선언됩니다.",
-    "Edges are declared in the same row as their target node":
-      "가장자리는 목표 노드가 있는 같은 행에 선언됩니다.",
-    "Edges are declared in their own row": "가장자리는 자신의 행에 선언됩니다.",
-    "Edges can also have ID's, classes, and attributes before the label":
-      "라벨 앞에 ID, 클래스 및 속성도 가질 수 있습니다.",
-    "Edges can be styled with dashed, dotted, or solid lines":
-      "가장자리는 점선, 점선 또는 실선으로 스타일이 지정될 수 있습니다.",
-    "Edges in Separate Rows": "각 행별로 간선",
-    "Edges in Source Node Row": "출발 노드 행에 간선",
-    "Edges in Target Node Row": "도착 노드 행에 간선",
-    Editable: "편집 가능",
-    Editor: "에디터",
-    Eggs: "달걀",
-    Email: "이메일",
-    Empty: "비어 있음",
-    "Enter your email": "이메일을 입력하세요",
-    "Equal To": "같음",
-    "Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":
-      "에세이 작성 과정 플로차트는 학생들을 초기 아이디어 생성, 개념 정리, 작성, 수정 단계를 거쳐 이끌어줍니다.",
-    "Everything in Free": "무료 플랜의 모든 것",
-    Excalidraw: "Excalidraw",
-    Export: "내보내기",
-    "Export High-Resolution Images": "고해상도 이미지 내보내기",
-    "Export to SVG": "SVG로 내보내기",
-    "Extra Large": "엄청나게 큼",
-    Extract: "발췌",
-    Feedback: "피드백",
-    "Follow Us on Twitter": "트위터에서 우리를 팔로우하기",
-    "Found a bug? Have a feature request? We would love to hear from you!":
-      "버그를 발견했습니까? 기능 요청이 있습니까? 우리는 당신과의 소통을 사랑합니다!",
-    Free: "무료",
-    Fullscreen: "전체 화면",
-    Futuristic: "미래 지향적",
-    "Get More Layouts": "더 많은 레이아웃을 보고 싶어요",
-    "Get personalized attention and support from our expert team":
-      "고급 지식과 지원을 받으세요",
-    "Getting Started": "시작하기",
-    Graph: "그래프",
-    Grid: "그리드",
-    "Have complex questions or issues? We're here to help.":
-      "복잡한 문제가 있나요? 여기에서 도와드리겠습니다.",
-    Help: "도움말",
-    History: "기록",
-    Home: "집",
-    "How are edges declared in this data?":
-      "이 데이터에서 간선은 어떻게 선언됩니까?",
-    "ID's": "식별자",
-    "If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":
-      "<0>Flowchart Fun</0>을 사용하는 것을 즐기고 있다면, 프로젝트를 지원해 주시기 바랍니다.",
-    "If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\:</0>":
-      "만약 가장자리를 만들려면, 이 줄을 들여쓰기하십시오. 만약 그렇지 않다면, 콜론을 백슬래시로 이스케이프하십시오 <0>\\:</0>",
-    "Image Export": "이미지 내보내기",
-    Images: "이미지",
-    "Import Data": "데이터 가져오기",
-    "Import data from a CSV file.": "CSV 파일에서 데이터를 가져옵니다.",
-    "Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":
-      "CSV 파일에서 데이터를 가져와 새로운 플로우 차트로 매핑합니다. 이것은 Lucidchart, Google Sheets, Visio 등의 다른 소스에서 데이터를 가져오는 데 좋은 방법입니다.",
-    "Import from Lucidchart, Visio, and other CSV files":
-      "Lucidchart, Visio 및 기타 CSV 파일에서 가져오기",
-    "Import your diagram it into Microsoft Visio using one of these CSV files.":
-      "이러한 CSV 파일 중 하나를 사용하여 다이어그램을 Microsoft Visio로 가져옵니다.",
-    "Important Changes are Coming to Flowchart Fun.":
-      "Flowchart Fun에는 중요한 변화가 오고 있습니다.",
-    "Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":
-      "데이터 가져오기는 프로 기능입니다. Flowchart Fun Pro을 월 3 달러 또는 연 30 달러로 업그레이드 할 수 있습니다.",
-    "Impress your audience with high-quality images produced by Flowchart Fun":
-      "Flowchart Fun으로 만든 고품질 이미지로 관객을 감탄시켜보세요",
-    "Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":
-      "<0>title</0> 속성을 사용하여 제목을 포함시키십시오. Visio 색상을 사용하려면 다음 중 하나와 같은 <1>roleType</1> 속성을 추가하십시오:",
-    "Indenting creates an edge": "들여 쓰기는 가장자리를 만듭니다",
-    Info: "정보",
-    Instruct: "지시하다",
-    Is: "있다",
-    Language: "언어",
-    Large: "크다",
-    Layout: "레이아웃",
-    "Layout Frozen": "레이아웃 동결",
-    "Layout is Frozen": "레이아웃이 고정됨",
-    "Leading References": "주목할만한 참고",
-    "Learn More": "더 알아보기",
-    "Learn Syntax": "구문 배우기",
-    "Learn about Flowchart Fun Pro": "Flowchart Fun Pro에 대해 알아보기",
-    "Left to Right": "왼쪽에서 오른쪽으로",
-    "Let's get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":
-      "시작해볼까요? 아래 이메일 주소를 입력하여 Flowchart Fun 계정을 만들고 강력한 기능을 사용해보세요.",
-    Light: "라이트",
-    "Light Mode": "라이트 모드",
-    Link: "링크",
-    "Load Theme": "테마 불러오기",
-    "Loading a theme will erase any custom styles you have applied":
-      "테마를 불러오면 사용자 정의 스타일이 모두 사라집니다.",
-    "Lock Zoom to Graph": "그래프에 룩 줌을 고정하다",
-    "Log In": "로그인",
-    "Log Out": "로그아웃",
-    "Log in to start creating flowcharts.":
-      "플로우 차트를 만들기 위해 로그인하세요.",
-    "Make a One-Time Donation": "한 번 선물하기",
-    "Make publicly accessible": "공개적으로 액세스 가능하게 만들기",
-    "Make sure you use the same email you will use to log in.":
-      "로그인할 때 사용할 같은 이메일을 사용하세요.",
-    "Map Data": "데이터 지도",
-    "Market understanding and competitive landscape maintenance for SaaS product development":
-      "SaaS 제품 개발에 대한 시장 이해 및 경쟁 경제동화 유지",
-    Medium: "중간",
-    Method: "방법",
-    Monospace: "Monospace",
-    Monthly: "월간",
-    "Multiple pointers on same line": "같은 줄에 여러 포인터",
-    Museum: "박물관",
-    Name: "이름",
-    New: "신규",
-    "New Email": "새 이메일",
-    "Next charge": "다음 청구 금액",
-    "No Edges": "노드 없음",
-    Node: "노드",
-    "Node Border Style": "노드 경계 스타일",
-    "Node Colors": "노드 색상",
-    "Node ID": "노드 ID",
-    "Node ID, Classes, Attributes": "노드 ID, 클래스, 속성",
-    "Node Label": "노드 라벨",
-    "Node Shapes": "노드 모양",
-    "Node and pointer on same line": "같은 줄에 노드와 포인터",
-    "Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":
-      "노드는 점선, 점점선 또는 이중선으로 스타일링할 수 있으며, border-none으로 테두리를 제거할 수 있습니다.",
-    "Not Empty": "비어 있지 않음",
-    "Office Hours": "근무 시간",
-    "Once in a while the magic link will end up in your spam folder. If you don't see it after a few minutes, check there or request a new link.":
-      "가끔 마술 링크가 스팸 폴더에 도착할 수도 있습니다. 몇 분 후에도 보이지 않으면 거기를 확인하거나 새로운 링크를 요청하십시오.",
-    "One-on-One Support": "일대일 지원",
-    "One-time Share Links": "일회성 공유 링크",
-    "Only available on this device.<0/>Clearing your browser cache will erase them.":
-      "이 기기에서만 사용 가능합니다.<0/>브라우저 캐시를 지우면 삭제됩니다.",
-    "Open Customer Portal": "고객 포털 열기",
-    "Organization Chart": "조직도",
-    "Paste the information you wish to convert to a flowchart":
-      "순서도로 변환하려는 정보를 붙여넣기",
-    Permalinks: "퍼머링크",
-    "Permanent Charts are a Pro Feature": "영구 차트는 프로 기능입니다",
-    "Permanent Flowcharts": "영구 플로우차트",
-    Persistent: "끈기 있는",
-    "Persistent Flowcharts": "영구 다이어그램",
-    Playbook: "플레이북",
-    "Pointer and container on same line": "같은 줄에 포인터와 컨테이너",
-    Pricing: "가격",
-    "Process for corporate social responsibility initiatives development and implementation across company operations":
-      "기업 사회적 책임 사업 개발 및 회사 운영 전반에 걸쳐 구현 프로세스",
-    "Processing Data": "데이터 처리",
-    Prompt: "프롬프트",
-    Public: "공용",
-    Random: "무작위",
-    "Read-only": "읽기 전용",
-    "Reference by Class": "클래스로 참조",
-    "Reference by ID": "ID로 참조",
-    "Reference by Label": "레이블로 참조",
-    References: "참조",
-    "References are used to create edges between nodes that are created elsewhere in the document":
-      "참조는 문서 내부에서 만들어진 노드 사이에 간선을 만들기 위해 사용됩니다.",
-    "Referencing a node by its exact label": "정확한 레이블로 노드를 참조하기",
-    "Referencing a node by its unique ID": "고유한 ID로 노드를 참조하기",
-    "Referencing multiple nodes with the same assigned class":
-      "동일한 할당 된 클래스로 여러 노드를 참조하기 ",
-    "Refresh Page": "페이지 새로 고침 ",
-    "Reload to Update": "업데이트하려면 다시 로드하기 ",
-    "Remove Image Watermark": "이미지 워터마크 제거",
-    Rename: "이름 바꾸기",
-    Reset: "재설정",
-    "Resume Subscription": "구독 재개",
-    Retro: "레트로",
-    Return: "반품",
-    "Right to Left": "오른쪽에서 왼쪽으로",
-    Roadmap: "로드맵",
-    Save: "구하다",
-    "Save 20% (2 months free!)": "20% 할인 (2개월 무료!)",
-    "Save time and automate certain tasks with this cutting-edge feature":
-      "이 최신 기능으로 시간을 절약하고 일부 작업을 자동화하세요.",
-    "Save time and streamline your workflow by importing data from other sources":
-      "다른 소스에서 데이터를 가져와 작업흐름을 최적화하여 시간을 절약하세요.",
-    "Select your plan!": "귀하의 계획을 선택하세요!",
-    "Send Feedback": "피드백 보내기",
-    Settings: "설정",
-    Shape: "모양 ",
-    "Sign Up": "가입",
-    "Sign Up Now": "지금 가입하세요",
-    Size: "크기",
-    Small: "작은",
-    "Some classes are available to help style your flowchart":
-      "스타일링하기 위한 몇 가지 클래스가 이용 가능합니다",
-    "Sorry! This page is only available in English.":
-      "죄송합니다! 이 페이지는 영어로만 제공됩니다.",
-    "Source Arrow": "소스 화살표",
-    "Source Column": "소스 열",
-    "Source Delimiter": "소스 구분자",
-    "Source/Target Arrow Shape": "소스/대상 화살표 모양",
-    Spacing: "간격",
-    "Special Attributes": "특별한 속성",
-    Start: "시작",
-    "Start Over": "처음부터 다시 시작하기",
-    Status: "상태",
-    "Store any data associated to a node": "노드에 관련된 모든 데이터 저장하기",
-    "Stored in the cloud": "클라우드에 저장",
-    "Stored on this computer": "이 컴퓨터에 저장됨",
-    "Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":
-      "Flowchart Fun을 사용하여 작업흐름을 최적화하고 프로 기능을 이용하여 프로세스 시각화를 단순화하세요. 지금 구독하고 우리의 강력한 프로 기능을 이용하세요!",
-    Style: "스타일",
-    "Style Classes": "스타일 클래스",
-    Submit: "보내다",
-    Subscribe: "구독",
-    Subscription: "구독",
-    "Subscription will end": "구독이 종료될 예정입니다.",
-    "Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":
-      "공급망 분석 및 최적화: 비용 감소, 효율 향상 및 이해관계자 협력",
-    "Target Arrow": "타겟 화살표",
-    "Target Column": "대상 열",
-    "Target Delimiter": "대상 구분자",
-    Temporary: "임시",
-    "Temporary Flowchart Warning": "임시 흐름도 경고",
-    "Temporary Flowcharts": "임시 다이어그램",
-    "Text Sizes": "텍스트 크기",
-    "Text followed by colon+space creates an edge with the text as the label":
-      "콜론 뒤에 공백이 따라오는 텍스트는 텍스트를 레이블로 하는 엣지를 만듭니다.",
-    "Text on a line creates a node with the text as the label":
-      "한 줄에 있는 텍스트는 텍스트를 레이블로 하는 노드를 만듭니다.",
-    "Text sizes include small, regular, large, and extra-large.":
-      "텍스트 사이즈는 작게, 일반, 크게, 그리고 엑스트라 크게가 있습니다.",
-    "Thank you for your feedback!": "피드백을 해주셔서 감사합니다!",
-    "The column that contains the edge label(s)": "엣지 레이블이 포함된 열",
-    "The column that contains the source node ID(s)":
-      "소스 노드 ID가 포함된 열",
-    "The column that contains the target node ID(s)":
-      "타겟 노드 ID가 포함된 열",
-    "The delimiter used to separate multiple source nodes":
-      "여러 개의 소스 노드를 구분하기 위해 사용되는 구분 기호",
-    "The delimiter used to separate multiple target nodes":
-      "여러 개의 목표 노드를 구분하기 위해 사용되는 구분 기호",
-    "The possible shapes are:": "가능한 모양은 다음과 같습니다:",
-    "Theme Editor": "테마 편집기",
-    "There are no edges in this data": "이 데이터에는 엣지가 없습니다.",
-    "This action cannot be undone.": "이 작업은 취소할 수 없습니다.",
-    "This app works by typing": "이 앱은 타이핑으로 작동합니다",
-    "This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":
-      "이 문서는 이 컴퓨터에만 저장됩니다. 문서를 저장하고 어디서든 접근할 수 있는 방법이 필요하면 저희 프로 플랜을 3달러에 체험해보세요!",
-    "This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":
-      "이 기능은 프로 사용자에게만 제공됩니다. <0>프로 사용자가 되어</0> 이 기능을 사용할 수 있게 하세요.",
-    "This is a temporary chart, it will be deleted when you clear your browser cache.":
-      "이것은 임시 차트로, 브라우저 캐시를 지우면 삭제됩니다.",
-    Tip: "팁",
-    "To fix this change one of the edge IDs":
-      "이를 수정하려면, 가장자리 ID 중 하나를 변경하십시오.",
-    "To fix this change one of the node IDs":
-      "이것을 수정하려면 노드 ID 중 하나를 변경하십시오",
-    "To fix this move one pointer to the next line":
-      "이것을 수정하려면 포인터를 다음 줄로 이동하십시오",
-    "To fix this move the pointer to the next line":
-      "이것을 수정하려면 포인터를 다음 줄로 이동하십시오",
-    "To fix this start the container <0/> on a different line":
-      "이것을 수정하려면 컨테이너 <0/>을 다른 줄에 시작하십시오",
-    "Top to Bottom": "위에서 아래로",
-    "Try again": "다시 시도하세요",
-    "Two edges have the same ID": "두 개의 간선이 같은 ID를 가지고 있습니다",
-    "Two nodes have the same ID": "두 개의 노드가 같은 ID를 가지고 있습니다",
-    Type: "유형",
-    Unfreeze: "녹이다",
-    "Unique text value to identify a node":
-      "노드를 식별하기 위한 고유한 텍스트 값",
-    "Unknown Parsing Error": "알 수 없는 구문 분석 오류",
-    "Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":
-      "Flowchart Fun Pro를 사용하여 창의력을 발휘하고 작업흐름을 최적화하세요. $3/월부터 시작합니다!",
-    "Update Email": "이메일 업데이트",
-    "Upgrade to Pro": "프로로 업그레이드",
-    "Upload your File": "파일을 업로드하세요.",
-    "Use AI to create flowcharts from data sets or algorithms":
-      "데이터 집합 또는 알고리즘에서 플로우차트를 만들기 위해 AI를 사용하세요.",
-    "Use AI to generate a flowchart from a prompt. <0>Learn More</0>":
-      "AI를 사용하여 프롬프트에세이에서 플로차트를 생성합니다. <0>자세히 알아보기</0>",
-    "Use classes to group nodes":
-      "노드를 그룹화하기 위해 클래스를 사용하십시오.",
-    "Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":
-      "새 탭에서 링크를 설정하기 위해 속성 <0>href</0>을 사용하십시오.",
-    "Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":
-      "노드의 이미지를 설정하려면 <0>src</0> 속성을 사용하세요. 이미지는 노드에 맞게 크기가 조정되므로 노드의 너비와 높이를 조절해야 할 수도 있습니다. CORS로 차단되지 않은 공개 이미지만 지원됩니다.",
-    "Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":
-      "노드의 너비와 높이를 명시적으로 설정하려면 <0>w</0> 및 <1>h</1> 속성을 사용하세요.",
-    "Use the customer portal to change your billing information.":
-      "청구 정보를 변경하려면 고객 포털을 사용하십시오.",
-    "Use this file for org charts, hierarchies, and other organizational structures.":
-      "조직도, 계층 구조 및 기타 조직 구조를 위해 이 파일을 사용하십시오.",
-    "Use this file for sequences, processes, and workflows.":
-      "시퀀스, 프로세스 및 워크플로우에 대해 이 파일을 사용하십시오.",
-    User: "사용자",
-    "View on Github": "Github에서 보기",
-    "Water evaporates from the Earth's surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":
-      "물은 지구의 표면에서 증발하여 공기로 올라가 비로 내려오고, 그 물은 강과 호수, 바다로 흐르고 다시 증발하여 공기로 돌아가 재생됩니다.",
-    "We use magic links to log you in. Enter your email below to get started.":
-      "우리는 마법 링크를 사용하여 로그인합니다. 아래에 이메일을 입력하여 시작하십시오.",
-    "Welcome to Flowchart Fun Pro!":
-      "Flowchart Fun Pro에 오신 것을 환영합니다!",
-    "Width and Height": "너비와 높이",
-    "Would you like to continue?": "계속하시겠습니까?",
-    Yearly: "연간",
-    "You already have a flowchart with this name.":
-      "이 이름을 가진 다이어그램이 이미 있습니다.",
-    "You are about to add {numNodes} nodes and {numEdges} edges to your graph.":
-      [
-        "그래프에 ",
-        ["numNodes"],
-        "개의 노드와 ",
-        ["numEdges"],
-        "개의 간선을 추가하려고 합니다.",
-      ],
-    'You can also wrap text in "\\( \\)" to create an edge':
-      '텍스트를 "\\( \\)"로 감싸서 가장자리를 만들 수도 있습니다.',
-    "You must log in to create a persistent flowchart.":
-      "끈기 있는 플로우차트를 만들기 위해 로그인해야합니다.",
-    "You're almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":
-      "거의 다 됐습니다! <0>Flowchart Fun Pro</0>의 전력을 풀어내는 마지막 한 단계만 남았습니다. 아래에 결제 정보를 입력하여 구독을 완료하고 오늘부터 놀라운 플로우 차트를 만들 수 있습니다.",
-    "Your Charts": "당신의 차트",
-    "Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":
-      "귀하의 구독이 더 이상 활성화되지 않습니다. 영구 차트를 만들고 편집하려면 Pro로 업그레이드하십시오.",
-    "Zoom In": "확대",
-    "Zoom Out": "축소하기",
-    "any text: before a colon creates a label":
-      "콜론 앞의 모든 텍스트는: 레이블을 생성합니다",
-  },
+  messages: JSON.parse(
+    '{"$3 per month":"월 3 달러","$30 per year":"연 30 달러","(before a colon creates a label)":"(레이블을 생성합니다)","<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":"<0>Flowchart Fun</0>은 <1>Tone Row</1>가 만든 오픈 소스 프로젝트입니다.","<0>Sign In</0> / <1>Sign Up</1> with email and password":"<0>로그인</0> / <1>회원가입</1> 이메일과 비밀번호로","<0>Starting August 28th you will need a subscription to create and edit charts. <1>Learn more</1>.</0>":"<0>8월 28일부터 차트를 만들고 편집하기 위해 구독이 필요합니다. <1>자세히 알아보기</1>.</0>","<0>You currently have a free account.<1/><2>Learn about our Pro Features and subscribe on our pricing page</2>.</0>":"<0>현재 당신은 무료 계정을 가지고 있습니다. <1/><2>프로 기능에 대해 알아보고 우리 가격 페이지에서 구독하세요</2>.</0>","A new version of the app is available. Please reload to update.":"새로운 버전의 앱이 사용 가능합니다. 업데이트하려면 다시로드하십시오.","About":"소개","Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":"어디서든 이 차트에 액세스하세요.<0/>동기화되어 유지되는 플로우차트를 공유하고 삽입하세요.","Accessible from any device":"모든 장치에서 액세스 가능","Account":"계정","Activate your account":"계정을 활성화하세요","Advanced":"고급의","Amount":"금액","An error occurred. Try resubmitting or email {0} directly.":["오류가 발생하였습니다. 다시 제출하거나 ",["0"],"으로 직접 이메일을 보내주십시오."],"Appearance":"외관","Apply Style":"스타일 적용","Attributes":"속성","Auto-Layouts":"자동 레이아웃","Back To Editor":"편집기로 돌아가기","Basic Flowchart":"기본 플로우 차트","Become a Github Sponsor":"깃허브 스폰서가 되기","Become a Pro User":"프로 사용자가 되기","Begin with a simple example showing how <0>Flowchart Fun</0> works.":"<0>Flowchart Fun</0>이 어떻게 작동하는지를 보여주는 간단한 예제로 시작하세요.","Blank":"빈칸","Blog":"블로그","Book a Meeting":"미팅 예약","Border":"경계","Bottom to Top":"아래에서 위로","Breadthfirst":"폭 우선","Cancel":"취소","Cancel your subscription. Your hosted charts will become read-only.":"구독을 취소하십시오. 귀하의 호스트 차트가 읽기 전용이 됩니다.","Certain attributes can be used to customize the appearance or functionality of elements.":"일부 속성은 요소의 모양 또는 기능을 사용자 정의하기 위해 사용할 수 있습니다.","Change Email Address":"이메일 주소 변경","Changelog":"변경 로그","Charts":"차트","Check your email for a link to log in.<0/>You can close this window.":"로그인할 링크가 들어있는 이메일을 확인하세요. 이 창은 닫을 수 있습니다.","Choose":"선택하세요","Choose a Theme":"테마를 선택하세요","Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":"에지의 소스와 목적지를 위한 다양한 화살표 모양을 선택합니다. 모양에는 삼각형, 삼각형-티, 원-삼각형, 삼각형-가위형, 삼각형-뒤곡선, 베이, 티, 사각형, 원, 다이아몬드, 쉐브론, 없음이 포함됩니다.","Choose the plan that\'s right for you and start creating amazing flowcharts with Flowchart Fun Pro":"당신에게 맞는 계획을 선택하고 Flowchart Fun Pro을 사용하여 놀라운 플로우차트를 만들어보세요.","Circle":"원","Classes":"클래스","Clay":"점토","Clear":"지우다","Clear text?":"텍스트를 지우시겠습니까?","Click on \\\\nLearn Syntax\\\\nto learn more":"\\\\n문법 습득\\\\n을 클릭하여 더 배우십시오","Click to preview":"미리보기를 클릭하세요","Clone":"클론","Close":"닫기","Color":"색상","Colors include red, orange, yellow, blue, purple, black, white, and gray.":"색상에는 빨강, 주황, 노랑, 파랑, 보라, 검정, 흰색, 회색이 포함됩니다.","Column":"열","Comic Book":"만화책","Comment":"댓글 달기","Concentric":"동심","Confirm New Email":"새 이메일 확인","Confirm your email address to sign in.":"로그인하려면 이메일 주소를 확인하세요.","Containers":"컨테이너","Containers are nodes that contain other nodes. They are declared using curly braces.":"컨테이너는 다른 노드를 포함하는 노드입니다. 중괄호를 사용하여 선언됩니다.","Continue":"계속하기","Convert this to a hosted chart?":"이것을 호스팅 된 차트로 변환할까요?","Convert to hosted chart?":"호스팅 차트로 변환하시겠습니까?","Copy":"복사","Copy PNG Image":"PNG 이미지 복사","Copy SVG Code":"SVG 코드 복사","Copy your mermaid.js code or open it directly in the mermaid.js live editor.":"mermaid.js 코드를 복사하거나 mermaid.js 라이브 편집기에서 직접 열어주세요.","Create Flowcharts from a Prompt with AI":"AI 프롬프트를 사용하여 다이어그램 생성","Create New Flowchart":"새로운 플로우차트 만들기","Create a New Flowchart":"새 다이어그램 만들기","Create professional-quality visual aids for presentations or training materials":"프레젠테이션 또는 교육 자료를 위한 전문 품질의 시각적 지원 작성","Creating an edge between two nodes is done by indenting the second node below the first":"두 노드 사이에 엣지를 만드는 것은 첫 번째 노드 아래에 두 번째 노드를 들여 쓰는 것으로 합니다.","Custom Sharing Options":"커스텀 공유 옵션","Customer Portal":"고객 포털","Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":"아래의 <0>Cytoscape CSS</0>를 편집하여 테마를 사용자 정의합니다. 스타일 문서는 곧 준비될 예정입니다! 일단은 최고의 자료는 <1>Cytoscape</1> 문서입니다. 막히면 <2>Discord</2>에 문의하세요.","Dark":"다크","Dark Mode":"다크 모드","Date":"날짜","Deleted when browser data is cleared":"브라우저 데이터가 지워지면 삭제됨","Describe the flowchart you wish to create":"만들려는 다이어그램 설명","Direction":"방향","Discard":"버리다","Do you want to cancel your subscription?":"구독을 취소하시겠습니까?","Do you want to delete this?":"본 항목을 삭제하시겠습니까?","Document":"문서","Download":"다운로드","Download JPG":"JPG 다운로드","Download PNG":"PNG 다운로드","Download SVG":"SVG 다운로드","Drag and drop a CSV file here, or click to select a file":"CSV 파일을 여기에 드래그 앤 드롭하거나 파일을 선택하려면 클릭하세요.","Draw an edge from multiple nodes by beginning the line with a reference":"참조를 시작하여 여러 노드로부터 엣지를 그립니다.","Drop the file here ...":"파일을 여기에 드롭하세요 ...","Easily import existing flowcharts from other software":"다른 소프트웨어에서 기존 플로우 차트를 쉽게 가져오기","Edge":"가장자리","Edge ID, Classes, Attributes":"가장자리 ID, 클래스, 속성","Edge Label":"가장자리 라벨","Edge Label Column":"가장자리 라벨 열","Edge Style":"가장자리 스타일","Edge missing indentation":"들여쓰기가 누락된 가장자리","Edges":"가장자리","Edges are declared in the same row as their source node":"가장자리는 소스 노드가 있는 같은 행에 선언됩니다.","Edges are declared in the same row as their target node":"가장자리는 목표 노드가 있는 같은 행에 선언됩니다.","Edges are declared in their own row":"가장자리는 자신의 행에 선언됩니다.","Edges can also have ID\'s, classes, and attributes before the label":"라벨 앞에 ID, 클래스 및 속성도 가질 수 있습니다.","Edges can be styled with dashed, dotted, or solid lines":"가장자리는 점선, 점선 또는 실선으로 스타일이 지정될 수 있습니다.","Edges in Separate Rows":"각 행별로 간선","Edges in Source Node Row":"출발 노드 행에 간선","Edges in Target Node Row":"도착 노드 행에 간선","Editable":"편집 가능","Editor":"에디터","Eggs":"달걀","Email":"이메일","Empty":"비어 있음","Enter your email":"이메일을 입력하세요","Enter your email address and we\'ll send you a magic link to sign in.":"이메일 주소를 입력하면 자동으로 로그인할 수 있는 링크를 보내드립니다.","Enter your email address below and we\'ll send you a link to reset your password.":"아래에 이메일 주소를 입력하면 비밀번호 재설정을 위한 링크를 보내드립니다.","Equal To":"같음","Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":"에세이 작성 과정 플로차트는 학생들을 초기 아이디어 생성, 개념 정리, 작성, 수정 단계를 거쳐 이끌어줍니다.","Everything in Free":"무료 플랜의 모든 것","Excalidraw":"Excalidraw","Export":"내보내기","Export High-Resolution Images":"고해상도 이미지 내보내기","Export to SVG":"SVG로 내보내기","Extra Large":"엄청나게 큼","Extract":"발췌","Feedback":"피드백","Follow Us on Twitter":"트위터에서 우리를 팔로우하기","Forgot your password?":"비밀번호를 잊으셨나요?","Found a bug? Have a feature request? We would love to hear from you!":"버그를 발견했습니까? 기능 요청이 있습니까? 우리는 당신과의 소통을 사랑합니다!","Free":"무료","Fullscreen":"전체 화면","Futuristic":"미래 지향적","Get More Layouts":"더 많은 레이아웃을 보고 싶어요","Get personalized attention and support from our expert team":"고급 지식과 지원을 받으세요","Getting Started":"시작하기","Graph":"그래프","Grid":"그리드","Have complex questions or issues? We\'re here to help.":"복잡한 문제가 있나요? 여기에서 도와드리겠습니다.","Help":"도움말","History":"기록","Home":"집","How are edges declared in this data?":"이 데이터에서 간선은 어떻게 선언됩니까?","ID\'s":"식별자","If an account with that email exists, we\'ve sent you an email with instructions on how to reset your password.":"해당 이메일로 계정이 있다면 비밀번호 재설정을 위한 안내를 포함한 이메일을 보내드립니다.","If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":"<0>Flowchart Fun</0>을 사용하는 것을 즐기고 있다면, 프로젝트를 지원해 주시기 바랍니다.","If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\\\:</0>":"만약 가장자리를 만들려면, 이 줄을 들여쓰기하십시오. 만약 그렇지 않다면, 콜론을 백슬래시로 이스케이프하십시오 <0>\\\\:</0>","Image Export":"이미지 내보내기","Images":"이미지","Import Data":"데이터 가져오기","Import data from a CSV file.":"CSV 파일에서 데이터를 가져옵니다.","Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":"CSV 파일에서 데이터를 가져와 새로운 플로우 차트로 매핑합니다. 이것은 Lucidchart, Google Sheets, Visio 등의 다른 소스에서 데이터를 가져오는 데 좋은 방법입니다.","Import from Lucidchart, Visio, and other CSV files":"Lucidchart, Visio 및 기타 CSV 파일에서 가져오기","Import your diagram it into Microsoft Visio using one of these CSV files.":"이러한 CSV 파일 중 하나를 사용하여 다이어그램을 Microsoft Visio로 가져옵니다.","Important Changes are Coming to Flowchart Fun.":"Flowchart Fun에는 중요한 변화가 오고 있습니다.","Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":"데이터 가져오기는 프로 기능입니다. Flowchart Fun Pro을 월 3 달러 또는 연 30 달러로 업그레이드 할 수 있습니다.","Impress your audience with high-quality images produced by Flowchart Fun":"Flowchart Fun으로 만든 고품질 이미지로 관객을 감탄시켜보세요","Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":"<0>title</0> 속성을 사용하여 제목을 포함시키십시오. Visio 색상을 사용하려면 다음 중 하나와 같은 <1>roleType</1> 속성을 추가하십시오:","Indenting creates an edge":"들여 쓰기는 가장자리를 만듭니다","Info":"정보","Instruct":"지시하다","Is":"있다","Language":"언어","Large":"크다","Layout":"레이아웃","Layout Frozen":"레이아웃 동결","Layout is Frozen":"레이아웃이 고정됨","Leading References":"주목할만한 참고","Learn More":"더 알아보기","Learn Syntax":"구문 배우기","Learn about Flowchart Fun Pro":"Flowchart Fun Pro에 대해 알아보기","Left to Right":"왼쪽에서 오른쪽으로","Let\'s get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":"시작해볼까요? 아래 이메일 주소를 입력하여 Flowchart Fun 계정을 만들고 강력한 기능을 사용해보세요.","Light":"라이트","Light Mode":"라이트 모드","Link":"링크","Load Theme":"테마 불러오기","Loading a theme will erase any custom styles you have applied":"테마를 불러오면 사용자 정의 스타일이 모두 사라집니다.","Lock Zoom to Graph":"그래프에 룩 줌을 고정하다","Log In":"로그인","Log Out":"로그아웃","Log in to start creating flowcharts.":"플로우 차트를 만들기 위해 로그인하세요.","Make a One-Time Donation":"한 번 선물하기","Make publicly accessible":"공개적으로 액세스 가능하게 만들기","Make sure you use the same email you will use to log in.":"로그인할 때 사용할 같은 이메일을 사용하세요.","Map Data":"데이터 지도","Market understanding and competitive landscape maintenance for SaaS product development":"SaaS 제품 개발에 대한 시장 이해 및 경쟁 경제동화 유지","Medium":"중간","Method":"방법","Monospace":"Monospace","Monthly":"월간","Multiple pointers on same line":"같은 줄에 여러 포인터","Museum":"박물관","Name":"이름","New":"신규","New Email":"새 이메일","Next charge":"다음 청구 금액","No Edges":"노드 없음","Node":"노드","Node Border Style":"노드 경계 스타일","Node Colors":"노드 색상","Node ID":"노드 ID","Node ID, Classes, Attributes":"노드 ID, 클래스, 속성","Node Label":"노드 라벨","Node Shapes":"노드 모양","Node and pointer on same line":"같은 줄에 노드와 포인터","Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":"노드는 점선, 점점선 또는 이중선으로 스타일링할 수 있으며, border-none으로 테두리를 제거할 수 있습니다.","Not Empty":"비어 있지 않음","Office Hours":"근무 시간","Once in a while the magic link will end up in your spam folder. If you don\'t see it after a few minutes, check there or request a new link.":"가끔 마술 링크가 스팸 폴더에 도착할 수도 있습니다. 몇 분 후에도 보이지 않으면 거기를 확인하거나 새로운 링크를 요청하십시오.","One-on-One Support":"일대일 지원","One-time Share Links":"일회성 공유 링크","Only available on this device.<0/>Clearing your browser cache will erase them.":"이 기기에서만 사용 가능합니다.<0/>브라우저 캐시를 지우면 삭제됩니다.","Open Customer Portal":"고객 포털 열기","Organization Chart":"조직도","Password":"비밀번호","Paste the information you wish to convert to a flowchart":"순서도로 변환하려는 정보를 붙여넣기","Permalinks":"퍼머링크","Permanent Charts are a Pro Feature":"영구 차트는 프로 기능입니다","Permanent Flowcharts":"영구 플로우차트","Persistent":"끈기 있는","Persistent Flowcharts":"영구 다이어그램","Playbook":"플레이북","Pointer and container on same line":"같은 줄에 포인터와 컨테이너","Pricing":"가격","Process for corporate social responsibility initiatives development and implementation across company operations":"기업 사회적 책임 사업 개발 및 회사 운영 전반에 걸쳐 구현 프로세스","Processing Data":"데이터 처리","Prompt":"프롬프트","Public":"공용","Random":"무작위","Read-only":"읽기 전용","Reference by Class":"클래스로 참조","Reference by ID":"ID로 참조","Reference by Label":"레이블로 참조","References":"참조","References are used to create edges between nodes that are created elsewhere in the document":"참조는 문서 내부에서 만들어진 노드 사이에 간선을 만들기 위해 사용됩니다.","Referencing a node by its exact label":"정확한 레이블로 노드를 참조하기","Referencing a node by its unique ID":"고유한 ID로 노드를 참조하기","Referencing multiple nodes with the same assigned class":"동일한 할당 된 클래스로 여러 노드를 참조하기 ","Refresh Page":"페이지 새로 고침 ","Reload to Update":"업데이트하려면 다시 로드하기 ","Remove Image Watermark":"이미지 워터마크 제거","Rename":"이름 바꾸기","Request Magic Link":"마법 링크 요청","Request Password Reset":"비밀번호 재설정 요청","Reset":"재설정","Reset Password":"비밀번호 재설정","Resume Subscription":"구독 재개","Retro":"레트로","Return":"반품","Right to Left":"오른쪽에서 왼쪽으로","Roadmap":"로드맵","Save":"구하다","Save 20% (2 months free!)":"20% 할인 (2개월 무료!)","Save time and automate certain tasks with this cutting-edge feature":"이 최신 기능으로 시간을 절약하고 일부 작업을 자동화하세요.","Save time and streamline your workflow by importing data from other sources":"다른 소스에서 데이터를 가져와 작업흐름을 최적화하여 시간을 절약하세요.","Select your plan!":"귀하의 계획을 선택하세요!","Send Feedback":"피드백 보내기","Settings":"설정","Shape":"모양 ","Sign In":"로그인","Sign Up":"가입","Sign in with <0>GitHub</0>":"<0>GitHub</0>으로 로그인","Sign in with <0>Google</0>":"<0>Google</0>으로 로그인","Size":"크기","Small":"작은","Some classes are available to help style your flowchart":"스타일링하기 위한 몇 가지 클래스가 이용 가능합니다","Sorry! This page is only available in English.":"죄송합니다! 이 페이지는 영어로만 제공됩니다.","Source Arrow":"소스 화살표","Source Column":"소스 열","Source Delimiter":"소스 구분자","Source/Target Arrow Shape":"소스/대상 화살표 모양","Spacing":"간격","Special Attributes":"특별한 속성","Start":"시작","Start Over":"처음부터 다시 시작하기","Status":"상태","Store any data associated to a node":"노드에 관련된 모든 데이터 저장하기","Stored in the cloud":"클라우드에 저장","Stored on this computer":"이 컴퓨터에 저장됨","Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":"Flowchart Fun을 사용하여 작업흐름을 최적화하고 프로 기능을 이용하여 프로세스 시각화를 단순화하세요. 지금 구독하고 우리의 강력한 프로 기능을 이용하세요!","Style":"스타일","Style Classes":"스타일 클래스","Submit":"보내다","Subscribe":"구독","Subscription":"구독","Subscription will end":"구독이 종료될 예정입니다.","Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":"공급망 분석 및 최적화: 비용 감소, 효율 향상 및 이해관계자 협력","Target Arrow":"타겟 화살표","Target Column":"대상 열","Target Delimiter":"대상 구분자","Temporary":"임시","Temporary Flowchart Warning":"임시 흐름도 경고","Temporary Flowcharts":"임시 다이어그램","Text Sizes":"텍스트 크기","Text followed by colon+space creates an edge with the text as the label":"콜론 뒤에 공백이 따라오는 텍스트는 텍스트를 레이블로 하는 엣지를 만듭니다.","Text on a line creates a node with the text as the label":"한 줄에 있는 텍스트는 텍스트를 레이블로 하는 노드를 만듭니다.","Text sizes include small, regular, large, and extra-large.":"텍스트 사이즈는 작게, 일반, 크게, 그리고 엑스트라 크게가 있습니다.","Thank you for your feedback!":"피드백을 해주셔서 감사합니다!","The column that contains the edge label(s)":"엣지 레이블이 포함된 열","The column that contains the source node ID(s)":"소스 노드 ID가 포함된 열","The column that contains the target node ID(s)":"타겟 노드 ID가 포함된 열","The delimiter used to separate multiple source nodes":"여러 개의 소스 노드를 구분하기 위해 사용되는 구분 기호","The delimiter used to separate multiple target nodes":"여러 개의 목표 노드를 구분하기 위해 사용되는 구분 기호","The possible shapes are:":"가능한 모양은 다음과 같습니다:","Theme Editor":"테마 편집기","There are no edges in this data":"이 데이터에는 엣지가 없습니다.","This action cannot be undone.":"이 작업은 취소할 수 없습니다.","This app works by typing":"이 앱은 타이핑으로 작동합니다","This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":"이 문서는 이 컴퓨터에만 저장됩니다. 문서를 저장하고 어디서든 접근할 수 있는 방법이 필요하면 저희 프로 플랜을 3달러에 체험해보세요!","This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":"이 기능은 프로 사용자에게만 제공됩니다. <0>프로 사용자가 되어</0> 이 기능을 사용할 수 있게 하세요.","This is a temporary chart, it will be deleted when you clear your browser cache.":"이것은 임시 차트로, 브라우저 캐시를 지우면 삭제됩니다.","Tip":"팁","To fix this change one of the edge IDs":"이를 수정하려면, 가장자리 ID 중 하나를 변경하십시오.","To fix this change one of the node IDs":"이것을 수정하려면 노드 ID 중 하나를 변경하십시오","To fix this move one pointer to the next line":"이것을 수정하려면 포인터를 다음 줄로 이동하십시오","To fix this move the pointer to the next line":"이것을 수정하려면 포인터를 다음 줄로 이동하십시오","To fix this start the container <0/> on a different line":"이것을 수정하려면 컨테이너 <0/>을 다른 줄에 시작하십시오","To learn more about why we require you to log in, please read <0>this blog post</0>.":"왜 로그인을 해야하는지 자세히 알아보려면 <0>이 블로그 글</0>을 읽어보세요.","Top to Bottom":"위에서 아래로","Try again":"다시 시도하세요","Two edges have the same ID":"두 개의 간선이 같은 ID를 가지고 있습니다","Two nodes have the same ID":"두 개의 노드가 같은 ID를 가지고 있습니다","Type":"유형","Unfreeze":"녹이다","Unique text value to identify a node":"노드를 식별하기 위한 고유한 텍스트 값","Unknown Parsing Error":"알 수 없는 구문 분석 오류","Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":"Flowchart Fun Pro를 사용하여 창의력을 발휘하고 작업흐름을 최적화하세요. $3/월부터 시작합니다!","Update Email":"이메일 업데이트","Upgrade to Pro":"프로로 업그레이드","Upload your File":"파일을 업로드하세요.","Use AI to create flowcharts from data sets or algorithms":"데이터 집합 또는 알고리즘에서 플로우차트를 만들기 위해 AI를 사용하세요.","Use AI to generate a flowchart from a prompt. <0>Learn More</0>":"AI를 사용하여 프롬프트에세이에서 플로차트를 생성합니다. <0>자세히 알아보기</0>","Use classes to group nodes":"노드를 그룹화하기 위해 클래스를 사용하십시오.","Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":"새 탭에서 링크를 설정하기 위해 속성 <0>href</0>을 사용하십시오.","Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":"노드의 이미지를 설정하려면 <0>src</0> 속성을 사용하세요. 이미지는 노드에 맞게 크기가 조정되므로 노드의 너비와 높이를 조절해야 할 수도 있습니다. CORS로 차단되지 않은 공개 이미지만 지원됩니다.","Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":"노드의 너비와 높이를 명시적으로 설정하려면 <0>w</0> 및 <1>h</1> 속성을 사용하세요.","Use the customer portal to change your billing information.":"청구 정보를 변경하려면 고객 포털을 사용하십시오.","Use this file for org charts, hierarchies, and other organizational structures.":"조직도, 계층 구조 및 기타 조직 구조를 위해 이 파일을 사용하십시오.","Use this file for sequences, processes, and workflows.":"시퀀스, 프로세스 및 워크플로우에 대해 이 파일을 사용하십시오.","User":"사용자","View on Github":"Github에서 보기","Water evaporates from the Earth\'s surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":"물은 지구의 표면에서 증발하여 공기로 올라가 비로 내려오고, 그 물은 강과 호수, 바다로 흐르고 다시 증발하여 공기로 돌아가 재생됩니다.","Welcome to Flowchart Fun Pro!":"Flowchart Fun Pro에 오신 것을 환영합니다!","Width and Height":"너비와 높이","Would you like to continue?":"계속하시겠습니까?","Yearly":"연간","You already have a flowchart with this name.":"이 이름을 가진 다이어그램이 이미 있습니다.","You are about to add {numNodes} nodes and {numEdges} edges to your graph.":["그래프에 ",["numNodes"],"개의 노드와 ",["numEdges"],"개의 간선을 추가하려고 합니다."],"You can also wrap text in \\"\\\\( \\\\)\\" to create an edge":"텍스트를 \\"\\\\( \\\\)\\"로 감싸서 가장자리를 만들 수도 있습니다.","You must log in to create a persistent flowchart.":"끈기 있는 플로우차트를 만들기 위해 로그인해야합니다.","You need to log in to access this page.":"페이지에 접근하려면 로그인해야합니다.","You\'re almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":"거의 다 됐습니다! <0>Flowchart Fun Pro</0>의 전력을 풀어내는 마지막 한 단계만 남았습니다. 아래에 결제 정보를 입력하여 구독을 완료하고 오늘부터 놀라운 플로우 차트를 만들 수 있습니다.","Your Charts":"당신의 차트","Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":"귀하의 구독이 더 이상 활성화되지 않습니다. 영구 차트를 만들고 편집하려면 Pro로 업그레이드하십시오.","Zoom In":"확대","Zoom Out":"축소하기","any text: before a colon creates a label":"콜론 앞의 모든 텍스트는: 레이블을 생성합니다","or":"또는"}'
+  ),
 };
