@@ -1,458 +1,5 @@
 /*eslint-disable*/ module.exports = {
-  messages: {
-    "$3 per month": "$3 por mês",
-    "$30 per year": "$30 por ano",
-    "(before a colon creates a label)":
-      "(antes de dois pontos cria uma etiqueta)",
-    "<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":
-      "<0>Flowchart Fun</0> é um projeto de código aberto feito por <1>Tone Row</1>",
-    "A new version of the app is available. Please reload to update.":
-      "Uma nova versão do aplicativo está disponível. Por favor, recarregue para atualizar.",
-    About: "Sobre",
-    "Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":
-      "Acesse esses gráficos de qualquer lugar.<0/>Compartilhe e incorpore fluxogramas que permanecem sincronizados.",
-    "Accessible from any device": "Acessível de qualquer dispositivo",
-    Account: "Conta",
-    "Activate your account": "Ative sua conta",
-    Advanced: "Avançado",
-    Amount: "Total",
-    "An error occurred. Try resubmitting or email {0} directly.": [
-      "Ocorreu um erro. Tente reenviar ou envie um e-mail direto para ",
-      ["0"],
-      ".",
-    ],
-    Appearance: "Aparência",
-    "Apply Style": "Aplicar Estilo",
-    Attributes: "Atributos",
-    "Auto-Layouts": "Layouts automáticos",
-    "Back To Editor": "Voltar ao editor",
-    "Basic Flowchart": "Fluxograma Básico",
-    "Become a Github Sponsor": "Seja um patrocinador do Github",
-    "Become a Pro User": "Se torne um usuário Pro",
-    "Begin with a simple example showing how <0>Flowchart Fun</0> works.":
-      "Comece com um exemplo simples mostrando como o <0>Flowchart Fun</0> funciona.",
-    Blank: "Branco",
-    Blog: "Blog",
-    "Book a Meeting": "Reserve uma Reunião",
-    Border: "Borda",
-    "Bottom to Top": "De baixo para cima",
-    Breadthfirst: "Por extensão",
-    Cancel: "Cancelar",
-    "Cancel your subscription. Your hosted charts will become read-only.":
-      "Cancele sua inscrição. Seus diagramas hospedados não serão modificáveis.",
-    "Certain attributes can be used to customize the appearance or functionality of elements.":
-      "Certos atributos podem ser usados para personalizar a aparência ou funcionalidade dos elementos.",
-    "Change Email Address": "Mude o endereço de email",
-    Changelog: "Registro de alterações",
-    Charts: "Diagramas",
-    "Check your email for a link to log in.<0/>You can close this window.":
-      "Verifique seu e-mail para um link para fazer login. Você pode fechar esta janela.",
-    "Choose a Theme": "Escolha um Tema",
-    "Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":
-      "Escolha entre uma variedade de formas de seta para a origem e o destino de uma aresta. As formas incluem triângulo, triângulo-tee, triângulo-círculo, triângulo-cruz, triângulo-curva-inversa, vee, tee, quadrado, círculo, diamante, chevron e nenhum.",
-    "Choose the plan that's right for you and start creating amazing flowcharts with Flowchart Fun Pro":
-      "Escolha o plano certo para você e comece a criar incríveis fluxogramas com o Flowchart Fun Pro.",
-    Circle: "Círculo",
-    Classes: "Classes",
-    Clay: "Argila",
-    Clear: "Apagar",
-    "Clear text?": "Limpar o texto?",
-    "Click on \\nLearn Syntax\\nto learn more":
-      "Clique em \\nAprenda Sintaxe\\n para aprender mais",
-    "Click to preview": "Clique para pré-visualizar",
-    Clone: "Clonar",
-    Close: "Fechar",
-    Color: "Cor",
-    "Colors include red, orange, yellow, blue, purple, black, white, and gray.":
-      "As cores incluem vermelho, laranja, amarelo, azul, roxo, preto, branco e cinza.",
-    Column: "Coluna",
-    "Comic Book": "História em Quadrinhos",
-    Comment: "Comente",
-    Concentric: "Concêntrico",
-    "Confirm New Email": "Confirme o novo email",
-    Containers: "Recipientes",
-    "Containers are nodes that contain other nodes. They are declared using curly braces.":
-      "Os containers são nós que contêm outros nós. Eles são declarados usando chaves.",
-    Continue: "Continuar",
-    "Convert this to a hosted chart?":
-      "Converter isso para um gráfico hospedado?",
-    "Convert to hosted chart?": "Converter em diagrama hospedado?",
-    Copy: "Copiar",
-    "Copy PNG Image": "Copiar imagem PNG",
-    "Copy SVG Code": "Copiar código SVG",
-    "Copy your mermaid.js code or open it directly in the mermaid.js live editor.":
-      "Copie seu código mermaid.js ou abra diretamente no editor ao vivo mermaid.js.",
-    "Create Flowcharts from a Prompt with AI":
-      "Crie diagramas usando prompts de IA",
-    "Create New Flowchart": "Criar Novo Fluxograma",
-    "Create a New Flowchart": "Criar um novo diagrama",
-    "Create professional-quality visual aids for presentations or training materials":
-      "Crie recursos visuais de qualidade profissional para apresentações ou materiais de treinamento",
-    "Creating an edge between two nodes is done by indenting the second node below the first":
-      "Criar uma aresta entre dois nós é feito indentando o segundo nó abaixo do primeiro",
-    "Custom Sharing Options": "Opções de Compartilhamento Personalizadas",
-    "Customer Portal": "Portal do cliente",
-    "Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":
-      "Personalize seu tema editando o <0>CSS do Cytoscape</0> abaixo. Nossa documentação de estilos está chegando em breve! Enquanto isso, o melhor recurso é a documentação do <1>Cytoscape</1>. Venha perguntar no <2>Discord</2> se ficar preso.",
-    Dark: "Escuro",
-    "Dark Mode": "Modo escuro",
-    Date: "Data",
-    "Deleted when browser data is cleared":
-      "Excluído quando os dados do navegador são apagados",
-    "Describe the flowchart you wish to create":
-      "Descreva o diagrama que você deseja criar",
-    Direction: "Direção",
-    Discard: "Descartar",
-    "Do you want to cancel your subscription?":
-      "Você deseja cancelar sua inscrição?",
-    "Do you want to delete this?": "Você deseja deletar isso?",
-    Document: "Documento",
-    "Don't have an account?": "Não tem uma conta?",
-    Download: "Baixar",
-    "Download JPG": "Baixar JPG",
-    "Download PNG": "Baixar PNG",
-    "Download SVG": "Baixar SVG",
-    "Drag and drop a CSV file here, or click to select a file":
-      "Arraste e solte um arquivo CSV aqui ou clique para selecionar um arquivo",
-    "Draw an edge from multiple nodes by beginning the line with a reference":
-      "Desenhe uma aresta de vários nós começando a linha com uma referência",
-    "Drop the file here ...": "Solte o arquivo aqui ...",
-    "Easily import existing flowcharts from other software":
-      "Importe facilmente fluxogramas existentes de outros softwares",
-    Edge: "Borda",
-    "Edge ID, Classes, Attributes": "ID de Borda, Classes, Atributos",
-    "Edge Label": "Rótulo de Borda",
-    "Edge Label Column": "Coluna de Rótulo de Borda",
-    "Edge Style": "Estilo de Borda",
-    "Edge missing indentation": "Recuo em falta na borda",
-    Edges: "Bordas",
-    "Edges are declared in the same row as their source node":
-      "As bordas são declaradas na mesma linha que seu nó de origem",
-    "Edges are declared in the same row as their target node":
-      "As bordas são declaradas na mesma linha que seu nó de destino",
-    "Edges are declared in their own row":
-      "As bordas são declaradas em sua própria linha",
-    "Edges can also have ID's, classes, and attributes before the label":
-      "As bordas também podem ter ID's, classes e atributos antes da etiqueta",
-    "Edges can be styled with dashed, dotted, or solid lines":
-      "As bordas podem ser estilizadas com linhas tracejadas, pontilhadas ou sólidas",
-    "Edges in Separate Rows": "Bordas em Linhas Separadas",
-    "Edges in Source Node Row": "Bordas na Linha do Nó de Origem",
-    "Edges in Target Node Row": "Bordas na Linha do Nó de Destino",
-    Editable: "Editável",
-    Editor: "Editor",
-    Eggs: "Ovos",
-    Email: "E-mail",
-    Empty: "Vazio",
-    "Enter your email": "Digite seu email",
-    "Equal To": "Igual a",
-    "Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":
-      "Diagrama de fluxo da escrita de ensaios, orientando os alunos através das etapas de brainstorming, delineamento, elaboração e revisão.",
-    "Everything in Free": "Tudo no Plano Gratuito",
-    Excalidraw: "Excalidraw",
-    Export: "Exportar",
-    "Export High-Resolution Images": "Exportar imagens de alta resolução",
-    "Export to SVG": "Exportar para SVG",
-    "Extra Large": "Extra Grande",
-    Extract: "Extrair",
-    Feedback: "Feedback",
-    "Follow Us on Twitter": "Siga-nos no Twitter",
-    "Found a bug? Have a feature request? We would love to hear from you!":
-      "Encontrou um bug? Tem uma solicitação de recurso? Adoraríamos ouvir de você!",
-    Free: "Grátis ",
-    Fullscreen: "Tela cheia",
-    Futuristic: "Futurista",
-    "Get More Layouts": "Obtenha Mais Leiautes",
-    "Get personalized attention and support from our expert team":
-      "Receba atenção e suporte personalizado de nossa equipe especializada",
-    "Getting Started": "Começando",
-    Graph: "Diagrama",
-    Grid: "Grade",
-    "Have complex questions or issues? We're here to help.":
-      "Tem questões ou problemas complexos? Estamos aqui para ajudar.",
-    Help: "Ajuda",
-    History: "Histórico",
-    Home: "Página inicial",
-    "How are edges declared in this data?":
-      "Como as arestas são declaradas nestes dados?",
-    "ID's": "IDs",
-    "If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":
-      "Se você gosta de usar <0>Flowchart Fun</0>, por favor, considere apoiar o projeto",
-    "If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\:</0>":
-      "Se você quiser criar uma borda, indente esta linha. Se não, escape o dois-pontos com uma barra invertida <0>\\:</0>",
-    "Image Export": "Exportação de imagem",
-    Images: "Imagens",
-    "Import Data": "Importar dados",
-    "Import data from a CSV file.": "Importar dados de um arquivo CSV.",
-    "Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":
-      "Importar dados de qualquer arquivo CSV e mapeá-los para um novo fluxograma. Esta é uma ótima maneira de importar dados de outras fontes, como Lucidchart, Google Sheets e Visio.",
-    "Import from Lucidchart, Visio, and other CSV files":
-      "Importar do Lucidchart, do Visio e de outros arquivos CSV",
-    "Import your diagram it into Microsoft Visio using one of these CSV files.":
-      "Importe seu diagrama para o Microsoft Visio usando um desses arquivos CSV.",
-    "Important Changes are Coming to Flowchart Fun.":
-      "Mudanças importantes estão chegando ao Flowchart Fun.",
-    "Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":
-      "Importar dados é um recurso avançado. Você pode atualizar para o Flowchart Fun Pro por apenas US $ 3/mês ou US $ 30/ano.",
-    "Impress your audience with high-quality images produced by Flowchart Fun":
-      "Impressione sua audiência com imagens de alta qualidade produzidas pelo Flowchart Fun",
-    "Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":
-      "Inclua um título usando um atributo <0>title</0>. Para usar a coloração do Visio, adicione um atributo <1>roleType</1> igual a uma das seguintes opções:",
-    "Indenting creates an edge": "O recuo cria um link",
-    Info: "Informações",
-    Instruct: "Instruir",
-    Is: "É",
-    Language: "Idioma",
-    Large: "Grande",
-    Layout: "Layout",
-    "Layout Frozen": "Layout Congelado",
-    "Layout is Frozen": "O layout está congelado",
-    "Leading References": "Principais Referências",
-    "Learn More": "Saber mais",
-    "Learn Syntax": "Aprender Sintaxe",
-    "Learn about Flowchart Fun Pro": "Saiba mais sobre o Flowchart Fun Pro",
-    "Left to Right": "Da esquerda para direita",
-    "Let's get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":
-      "Vamos começar! Digite seu endereço de email abaixo para criar sua conta no Flowchart Fun e começar a usar nossos recursos poderosos.",
-    Light: "Claro",
-    "Light Mode": "Modo claro",
-    Link: "Link",
-    "Load Theme": "Carregar Tema",
-    "Loading a theme will erase any custom styles you have applied":
-      "Carregar um tema apagará quaisquer estilos personalizados que você tenha aplicado",
-    "Lock Zoom to Graph": "Bloquear Zoom para o Gráfico",
-    "Log In": "Acessar",
-    "Log Out": "Deslogar",
-    "Log in to start creating flowcharts.":
-      "Faça o login para começar a criar fluxogramas.",
-    "Make a One-Time Donation": "Faça uma Doação Única",
-    "Make publicly accessible": "Tornar publicamente acessível",
-    "Make sure you use the same email you will use to log in.":
-      "Certifique-se de usar o mesmo e-mail que você usará para fazer o login.",
-    "Map Data": "Mapear Dados",
-    "Market understanding and competitive landscape maintenance for SaaS product development":
-      "Compreensão do mercado e manutenção do cenário competitivo para o desenvolvimento de produtos SaaS.",
-    Medium: "Médio",
-    Method: "Método",
-    Monospace: "Monospace",
-    Monthly: "Mensal",
-    "Multiple pointers on same line": "Múltiplos ponteiros na mesma linha",
-    Museum: "Museu",
-    Name: "Nome",
-    New: "Novo",
-    "New Email": "Novo Email",
-    "Next charge": "Próxima cobrança",
-    "No Edges": "Sem Bordas",
-    Node: "Nó",
-    "Node Border Style": "Estilo de Borda do Nó",
-    "Node Colors": "Cores do Nó",
-    "Node ID": "ID do Nó",
-    "Node ID, Classes, Attributes": "ID do Nó, Classes, Atributos",
-    "Node Label": "Rótulo do Nó",
-    "Node Shapes": "Formas do Nó",
-    "Node and pointer on same line": "Nó e ponteiro na mesma linha",
-    "Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":
-      "Nós podem ser estilizados com traços, pontilhados ou duplos. As bordas também podem ser removidas com border-none.",
-    "Not Empty": "Não Vazio",
-    "Office Hours": "Horário de trabalho",
-    "Once in a while the magic link will end up in your spam folder. If you don't see it after a few minutes, check there or request a new link.":
-      " vez em quando, o link mágico acabará em sua pasta de spam. Se você não o vir após alguns minutos, verifique lá ou solicite um novo link.",
-    "One-on-One Support": "Suporte Individual",
-    "One-time Share Links": "Links de compartilhamento único",
-    "Only available on this device.<0/>Clearing your browser cache will erase them.":
-      "Disponível apenas neste dispositivo.<0/>Limpar o cache do navegador irá apagá-los.",
-    "Open Customer Portal": "Abra o portal do cliente",
-    "Organization Chart": "Gráfico de Organização",
-    "Paste the information you wish to convert to a flowchart":
-      "Cole as informações que deseja converter em um fluxograma",
-    Permalinks: "Permalinks",
-    "Permanent Charts are a Pro Feature":
-      "Gráficos permanentes são um recurso Pro",
-    "Permanent Flowcharts": "Fluxogramas Permanentes",
-    Persistent: "Persistente",
-    "Persistent Flowcharts": "Diagramas Persistentes",
-    Playbook: "Cartilha",
-    "Pointer and container on same line": "Ponteiro e contêiner na mesma linha",
-    Pricing: "Preços",
-    "Process for corporate social responsibility initiatives development and implementation across company operations":
-      "Processo para o desenvolvimento e implementação de iniciativas de responsabilidade social corporativa em todas as operações da empresa.",
-    "Processing Data": "Processando Dados",
-    Prompt: "Estímulo",
-    Public: "Público",
-    Random: "Aleatório",
-    "Read-only": "Somente leitura",
-    "Reference by Class": "Referência por Classe",
-    "Reference by ID": "Referência por ID",
-    "Reference by Label": "Referência por Rótulo",
-    References: "Referências",
-    "References are used to create edges between nodes that are created elsewhere in the document":
-      "Referências são usadas para criar arestas entre nós que são criados em outro lugar no documento",
-    "Referencing a node by its exact label":
-      "Referenciando um nó pelo seu rótulo exato",
-    "Referencing a node by its unique ID":
-      "Referenciando um nó pelo seu ID único",
-    "Referencing multiple nodes with the same assigned class":
-      "Referenciando vários nós com a mesma classe atribuída",
-    "Refresh Page": "Atualizar Página",
-    "Reload to Update": "Recarregar para Atualizar",
-    "Remove Image Watermark": "Remover marca d'água da imagem",
-    Rename: "Renomear",
-    Reset: "Resetar",
-    "Resume Subscription": "Resumir inscrição",
-    Retro: "Retrô",
-    Return: "Retornar",
-    "Right to Left": "Da direita para esquerda",
-    Roadmap: "Roteiro",
-    Save: "Salvar",
-    "Save 20% (2 months free!)": "Economize 20% (2 meses grátis!)",
-    "Save time and automate certain tasks with this cutting-edge feature":
-      "Economize tempo e automatize certas tarefas com essa funcionalidade de ponta.",
-    "Save time and streamline your workflow by importing data from other sources":
-      "Economize tempo e simplifique seu fluxo de trabalho importando dados de outras fontes.",
-    "Select your plan!": "Selecione seu plano!",
-    "Send Feedback": "Enviar Feedback",
-    Settings: "Configurações",
-    Shape: "Forma",
-    "Sign Up": "Inscreva-se",
-    "Sign Up Now": "Cadastre-se agora",
-    Size: "Tamanho",
-    Small: "Pequeno",
-    "Some classes are available to help style your flowchart":
-      "Algumas classes estão disponíveis para ajudar a estilizar seu fluxograma",
-    "Sorry! This page is only available in English.":
-      "Sinto muito! Esta página só está disponível em inglês.",
-    "Source Arrow": "Seta de Origem",
-    "Source Column": "Coluna de Origem",
-    "Source Delimiter": "Delimitador de Origem",
-    "Source/Target Arrow Shape": "Forma da Seta de Origem/Destino",
-    Spacing: "Espaçamento",
-    "Special Attributes": "Atributos Especiais",
-    Start: "Início",
-    "Start Over": "Recomeçar",
-    Status: "Status",
-    "Store any data associated to a node":
-      "Armazenar quaisquer dados associados a um nó",
-    "Stored in the cloud": "Armazenado na nuvem",
-    "Stored on this computer": "Armazenado neste computador",
-    "Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":
-      "Simplifique seu fluxo de trabalho e simplifique sua visualização de processo com o Flowchart Fun. Assine agora e aproveite nossos poderosos recursos Pro!",
-    Style: "Estilo",
-    "Style Classes": "Classes de Estilo",
-    Submit: "Enviar",
-    Subscribe: "Inscrever-se",
-    Subscription: "Inscrição",
-    "Subscription will end": "Inscrição acabará",
-    "Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":
-      "Análise e otimização da cadeia de suprimentos: redução de custos, melhoria da eficiência e colaboração com os stakeholders.",
-    "Target Arrow": "Seta de Destino",
-    "Target Column": "Coluna Alvo",
-    "Target Delimiter": "Delimitador Alvo",
-    Temporary: "Temporário",
-    "Temporary Flowchart Warning": "Aviso de Fluxograma Temporário",
-    "Temporary Flowcharts": "Diagramas Temporários",
-    "Text Sizes": "Tamanhos de Texto",
-    "Text followed by colon+space creates an edge with the text as the label":
-      "Texto seguido de dois-pontos+espaço cria uma aresta com o texto como rótulo",
-    "Text on a line creates a node with the text as the label":
-      "Texto em uma linha cria um nó com o texto como rótulo",
-    "Text sizes include small, regular, large, and extra-large.":
-      "Os tamanhos de texto incluem pequeno, regular, grande e extra grande.",
-    "Thank you for your feedback!": "Agradecimentos pelo seu feedback!",
-    "The column that contains the edge label(s)":
-      "A coluna que contém o(s) rótulo(s) da aresta",
-    "The column that contains the source node ID(s)":
-      "A coluna que contém o(s) ID(s) do nó de origem",
-    "The column that contains the target node ID(s)":
-      "A coluna que contém o(s) ID(s) do nó de destino",
-    "The delimiter used to separate multiple source nodes":
-      "O delimitador usado para separar vários nós de origem",
-    "The delimiter used to separate multiple target nodes":
-      "O delimitador usado para separar vários nós de destino",
-    "The possible shapes are:": "As formas possíveis são:",
-    "Theme Editor": "Editor de Temas",
-    "There are no edges in this data": "Não há arestas nestes dados",
-    "This action cannot be undone.": "Esta ação não pode ser desfeita.",
-    "This app works by typing": "Este aplicativo funciona por digitação",
-    "This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":
-      "Este documento é salvo apenas neste computador. Se você quiser uma maneira de armazenar documentos e acessá-los de qualquer lugar, experimente nosso plano Pro por apenas US$ 3/mês!",
-    "This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":
-      "Esta funcionalidade está disponível apenas para usuários Pro. <0>Torne-se um usuário Pro</0> para desbloqueá-la.",
-    "This is a temporary chart, it will be deleted when you clear your browser cache.":
-      "Este é um gráfico temporário, ele será excluído quando você limpar o cache do seu navegador.",
-    Tip: "Dica",
-    "To fix this change one of the edge IDs":
-      "Para corrigir isso, altere um dos IDs de borda",
-    "To fix this change one of the node IDs":
-      "Para corrigir isso, altere um dos IDs de nó",
-    "To fix this move one pointer to the next line":
-      "Para corrigir isso, mova um ponteiro para a próxima linha",
-    "To fix this move the pointer to the next line":
-      "Para corrigir isso, mova o ponteiro para a próxima linha",
-    "To fix this start the container <0/> on a different line":
-      "Para corrigir isso, inicie o container <0/> em uma linha diferente",
-    "Top to Bottom": "De cima para baixo",
-    "Try again": "Tente novamente",
-    "Two edges have the same ID": "Dois bordos têm o mesmo ID",
-    "Two nodes have the same ID": "Dois nós têm o mesmo ID",
-    Type: "Tipo",
-    Unfreeze: "Descongelar",
-    "Unique text value to identify a node":
-      "Valor de texto único para identificar um nó",
-    "Unknown Parsing Error": "Erro de Análise Desconhecido",
-    "Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":
-      "Libere sua criatividade e simplifique seu fluxo de trabalho com o Flowchart Fun Pro - a partir de apenas $ 3 / mês!",
-    "Update Email": "Atualizar e-mail",
-    "Upgrade to Pro": "Atualizar para Pro",
-    "Upload your File": "Faça o upload do seu arquivo",
-    "Use AI to create flowcharts from data sets or algorithms":
-      "Use IA para criar fluxogramas a partir de conjuntos de dados ou algoritmos.",
-    "Use AI to generate a flowchart from a prompt. <0>Learn More</0>":
-      "Use IA para gerar um fluxograma a partir de uma solicitação. <0>Saiba mais</0>.",
-    "Use classes to group nodes": "Use classes para agrupar nós",
-    "Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":
-      "Use o atributo <0>href</0> para definir um link em um nó que abra em uma nova guia.",
-    "Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":
-      "Use o atributo <0>src</0> para definir a imagem de um nó. A imagem será dimensionada para caber no nó, então você pode precisar ajustar a largura e altura do nó para obter o resultado desejado. Apenas imagens públicas (não bloqueadas por CORS) são suportadas.",
-    "Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":
-      "Use os atributos <0>w</0> e <1>h</1> para definir explicitamente a largura e altura de um nó.",
-    "Use the customer portal to change your billing information.":
-      "Use o portal do cliente para alterar suas informações de cobrança.",
-    "Use this file for org charts, hierarchies, and other organizational structures.":
-      "Use este arquivo para organogramas, hierarquias e outras estruturas organizacionais.",
-    "Use this file for sequences, processes, and workflows.":
-      "Use este arquivo para sequências, processos e fluxos de trabalho.",
-    User: "Usuário",
-    "View on Github": "Visualizar no Github",
-    "Water evaporates from the Earth's surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":
-      "água evapora da superfície da Terra, sobe para a atmosfera e volta como precipitação. Esta água então corre para rios, lagos e oceanos, onde novamente evapora e é reciclada de volta para a atmosfera.",
-    "We use magic links to log you in. Enter your email below to get started.":
-      "Nós usamos links mágicos para logar você. Digite seu email abaixo para começar.",
-    "Welcome to Flowchart Fun Pro!": "Bem-vindo ao Flowchart Fun Pro!",
-    "Width and Height": "Largura e Altura",
-    "Would you like to continue?": "Você gostaria de continuar?",
-    Yearly: "Anualmente",
-    "You already have a flowchart with this name.":
-      "Você já tem um diagrama com este nome.",
-    "You are about to add {numNodes} nodes and {numEdges} edges to your graph.":
-      [
-        "Você está prestes a adicionar ",
-        ["numNodes"],
-        " nós e ",
-        ["numEdges"],
-        " arestas ao seu gráfico.",
-      ],
-    'You can also wrap text in "\\( \\)" to create an edge':
-      'Você também pode quebrar o texto dentro de "\\( \\)" para criar um link',
-    "You must log in to create a persistent flowchart.":
-      "Você deve fazer login para criar um fluxograma persistente.",
-    "You're almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":
-      "Você está quase lá! Apenas mais um passo para destravar o potencial total do <0>Flowchart Fun Pro</0>. Insira os detalhes do pagamento abaixo para completar sua assinatura e começar a criar fluxogramas incríveis hoje.",
-    "Your Charts": "Seus Gráficos",
-    "Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":
-      "Sua assinatura não está mais ativa. Se você deseja criar e editar gráficos permanentes, atualize para o Pro.",
-    "Zoom In": "Zoom In",
-    "Zoom Out": "Diminuir o zoom",
-    "any text: before a colon creates a label":
-      "qualquer texto: antes de dois pontos cria uma etiqueta",
-  },
+  messages: JSON.parse(
+    '{"$3 per month":"$3 por mês","$30 per year":"$30 por ano","(before a colon creates a label)":"(antes de dois pontos cria uma etiqueta)","<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":"<0>Flowchart Fun</0> é um projeto de código aberto feito por <1>Tone Row</1>","<0>Sign In</0> / <1>Sign Up</1> with email and password":"<0>Entrar</0> / <1>Cadastrar</1> com e-mail e senha","<0>Starting August 28th you will need a subscription to create and edit charts. <1>Learn more</1>.</0>":"<0>A partir de 28 de agosto, você precisará de uma assinatura para criar e editar gráficos. <1>Saiba mais</1>.</0>","<0>You currently have a free account.<1/><2>Learn about our Pro Features and subscribe on our pricing page</2>.</0>":"<0>Você atualmente tem uma conta gratuita. <1/><2>Saiba mais sobre nossos recursos Pro e assine nossa página de preços</2>.</0>","A new version of the app is available. Please reload to update.":"Uma nova versão do aplicativo está disponível. Por favor, recarregue para atualizar.","About":"Sobre","Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":"Acesse esses gráficos de qualquer lugar.<0/>Compartilhe e incorpore fluxogramas que permanecem sincronizados.","Accessible from any device":"Acessível de qualquer dispositivo","Account":"Conta","Activate your account":"Ative sua conta","Advanced":"Avançado","Amount":"Total","An error occurred. Try resubmitting or email {0} directly.":["Ocorreu um erro. Tente reenviar ou envie um e-mail direto para ",["0"],"."],"Appearance":"Aparência","Apply Style":"Aplicar Estilo","Attributes":"Atributos","Auto-Layouts":"Layouts automáticos","Back To Editor":"Voltar ao editor","Basic Flowchart":"Fluxograma Básico","Become a Github Sponsor":"Seja um patrocinador do Github","Become a Pro User":"Se torne um usuário Pro","Begin with a simple example showing how <0>Flowchart Fun</0> works.":"Comece com um exemplo simples mostrando como o <0>Flowchart Fun</0> funciona.","Blank":"Branco","Blog":"Blog","Book a Meeting":"Reserve uma Reunião","Border":"Borda","Bottom to Top":"De baixo para cima","Breadthfirst":"Por extensão","Cancel":"Cancelar","Cancel your subscription. Your hosted charts will become read-only.":"Cancele sua inscrição. Seus diagramas hospedados não serão modificáveis.","Certain attributes can be used to customize the appearance or functionality of elements.":"Certos atributos podem ser usados para personalizar a aparência ou funcionalidade dos elementos.","Change Email Address":"Mude o endereço de email","Changelog":"Registro de alterações","Charts":"Diagramas","Check your email for a link to log in.<0/>You can close this window.":"Verifique seu e-mail para um link para fazer login. Você pode fechar esta janela.","Choose":"Escolha","Choose a Theme":"Escolha um Tema","Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":"Escolha entre uma variedade de formas de seta para a origem e o destino de uma aresta. As formas incluem triângulo, triângulo-tee, triângulo-círculo, triângulo-cruz, triângulo-curva-inversa, vee, tee, quadrado, círculo, diamante, chevron e nenhum.","Choose the plan that\'s right for you and start creating amazing flowcharts with Flowchart Fun Pro":"Escolha o plano certo para você e comece a criar incríveis fluxogramas com o Flowchart Fun Pro.","Circle":"Círculo","Classes":"Classes","Clay":"Argila","Clear":"Apagar","Clear text?":"Limpar o texto?","Click on \\\\nLearn Syntax\\\\nto learn more":"Clique em \\\\nAprenda Sintaxe\\\\n para aprender mais","Click to preview":"Clique para pré-visualizar","Clone":"Clonar","Close":"Fechar","Color":"Cor","Colors include red, orange, yellow, blue, purple, black, white, and gray.":"As cores incluem vermelho, laranja, amarelo, azul, roxo, preto, branco e cinza.","Column":"Coluna","Comic Book":"História em Quadrinhos","Comment":"Comente","Concentric":"Concêntrico","Confirm New Email":"Confirme o novo email","Confirm your email address to sign in.":"Confirme seu endereço de e-mail para fazer login.","Containers":"Recipientes","Containers are nodes that contain other nodes. They are declared using curly braces.":"Os containers são nós que contêm outros nós. Eles são declarados usando chaves.","Continue":"Continuar","Convert this to a hosted chart?":"Converter isso para um gráfico hospedado?","Convert to hosted chart?":"Converter em diagrama hospedado?","Copy":"Copiar","Copy PNG Image":"Copiar imagem PNG","Copy SVG Code":"Copiar código SVG","Copy your mermaid.js code or open it directly in the mermaid.js live editor.":"Copie seu código mermaid.js ou abra diretamente no editor ao vivo mermaid.js.","Create Flowcharts from a Prompt with AI":"Crie diagramas usando prompts de IA","Create New Flowchart":"Criar Novo Fluxograma","Create a New Flowchart":"Criar um novo diagrama","Create professional-quality visual aids for presentations or training materials":"Crie recursos visuais de qualidade profissional para apresentações ou materiais de treinamento","Creating an edge between two nodes is done by indenting the second node below the first":"Criar uma aresta entre dois nós é feito indentando o segundo nó abaixo do primeiro","Custom Sharing Options":"Opções de Compartilhamento Personalizadas","Customer Portal":"Portal do cliente","Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":"Personalize seu tema editando o <0>CSS do Cytoscape</0> abaixo. Nossa documentação de estilos está chegando em breve! Enquanto isso, o melhor recurso é a documentação do <1>Cytoscape</1>. Venha perguntar no <2>Discord</2> se ficar preso.","Dark":"Escuro","Dark Mode":"Modo escuro","Date":"Data","Deleted when browser data is cleared":"Excluído quando os dados do navegador são apagados","Describe the flowchart you wish to create":"Descreva o diagrama que você deseja criar","Direction":"Direção","Discard":"Descartar","Do you want to cancel your subscription?":"Você deseja cancelar sua inscrição?","Do you want to delete this?":"Você deseja deletar isso?","Document":"Documento","Download":"Baixar","Download JPG":"Baixar JPG","Download PNG":"Baixar PNG","Download SVG":"Baixar SVG","Drag and drop a CSV file here, or click to select a file":"Arraste e solte um arquivo CSV aqui ou clique para selecionar um arquivo","Draw an edge from multiple nodes by beginning the line with a reference":"Desenhe uma aresta de vários nós começando a linha com uma referência","Drop the file here ...":"Solte o arquivo aqui ...","Easily import existing flowcharts from other software":"Importe facilmente fluxogramas existentes de outros softwares","Edge":"Borda","Edge ID, Classes, Attributes":"ID de Borda, Classes, Atributos","Edge Label":"Rótulo de Borda","Edge Label Column":"Coluna de Rótulo de Borda","Edge Style":"Estilo de Borda","Edge missing indentation":"Recuo em falta na borda","Edges":"Bordas","Edges are declared in the same row as their source node":"As bordas são declaradas na mesma linha que seu nó de origem","Edges are declared in the same row as their target node":"As bordas são declaradas na mesma linha que seu nó de destino","Edges are declared in their own row":"As bordas são declaradas em sua própria linha","Edges can also have ID\'s, classes, and attributes before the label":"As bordas também podem ter ID\'s, classes e atributos antes da etiqueta","Edges can be styled with dashed, dotted, or solid lines":"As bordas podem ser estilizadas com linhas tracejadas, pontilhadas ou sólidas","Edges in Separate Rows":"Bordas em Linhas Separadas","Edges in Source Node Row":"Bordas na Linha do Nó de Origem","Edges in Target Node Row":"Bordas na Linha do Nó de Destino","Editable":"Editável","Editor":"Editor","Eggs":"Ovos","Email":"E-mail","Empty":"Vazio","Enter your email":"Digite seu email","Enter your email address and we\'ll send you a magic link to sign in.":"Digite seu endereço de e-mail e enviaremos um link mágico para fazer login.","Enter your email address below and we\'ll send you a link to reset your password.":"Digite seu endereço de e-mail abaixo e enviaremos um link para redefinir sua senha.","Equal To":"Igual a","Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":"Diagrama de fluxo da escrita de ensaios, orientando os alunos através das etapas de brainstorming, delineamento, elaboração e revisão.","Everything in Free":"Tudo no Plano Gratuito","Excalidraw":"Excalidraw","Export":"Exportar","Export High-Resolution Images":"Exportar imagens de alta resolução","Export to SVG":"Exportar para SVG","Extra Large":"Extra Grande","Extract":"Extrair","Feedback":"Feedback","Follow Us on Twitter":"Siga-nos no Twitter","Forgot your password?":"Esqueceu sua senha?","Found a bug? Have a feature request? We would love to hear from you!":"Encontrou um bug? Tem uma solicitação de recurso? Adoraríamos ouvir de você!","Free":"Grátis ","Fullscreen":"Tela cheia","Futuristic":"Futurista","Get More Layouts":"Obtenha Mais Leiautes","Get personalized attention and support from our expert team":"Receba atenção e suporte personalizado de nossa equipe especializada","Getting Started":"Começando","Graph":"Diagrama","Grid":"Grade","Have complex questions or issues? We\'re here to help.":"Tem questões ou problemas complexos? Estamos aqui para ajudar.","Help":"Ajuda","History":"Histórico","Home":"Página inicial","How are edges declared in this data?":"Como as arestas são declaradas nestes dados?","ID\'s":"IDs","If an account with that email exists, we\'ve sent you an email with instructions on how to reset your password.":"Se uma conta com esse e-mail existir, enviamos um e-mail com instruções sobre como redefinir sua senha.","If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":"Se você gosta de usar <0>Flowchart Fun</0>, por favor, considere apoiar o projeto","If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\\\:</0>":"Se você quiser criar uma borda, indente esta linha. Se não, escape o dois-pontos com uma barra invertida <0>\\\\:</0>","Image Export":"Exportação de imagem","Images":"Imagens","Import Data":"Importar dados","Import data from a CSV file.":"Importar dados de um arquivo CSV.","Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":"Importar dados de qualquer arquivo CSV e mapeá-los para um novo fluxograma. Esta é uma ótima maneira de importar dados de outras fontes, como Lucidchart, Google Sheets e Visio.","Import from Lucidchart, Visio, and other CSV files":"Importar do Lucidchart, do Visio e de outros arquivos CSV","Import your diagram it into Microsoft Visio using one of these CSV files.":"Importe seu diagrama para o Microsoft Visio usando um desses arquivos CSV.","Important Changes are Coming to Flowchart Fun.":"Mudanças importantes estão chegando ao Flowchart Fun.","Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":"Importar dados é um recurso avançado. Você pode atualizar para o Flowchart Fun Pro por apenas US $ 3/mês ou US $ 30/ano.","Impress your audience with high-quality images produced by Flowchart Fun":"Impressione sua audiência com imagens de alta qualidade produzidas pelo Flowchart Fun","Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":"Inclua um título usando um atributo <0>title</0>. Para usar a coloração do Visio, adicione um atributo <1>roleType</1> igual a uma das seguintes opções:","Indenting creates an edge":"O recuo cria um link","Info":"Informações","Instruct":"Instruir","Is":"É","Language":"Idioma","Large":"Grande","Layout":"Layout","Layout Frozen":"Layout Congelado","Layout is Frozen":"O layout está congelado","Leading References":"Principais Referências","Learn More":"Saber mais","Learn Syntax":"Aprender Sintaxe","Learn about Flowchart Fun Pro":"Saiba mais sobre o Flowchart Fun Pro","Left to Right":"Da esquerda para direita","Let\'s get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":"Vamos começar! Digite seu endereço de email abaixo para criar sua conta no Flowchart Fun e começar a usar nossos recursos poderosos.","Light":"Claro","Light Mode":"Modo claro","Link":"Link","Load Theme":"Carregar Tema","Loading a theme will erase any custom styles you have applied":"Carregar um tema apagará quaisquer estilos personalizados que você tenha aplicado","Lock Zoom to Graph":"Bloquear Zoom para o Gráfico","Log In":"Acessar","Log Out":"Deslogar","Log in to start creating flowcharts.":"Faça o login para começar a criar fluxogramas.","Make a One-Time Donation":"Faça uma Doação Única","Make publicly accessible":"Tornar publicamente acessível","Make sure you use the same email you will use to log in.":"Certifique-se de usar o mesmo e-mail que você usará para fazer o login.","Map Data":"Mapear Dados","Market understanding and competitive landscape maintenance for SaaS product development":"Compreensão do mercado e manutenção do cenário competitivo para o desenvolvimento de produtos SaaS.","Medium":"Médio","Method":"Método","Monospace":"Monospace","Monthly":"Mensal","Multiple pointers on same line":"Múltiplos ponteiros na mesma linha","Museum":"Museu","Name":"Nome","New":"Novo","New Email":"Novo Email","Next charge":"Próxima cobrança","No Edges":"Sem Bordas","Node":"Nó","Node Border Style":"Estilo de Borda do Nó","Node Colors":"Cores do Nó","Node ID":"ID do Nó","Node ID, Classes, Attributes":"ID do Nó, Classes, Atributos","Node Label":"Rótulo do Nó","Node Shapes":"Formas do Nó","Node and pointer on same line":"Nó e ponteiro na mesma linha","Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":"Nós podem ser estilizados com traços, pontilhados ou duplos. As bordas também podem ser removidas com border-none.","Not Empty":"Não Vazio","Office Hours":"Horário de trabalho","Once in a while the magic link will end up in your spam folder. If you don\'t see it after a few minutes, check there or request a new link.":" vez em quando, o link mágico acabará em sua pasta de spam. Se você não o vir após alguns minutos, verifique lá ou solicite um novo link.","One-on-One Support":"Suporte Individual","One-time Share Links":"Links de compartilhamento único","Only available on this device.<0/>Clearing your browser cache will erase them.":"Disponível apenas neste dispositivo.<0/>Limpar o cache do navegador irá apagá-los.","Open Customer Portal":"Abra o portal do cliente","Organization Chart":"Gráfico de Organização","Password":"Senha","Paste the information you wish to convert to a flowchart":"Cole as informações que deseja converter em um fluxograma","Permalinks":"Permalinks","Permanent Charts are a Pro Feature":"Gráficos permanentes são um recurso Pro","Permanent Flowcharts":"Fluxogramas Permanentes","Persistent":"Persistente","Persistent Flowcharts":"Diagramas Persistentes","Playbook":"Cartilha","Pointer and container on same line":"Ponteiro e contêiner na mesma linha","Pricing":"Preços","Process for corporate social responsibility initiatives development and implementation across company operations":"Processo para o desenvolvimento e implementação de iniciativas de responsabilidade social corporativa em todas as operações da empresa.","Processing Data":"Processando Dados","Prompt":"Estímulo","Public":"Público","Random":"Aleatório","Read-only":"Somente leitura","Reference by Class":"Referência por Classe","Reference by ID":"Referência por ID","Reference by Label":"Referência por Rótulo","References":"Referências","References are used to create edges between nodes that are created elsewhere in the document":"Referências são usadas para criar arestas entre nós que são criados em outro lugar no documento","Referencing a node by its exact label":"Referenciando um nó pelo seu rótulo exato","Referencing a node by its unique ID":"Referenciando um nó pelo seu ID único","Referencing multiple nodes with the same assigned class":"Referenciando vários nós com a mesma classe atribuída","Refresh Page":"Atualizar Página","Reload to Update":"Recarregar para Atualizar","Remove Image Watermark":"Remover marca d\'água da imagem","Rename":"Renomear","Request Magic Link":"Solicitar Link Mágico","Request Password Reset":"Solicitar Redefinição de Senha","Reset":"Resetar","Reset Password":"Redefinir Senha","Resume Subscription":"Resumir inscrição","Retro":"Retrô","Return":"Retornar","Right to Left":"Da direita para esquerda","Roadmap":"Roteiro","Save":"Salvar","Save 20% (2 months free!)":"Economize 20% (2 meses grátis!)","Save time and automate certain tasks with this cutting-edge feature":"Economize tempo e automatize certas tarefas com essa funcionalidade de ponta.","Save time and streamline your workflow by importing data from other sources":"Economize tempo e simplifique seu fluxo de trabalho importando dados de outras fontes.","Select your plan!":"Selecione seu plano!","Send Feedback":"Enviar Feedback","Settings":"Configurações","Shape":"Forma","Sign In":"Entrar","Sign Up":"Inscreva-se","Sign in with <0>GitHub</0>":"Entrar com <0>GitHub</0>","Sign in with <0>Google</0>":"Entrar com <0>Google</0>","Size":"Tamanho","Small":"Pequeno","Some classes are available to help style your flowchart":"Algumas classes estão disponíveis para ajudar a estilizar seu fluxograma","Sorry! This page is only available in English.":"Sinto muito! Esta página só está disponível em inglês.","Source Arrow":"Seta de Origem","Source Column":"Coluna de Origem","Source Delimiter":"Delimitador de Origem","Source/Target Arrow Shape":"Forma da Seta de Origem/Destino","Spacing":"Espaçamento","Special Attributes":"Atributos Especiais","Start":"Início","Start Over":"Recomeçar","Status":"Status","Store any data associated to a node":"Armazenar quaisquer dados associados a um nó","Stored in the cloud":"Armazenado na nuvem","Stored on this computer":"Armazenado neste computador","Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":"Simplifique seu fluxo de trabalho e simplifique sua visualização de processo com o Flowchart Fun. Assine agora e aproveite nossos poderosos recursos Pro!","Style":"Estilo","Style Classes":"Classes de Estilo","Submit":"Enviar","Subscribe":"Inscrever-se","Subscription":"Inscrição","Subscription will end":"Inscrição acabará","Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":"Análise e otimização da cadeia de suprimentos: redução de custos, melhoria da eficiência e colaboração com os stakeholders.","Target Arrow":"Seta de Destino","Target Column":"Coluna Alvo","Target Delimiter":"Delimitador Alvo","Temporary":"Temporário","Temporary Flowchart Warning":"Aviso de Fluxograma Temporário","Temporary Flowcharts":"Diagramas Temporários","Text Sizes":"Tamanhos de Texto","Text followed by colon+space creates an edge with the text as the label":"Texto seguido de dois-pontos+espaço cria uma aresta com o texto como rótulo","Text on a line creates a node with the text as the label":"Texto em uma linha cria um nó com o texto como rótulo","Text sizes include small, regular, large, and extra-large.":"Os tamanhos de texto incluem pequeno, regular, grande e extra grande.","Thank you for your feedback!":"Agradecimentos pelo seu feedback!","The column that contains the edge label(s)":"A coluna que contém o(s) rótulo(s) da aresta","The column that contains the source node ID(s)":"A coluna que contém o(s) ID(s) do nó de origem","The column that contains the target node ID(s)":"A coluna que contém o(s) ID(s) do nó de destino","The delimiter used to separate multiple source nodes":"O delimitador usado para separar vários nós de origem","The delimiter used to separate multiple target nodes":"O delimitador usado para separar vários nós de destino","The possible shapes are:":"As formas possíveis são:","Theme Editor":"Editor de Temas","There are no edges in this data":"Não há arestas nestes dados","This action cannot be undone.":"Esta ação não pode ser desfeita.","This app works by typing":"Este aplicativo funciona por digitação","This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":"Este documento é salvo apenas neste computador. Se você quiser uma maneira de armazenar documentos e acessá-los de qualquer lugar, experimente nosso plano Pro por apenas US$ 3/mês!","This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":"Esta funcionalidade está disponível apenas para usuários Pro. <0>Torne-se um usuário Pro</0> para desbloqueá-la.","This is a temporary chart, it will be deleted when you clear your browser cache.":"Este é um gráfico temporário, ele será excluído quando você limpar o cache do seu navegador.","Tip":"Dica","To fix this change one of the edge IDs":"Para corrigir isso, altere um dos IDs de borda","To fix this change one of the node IDs":"Para corrigir isso, altere um dos IDs de nó","To fix this move one pointer to the next line":"Para corrigir isso, mova um ponteiro para a próxima linha","To fix this move the pointer to the next line":"Para corrigir isso, mova o ponteiro para a próxima linha","To fix this start the container <0/> on a different line":"Para corrigir isso, inicie o container <0/> em uma linha diferente","To learn more about why we require you to log in, please read <0>this blog post</0>.":"Para saber mais sobre por que precisamos que você faça login, leia <0>este post no blog</0>.","Top to Bottom":"De cima para baixo","Try again":"Tente novamente","Two edges have the same ID":"Dois bordos têm o mesmo ID","Two nodes have the same ID":"Dois nós têm o mesmo ID","Type":"Tipo","Unfreeze":"Descongelar","Unique text value to identify a node":"Valor de texto único para identificar um nó","Unknown Parsing Error":"Erro de Análise Desconhecido","Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":"Libere sua criatividade e simplifique seu fluxo de trabalho com o Flowchart Fun Pro - a partir de apenas $ 3 / mês!","Update Email":"Atualizar e-mail","Upgrade to Pro":"Atualizar para Pro","Upload your File":"Faça o upload do seu arquivo","Use AI to create flowcharts from data sets or algorithms":"Use IA para criar fluxogramas a partir de conjuntos de dados ou algoritmos.","Use AI to generate a flowchart from a prompt. <0>Learn More</0>":"Use IA para gerar um fluxograma a partir de uma solicitação. <0>Saiba mais</0>.","Use classes to group nodes":"Use classes para agrupar nós","Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":"Use o atributo <0>href</0> para definir um link em um nó que abra em uma nova guia.","Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":"Use o atributo <0>src</0> para definir a imagem de um nó. A imagem será dimensionada para caber no nó, então você pode precisar ajustar a largura e altura do nó para obter o resultado desejado. Apenas imagens públicas (não bloqueadas por CORS) são suportadas.","Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":"Use os atributos <0>w</0> e <1>h</1> para definir explicitamente a largura e altura de um nó.","Use the customer portal to change your billing information.":"Use o portal do cliente para alterar suas informações de cobrança.","Use this file for org charts, hierarchies, and other organizational structures.":"Use este arquivo para organogramas, hierarquias e outras estruturas organizacionais.","Use this file for sequences, processes, and workflows.":"Use este arquivo para sequências, processos e fluxos de trabalho.","User":"Usuário","View on Github":"Visualizar no Github","Water evaporates from the Earth\'s surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":"água evapora da superfície da Terra, sobe para a atmosfera e volta como precipitação. Esta água então corre para rios, lagos e oceanos, onde novamente evapora e é reciclada de volta para a atmosfera.","Welcome to Flowchart Fun Pro!":"Bem-vindo ao Flowchart Fun Pro!","Width and Height":"Largura e Altura","Would you like to continue?":"Você gostaria de continuar?","Yearly":"Anualmente","You already have a flowchart with this name.":"Você já tem um diagrama com este nome.","You are about to add {numNodes} nodes and {numEdges} edges to your graph.":["Você está prestes a adicionar ",["numNodes"]," nós e ",["numEdges"]," arestas ao seu gráfico."],"You can also wrap text in \\"\\\\( \\\\)\\" to create an edge":"Você também pode quebrar o texto dentro de \\"\\\\( \\\\)\\" para criar um link","You must log in to create a persistent flowchart.":"Você deve fazer login para criar um fluxograma persistente.","You need to log in to access this page.":"Você precisa fazer login para acessar esta página.","You\'re almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":"Você está quase lá! Apenas mais um passo para destravar o potencial total do <0>Flowchart Fun Pro</0>. Insira os detalhes do pagamento abaixo para completar sua assinatura e começar a criar fluxogramas incríveis hoje.","Your Charts":"Seus Gráficos","Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":"Sua assinatura não está mais ativa. Se você deseja criar e editar gráficos permanentes, atualize para o Pro.","Zoom In":"Zoom In","Zoom Out":"Diminuir o zoom","any text: before a colon creates a label":"qualquer texto: antes de dois pontos cria uma etiqueta","or":"ou"}'
+  ),
 };
