@@ -1,432 +1,5 @@
 /*eslint-disable*/ module.exports = {
-  messages: {
-    "$3 per month": "每月3美元",
-    "$30 per year": "每年30美元",
-    "(before a colon creates a label)": "(在冒号前创建一个标签)",
-    "<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":
-      "<0>Flowchart Fun</0>是由<1>Tone Row</1>制作的开源项目",
-    "A new version of the app is available. Please reload to update.":
-      "一个新版本的应用程序可用。请重新加载以更新。",
-    About: "关于",
-    "Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":
-      "从任何地方访问这些图表。<0/>共享并嵌入保持同步的流程图。",
-    "Accessible from any device": "可从任何设备访问",
-    Account: "帐户",
-    "Activate your account": "激活您的账户",
-    Advanced: "先进的",
-    Amount: "数量",
-    "An error occurred. Try resubmitting or email {0} directly.": [
-      "发生了一个错误。请尝试重新提交或直接发送电子邮件至",
-      ["0"],
-      "。",
-    ],
-    Appearance: "外观",
-    "Apply Style": "应用样式",
-    Attributes: "属性",
-    "Auto-Layouts": "自动布局",
-    "Back To Editor": "返回编辑器",
-    "Basic Flowchart": "基本流程图",
-    "Become a Github Sponsor": "成为Github赞助商",
-    "Become a Pro User": "成为专业用户",
-    "Begin with a simple example showing how <0>Flowchart Fun</0> works.":
-      "以一个简单的例子开始来展示<0>流程图的乐趣</0>是如何工作的。",
-    Blank: "空白",
-    Blog: "博客",
-    "Book a Meeting": "预订会议",
-    Border: "边框",
-    "Bottom to Top": "从下到上",
-    Breadthfirst: "宽度优先",
-    Cancel: "取消",
-    "Cancel your subscription. Your hosted charts will become read-only.":
-      "取消订阅。您的托管图表将变为只读。",
-    "Certain attributes can be used to customize the appearance or functionality of elements.":
-      "某些属性可用于自定义元素的外观或功能。",
-    "Change Email Address": "更改电子邮件地址",
-    Changelog: "变更日志",
-    Charts: "图表",
-    "Check your email for a link to log in.<0/>You can close this window.":
-      "检查您的电子邮件以获取登录链接。您可以关闭此窗口。",
-    "Choose a Theme": "选择一个主题",
-    "Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":
-      "为边缘的源和目标选择各种箭头形状。形状包括三角形，三角形-T，圆形-三角形，三角形-十字，三角形-后曲线，V形，T形，正方形，圆形，菱形，雪花形，无。",
-    "Choose the plan that's right for you and start creating amazing flowcharts with Flowchart Fun Pro":
-      "选择适合您的计划，开始使用Flowchart Fun Pro创建出色的流程图。",
-    Circle: "圆圈",
-    Classes: "类",
-    Clay: "粘土",
-    Clear: "清除",
-    "Clear text?": "清除文字？",
-    "Click on \\nLearn Syntax\\nto learn more": "点击 \\n学习语法\\n了解更多",
-    "Click to preview": "点击预览",
-    Clone: "克隆",
-    Close: "关闭",
-    Color: "颜色",
-    "Colors include red, orange, yellow, blue, purple, black, white, and gray.":
-      "颜色包括红色，橙色，黄色，蓝色，紫色，黑色，白色和灰色。",
-    Column: "列",
-    "Comic Book": "漫画书",
-    Comment: "评论",
-    Concentric: "同心",
-    "Confirm New Email": "确认新电子邮件",
-    Containers: "容器",
-    "Containers are nodes that contain other nodes. They are declared using curly braces.":
-      "容器是包含其他节点的节点。它们使用大括号声明。",
-    Continue: "继续",
-    "Convert this to a hosted chart?": "将此转换为托管图表？",
-    "Convert to hosted chart?": "是否转换为托管图表？",
-    Copy: "复制",
-    "Copy PNG Image": "复制PNG图像",
-    "Copy SVG Code": "复制 SVG 代码",
-    "Copy your mermaid.js code or open it directly in the mermaid.js live editor.":
-      "复制您的mermaid.js代码或直接在mermaid.js实时编辑器中打开它。",
-    "Create Flowcharts from a Prompt with AI": "使用 AI 根据提示创建图表",
-    "Create New Flowchart": "创建新流程图",
-    "Create a New Flowchart": "创建新图表",
-    "Create professional-quality visual aids for presentations or training materials":
-      "创建专业质量的演示文稿或培训材料",
-    "Creating an edge between two nodes is done by indenting the second node below the first":
-      "在两个节点之间创建边缘是通过将第二个节点缩进第一个节点来完成的",
-    "Custom Sharing Options": "自定义分享选项",
-    "Customer Portal": "客户门户",
-    "Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":
-      "通过编辑<0>Cytoscape CSS</0>来自定义主题，我们的样式文档即将推出！与此同时，最好的资源是<1>Cytoscape</1>文档。如果你遇到困难，可以在<2>Discord</2>上提问。",
-    Dark: "深色",
-    "Dark Mode": "深色模式",
-    Date: "日期",
-    "Deleted when browser data is cleared": "清除浏览器数据时删除",
-    "Describe the flowchart you wish to create": "描述您要创建的图表",
-    Direction: "方向",
-    Discard: "Discard",
-    "Do you want to cancel your subscription?": "您要取消订阅吗？",
-    "Do you want to delete this?": "您要将其删除吗？",
-    Document: "文档",
-    "Don't have an account?": "没有帐户？",
-    Download: "下载",
-    "Download JPG": "下载 JPG",
-    "Download PNG": "下载 PNG",
-    "Download SVG": "下载 SVG",
-    "Drag and drop a CSV file here, or click to select a file":
-      "将CSV文件拖放到此处，或单击以选择文件",
-    "Draw an edge from multiple nodes by beginning the line with a reference":
-      "通过引用开始行从多个节点绘制边缘",
-    "Drop the file here ...": "將檔案拖放到這裡...",
-    "Easily import existing flowcharts from other software":
-      "轻松从其他软件导入现有流程图",
-    Edge: "邊緣",
-    "Edge ID, Classes, Attributes": "邊緣ID、類別和屬性",
-    "Edge Label": "邊緣標籤",
-    "Edge Label Column": "邊緣標籤欄",
-    "Edge Style": "邊緣樣式",
-    "Edge missing indentation": "缺少缩进的边",
-    Edges: "边",
-    "Edges are declared in the same row as their source node":
-      "边声明在与源节点相同的行中",
-    "Edges are declared in the same row as their target node":
-      "边声明在与目标节点相同的行中",
-    "Edges are declared in their own row": "边声明在自己的行中",
-    "Edges can also have ID's, classes, and attributes before the label":
-      "边在标签之前可以有ID，类和属性",
-    "Edges can be styled with dashed, dotted, or solid lines":
-      "边可以用虚线，点线或实线样式",
-    "Edges in Separate Rows": "边在单独的行",
-    "Edges in Source Node Row": "边在源节点行",
-    "Edges in Target Node Row": "边在目标节点行",
-    Editable: "可编辑",
-    Editor: "编辑器",
-    Eggs: "蛋",
-    Email: "电子邮件",
-    Empty: "空",
-    "Enter your email": "请输入您的电子邮件",
-    "Equal To": "等于",
-    "Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":
-      "文章写作流程流程图，指导学生进行头脑风暴，大纲制定，起草和修改阶段",
-    "Everything in Free": "免费计划中的一切",
-    Excalidraw: "Excalidraw",
-    Export: "导出",
-    "Export High-Resolution Images": "导出高分辨率图像",
-    "Export to SVG": "导出为SVG",
-    "Extra Large": "超大",
-    Extract: "提取",
-    Feedback: "反馈",
-    "Follow Us on Twitter": "在Twitter上关注我们",
-    "Found a bug? Have a feature request? We would love to hear from you!":
-      "如果发现了一个 bug？有功能请求？我们很乐意听到您的意见！",
-    Free: "免费",
-    Fullscreen: "全屏",
-    Futuristic: "未来主义的",
-    "Get More Layouts": "获取更多布局",
-    "Get personalized attention and support from our expert team":
-      "从我们的专家团队获得个性化的关注和支持",
-    "Getting Started": "开始",
-    Graph: "图表",
-    Grid: "网格",
-    "Have complex questions or issues? We're here to help.":
-      "有复杂的问题或问题吗？我们在这里帮助你。",
-    Help: "帮助",
-    History: "历史",
-    Home: "主页",
-    "How are edges declared in this data?": "在这个数据中如何声明边缘？",
-    "ID's": "ID",
-    "If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":
-      "如果您喜欢使用<0>Flowchart Fun</0>，请考虑支持该项目",
-    "If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\:</0>":
-      "如果你想创建一个边，缩进这一行。如果不，用反斜杠转义冒号<0>\\:</0>",
-    "Image Export": "图像导出",
-    Images: "图像",
-    "Import Data": "导入数据",
-    "Import data from a CSV file.": "从CSV文件导入数据。",
-    "Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":
-      "從任何CSV檔案匯入資料並將其映射到新的流程圖。這是從Lucidchart、Google Sheets和Visio等其他來源匯入資料的一個很棒的方法。",
-    "Import from Lucidchart, Visio, and other CSV files":
-      "從Lucidchart、Visio和其他CSV檔案匯入",
-    "Import your diagram it into Microsoft Visio using one of these CSV files.":
-      "使用其中一个CSV文件将您的图表导入到Microsoft Visio中。",
-    "Important Changes are Coming to Flowchart Fun.":
-      "流程图乐趣将迎来重要变化。",
-    "Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":
-      "匯入資料是專業功能。您可以每月以3美元或每年30美元升級至Flowchart Fun Pro。",
-    "Impress your audience with high-quality images produced by Flowchart Fun":
-      "用Flowchart Fun生成的高质量图片给你的听众留下深刻印象",
-    "Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":
-      "使用<0>title</0>属性添加标题。要使用 Visio 颜色，请添加一个等于以下内容之一的<1>roleType</1>属性：",
-    "Indenting creates an edge": "缩进创建链接",
-    Info: "信息",
-    Instruct: "指导",
-    Is: "是",
-    Language: "语言",
-    Large: "大",
-    Layout: "布局",
-    "Layout Frozen": "布局已冻结",
-    "Layout is Frozen": "布局被冻结",
-    "Leading References": "主要參考",
-    "Learn More": "学到更多",
-    "Learn Syntax": "學習語法",
-    "Learn about Flowchart Fun Pro": "了解关于Flowchart Fun Pro",
-    "Left to Right": "从左到右",
-    "Let's get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":
-      "让我们开始吧！在下面输入您的电子邮件地址，创建您的Flowchart Fun帐户，开始使用我们强大的功能。",
-    Light: "浅色",
-    "Light Mode": "浅色模式",
-    Link: "链接",
-    "Load Theme": "加载主题",
-    "Loading a theme will erase any custom styles you have applied":
-      "加载主题将擦除您应用的任何自定义样式",
-    "Lock Zoom to Graph": "锁定缩放到图表",
-    "Log In": "登录",
-    "Log Out": "登出",
-    "Log in to start creating flowcharts.": "登入以開始創建流程圖。",
-    "Make a One-Time Donation": "进行一次性捐赠",
-    "Make publicly accessible": "设为公开访问",
-    "Make sure you use the same email you will use to log in.":
-      "確保您使用的是與登錄時相同的電子郵件。",
-    "Map Data": "對應資料",
-    "Market understanding and competitive landscape maintenance for SaaS product development":
-      "为SaaS产品开发提供市场理解和竞争环境维护",
-    Medium: "中等",
-    Method: "方法",
-    Monospace: "Monospace",
-    Monthly: "每月",
-    "Multiple pointers on same line": "同一行上的多个指针",
-    Museum: "博物馆",
-    Name: "姓名",
-    New: "新",
-    "New Email": "新邮件",
-    "Next charge": "下次扣费",
-    "No Edges": "沒有邊緣",
-    Node: "節點",
-    "Node Border Style": "节点边框样式",
-    "Node Colors": "节点颜色",
-    "Node ID": "节点ID",
-    "Node ID, Classes, Attributes": "节点ID、类、属性",
-    "Node Label": "节点标签",
-    "Node Shapes": "节点形状",
-    "Node and pointer on same line": "同一行上的节点和指针",
-    "Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":
-      "节点可以使用虚线、点线或双线进行样式设置，也可以使用border-none来移除边框。",
-    "Not Empty": "不为空",
-    "Office Hours": "工作时间",
-    "Once in a while the magic link will end up in your spam folder. If you don't see it after a few minutes, check there or request a new link.":
-      "偶尔，魔法链接会被放入您的垃圾邮件文件夹。如果几分钟后仍然没有收到，请检查垃圾邮件文件夹，或者重新请求新的链接。",
-    "One-on-One Support": "一对一支持",
-    "One-time Share Links": "一次性分享链接",
-    "Only available on this device.<0/>Clearing your browser cache will erase them.":
-      "仅在此设备上可用。<0/>清除浏览器缓存将抹去它们。",
-    "Open Customer Portal": "打开客户门户",
-    "Organization Chart": "组织结构图",
-    "Paste the information you wish to convert to a flowchart":
-      "将要转换的信息粘贴到流程图中",
-    Permalinks: "永久链接",
-    "Permanent Charts are a Pro Feature": "永久图表是专业功能",
-    "Permanent Flowcharts": "永久流程图",
-    Persistent: "坚持不懈",
-    "Persistent Flowcharts": "持久图",
-    Playbook: "剧本",
-    "Pointer and container on same line": "同一行上的指针和容器",
-    Pricing: "价钱",
-    "Process for corporate social responsibility initiatives development and implementation across company operations":
-      "公司社会责任倡议的发展和在公司业务中的实施过程",
-    "Processing Data": "处理数据",
-    Prompt: "提示",
-    Public: "公开",
-    Random: "随机",
-    "Read-only": "只读",
-    "Reference by Class": "按类引用",
-    "Reference by ID": "按 ID 参考",
-    "Reference by Label": "按标签参考",
-    References: "参考",
-    "References are used to create edges between nodes that are created elsewhere in the document":
-      "参考用于在文档中其他位置创建的节点之间创建边",
-    "Referencing a node by its exact label": "通过其确切标签引用节点",
-    "Referencing a node by its unique ID": "通过其唯一ID引用节点",
-    "Referencing multiple nodes with the same assigned class":
-      "使用相同分配的类引用多个节点",
-    "Refresh Page": "刷新页面",
-    "Reload to Update": "重新加载以更新",
-    "Remove Image Watermark": "去除图片水印",
-    Rename: "重命名",
-    Reset: "重置",
-    "Resume Subscription": "恢复订阅",
-    Retro: "复古的",
-    Return: "返回",
-    "Right to Left": "从右到左",
-    Roadmap: "路线图",
-    Save: "救球",
-    "Save 20% (2 months free!)": "節省20％（兩個月免費！）",
-    "Save time and automate certain tasks with this cutting-edge feature":
-      "通過這個尖端功能節省時間並自動化某些任務",
-    "Save time and streamline your workflow by importing data from other sources":
-      "通過從其他來源導入數據來節省時間並簡化您的工作流程",
-    "Select your plan!": "選擇您的計劃！",
-    "Send Feedback": "发送反馈",
-    Settings: "设置",
-    Shape: "形状",
-    "Sign Up": "注册",
-    "Sign Up Now": "现在注册",
-    Size: "尺寸",
-    Small: "小",
-    "Some classes are available to help style your flowchart":
-      "有些类可以帮助您设计流程图",
-    "Sorry! This page is only available in English.":
-      "抱歉！此页面只有英语版。",
-    "Source Arrow": "源箭头",
-    "Source Column": "源列",
-    "Source Delimiter": "源分隔符",
-    "Source/Target Arrow Shape": "源/目标箭头形状",
-    Spacing: "间距",
-    "Special Attributes": "特殊属性",
-    Start: "开始",
-    "Start Over": "重新開始",
-    Status: "状态",
-    "Store any data associated to a node": "將任何與節點相關的資料儲存",
-    "Stored in the cloud": "存储在云端",
-    "Stored on this computer": "保存在这台电脑上",
-    "Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":
-      "通過Flowchart Fun簡化您的工作流程並簡化您的流程視覺化。 立即訂閱並充分利用我們強大的Pro功能！",
-    Style: "风格",
-    "Style Classes": "樣式類別",
-    Submit: "提交",
-    Subscribe: "订阅",
-    Subscription: "订阅",
-    "Subscription will end": "订阅即将到期",
-    "Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":
-      "供应链分析和优化：成本降低，效率提高和利益相关者协作",
-    "Target Arrow": "目标箭头",
-    "Target Column": "目標欄",
-    "Target Delimiter": "目標分隔符",
-    Temporary: "临时",
-    "Temporary Flowchart Warning": "暂时流程图警告",
-    "Temporary Flowcharts": "临时图表",
-    "Text Sizes": "文字大小",
-    "Text followed by colon+space creates an edge with the text as the label":
-      "以冒号加空格结尾的文本将创建一个边，文本作为标签",
-    "Text on a line creates a node with the text as the label":
-      "在一行中的文本将创建一个节点，文本作为标签",
-    "Text sizes include small, regular, large, and extra-large.":
-      "文本大小包括小、正常、大和超大。",
-    "Thank you for your feedback!": "感谢您的反馈!",
-    "The column that contains the edge label(s)": "包含边标签的列",
-    "The column that contains the source node ID(s)": "包含源节点ID的列",
-    "The column that contains the target node ID(s)": "包含目标节点ID的列",
-    "The delimiter used to separate multiple source nodes":
-      "用于分隔多个源节点的分隔符",
-    "The delimiter used to separate multiple target nodes":
-      "用于分隔多个目标节点的分隔符",
-    "The possible shapes are:": "可能的形状是：",
-    "Theme Editor": "主题编辑器",
-    "There are no edges in this data": "此数据中没有边",
-    "This action cannot be undone.": "此操作无法撤销。",
-    "This app works by typing": "这款应用的工作原理是输入",
-    "This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":
-      "此文档仅保存在此计算机上。如果您想要一种存储文档并随时访问它们的方法，请尝试我们的Pro计划，每月只需3美元！",
-    "This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":
-      "只有专业用户才能使用此功能。 <0>成为专业用户</0>解锁。",
-    "This is a temporary chart, it will be deleted when you clear your browser cache.":
-      "这是一个临时图表，当您清除浏览器缓存时将被删除。",
-    Tip: "提示",
-    "To fix this change one of the edge IDs":
-      "为了修复这个，改变其中一个边的ID",
-    "To fix this change one of the node IDs": "要修复这个，更改其中一个节点ID",
-    "To fix this move one pointer to the next line":
-      "要修复这个，将指针移动到下一行",
-    "To fix this move the pointer to the next line":
-      "要修复这个，将指针移动到下一行",
-    "To fix this start the container <0/> on a different line":
-      "要修复这个，将容器<0/>放在另一行",
-    "Top to Bottom": "从上到下",
-    "Try again": "重试",
-    "Two edges have the same ID": "两个边有相同的ID",
-    "Two nodes have the same ID": "两个节点有相同的ID",
-    Type: "类型",
-    Unfreeze: "解冻",
-    "Unique text value to identify a node": "用于标识节点的唯一文本值",
-    "Unknown Parsing Error": "未知的解析错误",
-    "Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":
-      "釋放您的創意並簡化您的工作流程，從每月只需3美元開始，Flowchart Fun Pro！",
-    "Update Email": "更新电子邮件",
-    "Upgrade to Pro": "升级到Pro",
-    "Upload your File": "上传您的文件",
-    "Use AI to create flowcharts from data sets or algorithms":
-      "使用AI從數據集或算法創建流程圖",
-    "Use AI to generate a flowchart from a prompt. <0>Learn More</0>":
-      "使用AI从提示中生成流程图。<0>了解更多</0>",
-    "Use classes to group nodes": "使用类来分组节点",
-    "Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":
-      "使用属性<0>href</0>在节点上设置一个在新标签页中打开的链接。",
-    "Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":
-      "使用属性<0>src</0>来设置节点的图像，图像将被缩放以适应节点，因此您可能需要调整节点的宽度和高度以获得期望的结果。仅支持公共图像（不受CORS阻止）。",
-    "Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":
-      "使用属性<0>w</0>和<1>h</1>显式设置节点的宽度和高度。",
-    "Use the customer portal to change your billing information.":
-      "使用客户门户更改您的账单信息。",
-    "Use this file for org charts, hierarchies, and other organizational structures.":
-      "使用此文件制作组织图、层次结构和其他组织结构。",
-    "Use this file for sequences, processes, and workflows.":
-      "使用此文件进行顺序、流程和工作流程。",
-    User: "用户",
-    "View on Github": "在 Github 上查看",
-    "Water evaporates from the Earth's surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":
-      "水从地球表面蒸发，升入大气中，然后以降水形式回落。这些水然后流入河流，湖泊和海洋，在那里又蒸发，重新回到大气中循环。",
-    "We use magic links to log you in. Enter your email below to get started.":
-      "我们使用魔术链接来登录您。请在下面输入您的电子邮件以开始。",
-    "Welcome to Flowchart Fun Pro!": "欢迎来到流程图乐趣Pro！",
-    "Width and Height": "宽度和高度",
-    "Would you like to continue?": "您想继续吗？",
-    Yearly: "每年",
-    "You already have a flowchart with this name.": "您已有同名的图表。",
-    "You are about to add {numNodes} nodes and {numEdges} edges to your graph.":
-      ["您即将为您的图添加", ["numNodes"], "个节点和", ["numEdges"], "条边。"],
-    'You can also wrap text in "\\( \\)" to create an edge':
-      "您还可以将文本包装在\\n“\\( \\)”中以创建链接",
-    "You must log in to create a persistent flowchart.":
-      "您必须登录才能创建一个持久的流程图。",
-    "You're almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":
-      "您已经快完成了！只需要一步就可以解锁 <0>Flowchart Fun Pro</0> 的全部功能。请在下面输入您的付款详情来完成您的订阅，并今天开始创建令人惊叹的流程图。",
-    "Your Charts": "您的图表",
-    "Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":
-      "您的订阅已不再有效。如果您想创建和编辑永久图表，请升级到Pro。",
-    "Zoom In": "放大",
-    "Zoom Out": "縮小",
-    "any text: before a colon creates a label":
-      "任何文本： 在冒号前创建一个标签",
-  },
+  messages: JSON.parse(
+    '{"$3 per month":"每月3美元","$30 per year":"每年30美元","(before a colon creates a label)":"(在冒号前创建一个标签)","<0>Flowchart Fun</0> is an open source project made by <1>Tone Row</1>":"<0>Flowchart Fun</0>是由<1>Tone Row</1>制作的开源项目","<0>Sign In</0> / <1>Sign Up</1> with email and password":"<0>登录</0> / <1>注册</1> 使用电子邮件和密码","<0>Starting August 28th you will need a subscription to create and edit charts. <1>Learn more</1>.</0>":"<0>从8月28日开始，您需要订阅才能创建和编辑图表。<1>了解更多</1>。</0>","<0>You currently have a free account.<1/><2>Learn about our Pro Features and subscribe on our pricing page</2>.</0>":"<0>您目前拥有免费帐户。<1/><2>了解我们的Pro功能，并在我们的定价页面上订阅</2>。</0>","A new version of the app is available. Please reload to update.":"一个新版本的应用程序可用。请重新加载以更新。","About":"关于","Access these charts from anywhere.<0/>Share and embed flowcharts that stay in sync.":"从任何地方访问这些图表。<0/>共享并嵌入保持同步的流程图。","Accessible from any device":"可从任何设备访问","Account":"帐户","Activate your account":"激活您的账户","Advanced":"先进的","Amount":"数量","An error occurred. Try resubmitting or email {0} directly.":["发生了一个错误。请尝试重新提交或直接发送电子邮件至",["0"],"。"],"Appearance":"外观","Apply Style":"应用样式","Attributes":"属性","Auto-Layouts":"自动布局","Back To Editor":"返回编辑器","Basic Flowchart":"基本流程图","Become a Github Sponsor":"成为Github赞助商","Become a Pro User":"成为专业用户","Begin with a simple example showing how <0>Flowchart Fun</0> works.":"以一个简单的例子开始来展示<0>流程图的乐趣</0>是如何工作的。","Blank":"空白","Blog":"博客","Book a Meeting":"预订会议","Border":"边框","Bottom to Top":"从下到上","Breadthfirst":"宽度优先","Cancel":"取消","Cancel your subscription. Your hosted charts will become read-only.":"取消订阅。您的托管图表将变为只读。","Certain attributes can be used to customize the appearance or functionality of elements.":"某些属性可用于自定义元素的外观或功能。","Change Email Address":"更改电子邮件地址","Changelog":"变更日志","Charts":"图表","Check your email for a link to log in.<0/>You can close this window.":"检查您的电子邮件以获取登录链接。您可以关闭此窗口。","Choose":"選擇","Choose a Theme":"选择一个主题","Choose from a variety of arrow shapes for the source and target of an edge. Shapes include triangle, triangle-tee, circle-triangle, triangle-cross, triangle-backcurve, vee, tee, square, circle, diamond, chevron, none. .":"为边缘的源和目标选择各种箭头形状。形状包括三角形，三角形-T，圆形-三角形，三角形-十字，三角形-后曲线，V形，T形，正方形，圆形，菱形，雪花形，无。","Choose the plan that\'s right for you and start creating amazing flowcharts with Flowchart Fun Pro":"选择适合您的计划，开始使用Flowchart Fun Pro创建出色的流程图。","Circle":"圆圈","Classes":"类","Clay":"粘土","Clear":"清除","Clear text?":"清除文字？","Click on \\\\nLearn Syntax\\\\nto learn more":"点击 \\\\n学习语法\\\\n了解更多","Click to preview":"点击预览","Clone":"克隆","Close":"关闭","Color":"颜色","Colors include red, orange, yellow, blue, purple, black, white, and gray.":"颜色包括红色，橙色，黄色，蓝色，紫色，黑色，白色和灰色。","Column":"列","Comic Book":"漫画书","Comment":"评论","Concentric":"同心","Confirm New Email":"确认新电子邮件","Confirm your email address to sign in.":"確認您的電子郵件地址以登入。","Containers":"容器","Containers are nodes that contain other nodes. They are declared using curly braces.":"容器是包含其他节点的节点。它们使用大括号声明。","Continue":"继续","Convert this to a hosted chart?":"将此转换为托管图表？","Convert to hosted chart?":"是否转换为托管图表？","Copy":"复制","Copy PNG Image":"复制PNG图像","Copy SVG Code":"复制 SVG 代码","Copy your mermaid.js code or open it directly in the mermaid.js live editor.":"复制您的mermaid.js代码或直接在mermaid.js实时编辑器中打开它。","Create Flowcharts from a Prompt with AI":"使用 AI 根据提示创建图表","Create New Flowchart":"创建新流程图","Create a New Flowchart":"创建新图表","Create professional-quality visual aids for presentations or training materials":"创建专业质量的演示文稿或培训材料","Creating an edge between two nodes is done by indenting the second node below the first":"在两个节点之间创建边缘是通过将第二个节点缩进第一个节点来完成的","Custom Sharing Options":"自定义分享选项","Customer Portal":"客户门户","Customize your theme by editing the <0>Cytoscape CSS</0> below. Our styling documentation is coming soon! In the meantime, the best resource is the <1>Cytoscape</1> documentation. Come ask questions in the <2>Discord</2> if you get stuck.":"通过编辑<0>Cytoscape CSS</0>来自定义主题，我们的样式文档即将推出！与此同时，最好的资源是<1>Cytoscape</1>文档。如果你遇到困难，可以在<2>Discord</2>上提问。","Dark":"深色","Dark Mode":"深色模式","Date":"日期","Deleted when browser data is cleared":"清除浏览器数据时删除","Describe the flowchart you wish to create":"描述您要创建的图表","Direction":"方向","Discard":"Discard","Do you want to cancel your subscription?":"您要取消订阅吗？","Do you want to delete this?":"您要将其删除吗？","Document":"文档","Download":"下载","Download JPG":"下载 JPG","Download PNG":"下载 PNG","Download SVG":"下载 SVG","Drag and drop a CSV file here, or click to select a file":"将CSV文件拖放到此处，或单击以选择文件","Draw an edge from multiple nodes by beginning the line with a reference":"通过引用开始行从多个节点绘制边缘","Drop the file here ...":"將檔案拖放到這裡...","Easily import existing flowcharts from other software":"轻松从其他软件导入现有流程图","Edge":"邊緣","Edge ID, Classes, Attributes":"邊緣ID、類別和屬性","Edge Label":"邊緣標籤","Edge Label Column":"邊緣標籤欄","Edge Style":"邊緣樣式","Edge missing indentation":"缺少缩进的边","Edges":"边","Edges are declared in the same row as their source node":"边声明在与源节点相同的行中","Edges are declared in the same row as their target node":"边声明在与目标节点相同的行中","Edges are declared in their own row":"边声明在自己的行中","Edges can also have ID\'s, classes, and attributes before the label":"边在标签之前可以有ID，类和属性","Edges can be styled with dashed, dotted, or solid lines":"边可以用虚线，点线或实线样式","Edges in Separate Rows":"边在单独的行","Edges in Source Node Row":"边在源节点行","Edges in Target Node Row":"边在目标节点行","Editable":"可编辑","Editor":"编辑器","Eggs":"蛋","Email":"电子邮件","Empty":"空","Enter your email":"请输入您的电子邮件","Enter your email address and we\'ll send you a magic link to sign in.":"輸入您的電子郵件地址，我們將發送給您一個魔法鏈接以登入。","Enter your email address below and we\'ll send you a link to reset your password.":"在下面輸入您的電子郵件地址，我們將發送給您一個重置密碼的鏈接。","Equal To":"等于","Essay writing process flowchart, guiding students through brainstorming, outlining, drafting, and revising stages":"文章写作流程流程图，指导学生进行头脑风暴，大纲制定，起草和修改阶段","Everything in Free":"免费计划中的一切","Excalidraw":"Excalidraw","Export":"导出","Export High-Resolution Images":"导出高分辨率图像","Export to SVG":"导出为SVG","Extra Large":"超大","Extract":"提取","Feedback":"反馈","Follow Us on Twitter":"在Twitter上关注我们","Forgot your password?":"忘記密碼了？","Found a bug? Have a feature request? We would love to hear from you!":"如果发现了一个 bug？有功能请求？我们很乐意听到您的意见！","Free":"免费","Fullscreen":"全屏","Futuristic":"未来主义的","Get More Layouts":"获取更多布局","Get personalized attention and support from our expert team":"从我们的专家团队获得个性化的关注和支持","Getting Started":"开始","Graph":"图表","Grid":"网格","Have complex questions or issues? We\'re here to help.":"有复杂的问题或问题吗？我们在这里帮助你。","Help":"帮助","History":"历史","Home":"主页","How are edges declared in this data?":"在这个数据中如何声明边缘？","ID\'s":"ID","If an account with that email exists, we\'ve sent you an email with instructions on how to reset your password.":"如果該電子郵件存在該帳戶，我們已經發送給您一封電子郵件，其中包含如何重置您的密碼的說明。","If you enjoy using <0>Flowchart Fun</0>, please consider supporting the project":"如果您喜欢使用<0>Flowchart Fun</0>，请考虑支持该项目","If you mean to create an edge, indent this line. If not, escape the colon with a backslash <0>\\\\:</0>":"如果你想创建一个边，缩进这一行。如果不，用反斜杠转义冒号<0>\\\\:</0>","Image Export":"图像导出","Images":"图像","Import Data":"导入数据","Import data from a CSV file.":"从CSV文件导入数据。","Import data from any CSV file and map it to a new flowchart. This is a great way to import data from other sources like Lucidchart, Google Sheets, and Visio.":"從任何CSV檔案匯入資料並將其映射到新的流程圖。這是從Lucidchart、Google Sheets和Visio等其他來源匯入資料的一個很棒的方法。","Import from Lucidchart, Visio, and other CSV files":"從Lucidchart、Visio和其他CSV檔案匯入","Import your diagram it into Microsoft Visio using one of these CSV files.":"使用其中一个CSV文件将您的图表导入到Microsoft Visio中。","Important Changes are Coming to Flowchart Fun.":"流程图乐趣将迎来重要变化。","Importing data is a pro feature. You can upgrade to Flowchart Fun Pro for just $3/month or $30/year.":"匯入資料是專業功能。您可以每月以3美元或每年30美元升級至Flowchart Fun Pro。","Impress your audience with high-quality images produced by Flowchart Fun":"用Flowchart Fun生成的高质量图片给你的听众留下深刻印象","Include a title using a <0>title</0> attribute. To use Visio coloring, add a <1>roleType</1> attribute equal to one of the following:":"使用<0>title</0>属性添加标题。要使用 Visio 颜色，请添加一个等于以下内容之一的<1>roleType</1>属性：","Indenting creates an edge":"缩进创建链接","Info":"信息","Instruct":"指导","Is":"是","Language":"语言","Large":"大","Layout":"布局","Layout Frozen":"布局已冻结","Layout is Frozen":"布局被冻结","Leading References":"主要參考","Learn More":"学到更多","Learn Syntax":"學習語法","Learn about Flowchart Fun Pro":"了解关于Flowchart Fun Pro","Left to Right":"从左到右","Let\'s get started! Enter your email address below to create your Flowchart Fun account and start using our powerful features.":"让我们开始吧！在下面输入您的电子邮件地址，创建您的Flowchart Fun帐户，开始使用我们强大的功能。","Light":"浅色","Light Mode":"浅色模式","Link":"链接","Load Theme":"加载主题","Loading a theme will erase any custom styles you have applied":"加载主题将擦除您应用的任何自定义样式","Lock Zoom to Graph":"锁定缩放到图表","Log In":"登录","Log Out":"登出","Log in to start creating flowcharts.":"登入以開始創建流程圖。","Make a One-Time Donation":"进行一次性捐赠","Make publicly accessible":"设为公开访问","Make sure you use the same email you will use to log in.":"確保您使用的是與登錄時相同的電子郵件。","Map Data":"對應資料","Market understanding and competitive landscape maintenance for SaaS product development":"为SaaS产品开发提供市场理解和竞争环境维护","Medium":"中等","Method":"方法","Monospace":"Monospace","Monthly":"每月","Multiple pointers on same line":"同一行上的多个指针","Museum":"博物馆","Name":"姓名","New":"新","New Email":"新邮件","Next charge":"下次扣费","No Edges":"沒有邊緣","Node":"節點","Node Border Style":"节点边框样式","Node Colors":"节点颜色","Node ID":"节点ID","Node ID, Classes, Attributes":"节点ID、类、属性","Node Label":"节点标签","Node Shapes":"节点形状","Node and pointer on same line":"同一行上的节点和指针","Nodes can be styled with dashed, dotted, or double. Borders can also be removed with border-none.":"节点可以使用虚线、点线或双线进行样式设置，也可以使用border-none来移除边框。","Not Empty":"不为空","Office Hours":"工作时间","Once in a while the magic link will end up in your spam folder. If you don\'t see it after a few minutes, check there or request a new link.":"偶尔，魔法链接会被放入您的垃圾邮件文件夹。如果几分钟后仍然没有收到，请检查垃圾邮件文件夹，或者重新请求新的链接。","One-on-One Support":"一对一支持","One-time Share Links":"一次性分享链接","Only available on this device.<0/>Clearing your browser cache will erase them.":"仅在此设备上可用。<0/>清除浏览器缓存将抹去它们。","Open Customer Portal":"打开客户门户","Organization Chart":"组织结构图","Password":"密碼","Paste the information you wish to convert to a flowchart":"将要转换的信息粘贴到流程图中","Permalinks":"永久链接","Permanent Charts are a Pro Feature":"永久图表是专业功能","Permanent Flowcharts":"永久流程图","Persistent":"坚持不懈","Persistent Flowcharts":"持久图","Playbook":"剧本","Pointer and container on same line":"同一行上的指针和容器","Pricing":"价钱","Process for corporate social responsibility initiatives development and implementation across company operations":"公司社会责任倡议的发展和在公司业务中的实施过程","Processing Data":"处理数据","Prompt":"提示","Public":"公开","Random":"随机","Read-only":"只读","Reference by Class":"按类引用","Reference by ID":"按 ID 参考","Reference by Label":"按标签参考","References":"参考","References are used to create edges between nodes that are created elsewhere in the document":"参考用于在文档中其他位置创建的节点之间创建边","Referencing a node by its exact label":"通过其确切标签引用节点","Referencing a node by its unique ID":"通过其唯一ID引用节点","Referencing multiple nodes with the same assigned class":"使用相同分配的类引用多个节点","Refresh Page":"刷新页面","Reload to Update":"重新加载以更新","Remove Image Watermark":"去除图片水印","Rename":"重命名","Request Magic Link":"請求魔法鏈接","Request Password Reset":"請求密碼重置","Reset":"重置","Reset Password":"重置密碼","Resume Subscription":"恢复订阅","Retro":"复古的","Return":"返回","Right to Left":"从右到左","Roadmap":"路线图","Save":"救球","Save 20% (2 months free!)":"節省20％（兩個月免費！）","Save time and automate certain tasks with this cutting-edge feature":"通過這個尖端功能節省時間並自動化某些任務","Save time and streamline your workflow by importing data from other sources":"通過從其他來源導入數據來節省時間並簡化您的工作流程","Select your plan!":"選擇您的計劃！","Send Feedback":"发送反馈","Settings":"设置","Shape":"形状","Sign In":"登錄","Sign Up":"注册","Sign in with <0>GitHub</0>":"使用<0>GitHub</0>登录","Sign in with <0>Google</0>":"使用<0>Google</0>登录","Size":"尺寸","Small":"小","Some classes are available to help style your flowchart":"有些类可以帮助您设计流程图","Sorry! This page is only available in English.":"抱歉！此页面只有英语版。","Source Arrow":"源箭头","Source Column":"源列","Source Delimiter":"源分隔符","Source/Target Arrow Shape":"源/目标箭头形状","Spacing":"间距","Special Attributes":"特殊属性","Start":"开始","Start Over":"重新開始","Status":"状态","Store any data associated to a node":"將任何與節點相關的資料儲存","Stored in the cloud":"存储在云端","Stored on this computer":"保存在这台电脑上","Streamline your workflow and simplify your process visualization with Flowchart Fun. Subscribe now and take advantage of our powerful Pro features!":"通過Flowchart Fun簡化您的工作流程並簡化您的流程視覺化。 立即訂閱並充分利用我們強大的Pro功能！","Style":"风格","Style Classes":"樣式類別","Submit":"提交","Subscribe":"订阅","Subscription":"订阅","Subscription will end":"订阅即将到期","Supply chain analysis and optimization: cost reduction, efficiency improvement, and stakeholder collaboration":"供应链分析和优化：成本降低，效率提高和利益相关者协作","Target Arrow":"目标箭头","Target Column":"目標欄","Target Delimiter":"目標分隔符","Temporary":"临时","Temporary Flowchart Warning":"暂时流程图警告","Temporary Flowcharts":"临时图表","Text Sizes":"文字大小","Text followed by colon+space creates an edge with the text as the label":"以冒号加空格结尾的文本将创建一个边，文本作为标签","Text on a line creates a node with the text as the label":"在一行中的文本将创建一个节点，文本作为标签","Text sizes include small, regular, large, and extra-large.":"文本大小包括小、正常、大和超大。","Thank you for your feedback!":"感谢您的反馈!","The column that contains the edge label(s)":"包含边标签的列","The column that contains the source node ID(s)":"包含源节点ID的列","The column that contains the target node ID(s)":"包含目标节点ID的列","The delimiter used to separate multiple source nodes":"用于分隔多个源节点的分隔符","The delimiter used to separate multiple target nodes":"用于分隔多个目标节点的分隔符","The possible shapes are:":"可能的形状是：","Theme Editor":"主题编辑器","There are no edges in this data":"此数据中没有边","This action cannot be undone.":"此操作无法撤销。","This app works by typing":"这款应用的工作原理是输入","This document is only saved on this computer. If you want a way to store documents and access them wherever you go try our Pro plan for only $3/month!":"此文档仅保存在此计算机上。如果您想要一种存储文档并随时访问它们的方法，请尝试我们的Pro计划，每月只需3美元！","This feature is only available to pro users. <0>Become a pro user</0> to unlock it.":"只有专业用户才能使用此功能。 <0>成为专业用户</0>解锁。","This is a temporary chart, it will be deleted when you clear your browser cache.":"这是一个临时图表，当您清除浏览器缓存时将被删除。","Tip":"提示","To fix this change one of the edge IDs":"为了修复这个，改变其中一个边的ID","To fix this change one of the node IDs":"要修复这个，更改其中一个节点ID","To fix this move one pointer to the next line":"要修复这个，将指针移动到下一行","To fix this move the pointer to the next line":"要修复这个，将指针移动到下一行","To fix this start the container <0/> on a different line":"要修复这个，将容器<0/>放在另一行","To learn more about why we require you to log in, please read <0>this blog post</0>.":"要了解更多關於我們為什麼要求您登錄的原因，請閱讀<0>這篇博客文章</0>。","Top to Bottom":"从上到下","Try again":"重试","Two edges have the same ID":"两个边有相同的ID","Two nodes have the same ID":"两个节点有相同的ID","Type":"类型","Unfreeze":"解冻","Unique text value to identify a node":"用于标识节点的唯一文本值","Unknown Parsing Error":"未知的解析错误","Unleash your creativity and streamline your workflow with Flowchart Fun Pro – starting at just $3/month!":"釋放您的創意並簡化您的工作流程，從每月只需3美元開始，Flowchart Fun Pro！","Update Email":"更新电子邮件","Upgrade to Pro":"升级到Pro","Upload your File":"上传您的文件","Use AI to create flowcharts from data sets or algorithms":"使用AI從數據集或算法創建流程圖","Use AI to generate a flowchart from a prompt. <0>Learn More</0>":"使用AI从提示中生成流程图。<0>了解更多</0>","Use classes to group nodes":"使用类来分组节点","Use the attribute <0>href</0> to set a link on a node that opens in a new tab.":"使用属性<0>href</0>在节点上设置一个在新标签页中打开的链接。","Use the attribute <0>src</0> to set the image of a node. The image will be scaled to fit the node, so you may need to adjust the width and height of the node to get the desired result. Only public images (not blocked by CORS) are supported.":"使用属性<0>src</0>来设置节点的图像，图像将被缩放以适应节点，因此您可能需要调整节点的宽度和高度以获得期望的结果。仅支持公共图像（不受CORS阻止）。","Use the attributes <0>w</0> and <1>h</1> to explicitly set the width and height of a node.":"使用属性<0>w</0>和<1>h</1>显式设置节点的宽度和高度。","Use the customer portal to change your billing information.":"使用客户门户更改您的账单信息。","Use this file for org charts, hierarchies, and other organizational structures.":"使用此文件制作组织图、层次结构和其他组织结构。","Use this file for sequences, processes, and workflows.":"使用此文件进行顺序、流程和工作流程。","User":"用户","View on Github":"在 Github 上查看","Water evaporates from the Earth\'s surface, rises into the atmosphere and falls back down as precipitation. This water then runs off into rivers, lakes and oceans, where it again evaporates and is recycled back into the atmosphere.":"水从地球表面蒸发，升入大气中，然后以降水形式回落。这些水然后流入河流，湖泊和海洋，在那里又蒸发，重新回到大气中循环。","Welcome to Flowchart Fun Pro!":"欢迎来到流程图乐趣Pro！","Width and Height":"宽度和高度","Would you like to continue?":"您想继续吗？","Yearly":"每年","You already have a flowchart with this name.":"您已有同名的图表。","You are about to add {numNodes} nodes and {numEdges} edges to your graph.":["您即将为您的图添加",["numNodes"],"个节点和",["numEdges"],"条边。"],"You can also wrap text in \\"\\\\( \\\\)\\" to create an edge":"您还可以将文本包装在\\\\n“\\\\( \\\\)”中以创建链接","You must log in to create a persistent flowchart.":"您必须登录才能创建一个持久的流程图。","You need to log in to access this page.":"您需要登录才能访问此页面。","You\'re almost there! Just one more step to unlock the full potential of <0>Flowchart Fun Pro</0>. Enter your payment details below to complete your subscription and start creating amazing flowcharts today.":"您已经快完成了！只需要一步就可以解锁 <0>Flowchart Fun Pro</0> 的全部功能。请在下面输入您的付款详情来完成您的订阅，并今天开始创建令人惊叹的流程图。","Your Charts":"您的图表","Your subscription is no longer active. If you want to create and edit permanent charts upgrade to Pro.":"您的订阅已不再有效。如果您想创建和编辑永久图表，请升级到Pro。","Zoom In":"放大","Zoom Out":"縮小","any text: before a colon creates a label":"任何文本： 在冒号前创建一个标签","or":"或"}'
+  ),
 };
